% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{alpha2col}
\alias{alpha2col}
\title{set R color alpha value}
\usage{
alpha2col(x, alpha = 1, maxValue = 1, ...)
}
\arguments{
\item{x}{R compatible color, either a color name, or hex value, or
a mixture of the two. Any value compatible with \code{grDevices::col2rgb()}.}

\item{alpha}{numeric alpha transparency to use per x color. alpha is
recycled to length(x) as needed.}

\item{maxValue}{numeric maximum value to return, useful when the downstream
alpha range should be 255. By default maxValue=1 is returned.}

\item{...}{Additional arguments are ignored.}
}
\value{
\code{character} vector of R colors, with alpha values.
}
\description{
Define the alpha transparency per R color
}
\examples{
withr::with_par(list("mfrow"=c(2,2)), {
for (alpha in c(1, 0.8, 0.5, 0.2)) {
   nullPlot(plotAreaTitle=paste0("alpha=", alpha),
      doMargins=FALSE);
   usrBox(fill=alpha2col("yellow",
      alpha=alpha));
}
})

}
\seealso{
Other jam color functions:
\code{\link[=applyCLrange]{applyCLrange()}},
\code{\link[=col2alpha]{col2alpha()}},
\code{\link[=col2hcl]{col2hcl()}},
\code{\link[=col2hsl]{col2hsl()}},
\code{\link[=col2hsv]{col2hsv()}},
\code{\link[=color2gradient]{color2gradient()}},
\code{\link[=fixYellow]{fixYellow()}},
\code{\link[=fixYellowHue]{fixYellowHue()}},
\code{\link[=getColorRamp]{getColorRamp()}},
\code{\link[=hcl2col]{hcl2col()}},
\code{\link[=hsl2col]{hsl2col()}},
\code{\link[=hsv2col]{hsv2col()}},
\code{\link[=isColor]{isColor()}},
\code{\link[=kable_coloring]{kable_coloring()}},
\code{\link[=makeColorDarker]{makeColorDarker()}},
\code{\link[=rainbow2]{rainbow2()}},
\code{\link[=rgb2col]{rgb2col()}},
\code{\link[=setCLranges]{setCLranges()}},
\code{\link[=setTextContrastColor]{setTextContrastColor()}},
\code{\link[=showColors]{showColors()}},
\code{\link[=unalpha]{unalpha()}},
\code{\link[=warpRamp]{warpRamp()}}
}
\concept{jam color functions}
