% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{applyLceiling}
\alias{applyLceiling}
\title{Apply a lightness ceiling to R color}
\usage{
applyLceiling(x, Lceiling = 77, ...)
}
\arguments{
\item{x}{vector of R colors, either a text value from colors() or a
valid hex color, for example "#AA3399".}

\item{Lceiling}{integer length=1, indicating the maximum allowed L value
when colors are converted to HCL color space.}
}
\value{
vector of R colors
}
\description{
Apply a lightness ceiling to R color
}
\details{
This function takes an R color, converts to HCL colorspace, then applies
a ceiling to the L lightness value for all colors. It is primarily intended
to make text colors dark enough for display on a light background color,
for visible contrast.
}
\examples{
x <- c("yellow","orange","white","grey20");
applyLceiling(x);
printDebug(applyLceiling(x));

}
