% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-plots.r
\name{breakDensity}
\alias{breakDensity}
\title{Calculate more detailed density of numeric values}
\usage{
breakDensity(
  x,
  breaks = length(x)/3,
  bw = NULL,
  width = NULL,
  densityBreaksFactor = 3,
  weightFactor = 1,
  addZeroEnds = TRUE,
  baseline = 0,
  floorBaseline = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{numeric vector}

\item{breaks}{numeric breaks as described for \code{stats::density()} except
that single integer value is multiplied by \code{densityBreaksFactor}.}

\item{bw}{character name of a bandwidth function, or NULL.}

\item{width}{NULL or numeric value indicating the width of breaks to
apply.}

\item{densityBreaksFactor}{numeric factor to adjust the width of
density breaks, where higher values result in less detail.}

\item{weightFactor}{optional vector of weights \code{length(x)} to apply
to the density calculation.}

\item{addZeroEnds}{logical indicating whether the start and end value
should always be zero, which can be helpful for creating a polygon.}

\item{baseline}{optional numeric value indicating the expected baseline,
which is typically zero, but can be set to a higher value to indicate
a "noise floor".}

\item{floorBaseline}{logical indicating whether to apply a noise floor
to the output data.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional parameters are sent to \code{stats::density()}.}
}
\description{
Calculate more detailed density of numeric values
}
\details{
This function is a drop-in replacement for \code{stats::density()},
simply to provide a quick alternative that defaults to a higher
level of detail. Detail can be adjusted using \code{densityBreaksFactor},
where higher values will use a wider step size, thus lowering
the detail in the output.
}
\seealso{
Other jam practical functions: 
\code{\link{applyCLrange}()},
\code{\link{checkLightMode}()},
\code{\link{colNum2excelName}()},
\code{\link{exp2signed}()},
\code{\link{fileInfo}()},
\code{\link{fixYellowHue}()},
\code{\link{fixYellow}()},
\code{\link{getAxisLabel}()},
\code{\link{handleArgsText}()},
\code{\link{isFALSEV}()},
\code{\link{isTRUEV}()},
\code{\link{jamba}},
\code{\link{jargs}()},
\code{\link{kable_coloring}()},
\code{\link{log2signed}()},
\code{\link{make_html_styles}()},
\code{\link{make_styles}()},
\code{\link{mergeAllXY}()},
\code{\link{minorLogTicks}()},
\code{\link{newestFile}()},
\code{\link{printDebugI}()},
\code{\link{printDebug}()},
\code{\link{renameColumn}()},
\code{\link{rmInfinite}()},
\code{\link{rmNA}()},
\code{\link{rmNULL}()},
\code{\link{sclass}()},
\code{\link{sdima}()},
\code{\link{sdim}()},
\code{\link{setCLranges}()},
\code{\link{setPrompt}()},
\code{\link{ssdima}()},
\code{\link{ssdim}()}
}
\concept{jam practical functions}
