% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-clrange.R
\name{checkLightMode}
\alias{checkLightMode}
\title{check lightMode for light background color}
\usage{
checkLightMode(
  lightMode = getOption("jam.lightMode"),
  positron_default = getOption("jam.lightMode.positron", TRUE),
  html_default = getOption("jam.lightMode.html", TRUE),
  knitr_default = getOption("jam.lightMode.knitr", TRUE),
  other_default = getOption("jam.lightMode.other", FALSE),
  ...
)
}
\arguments{
\item{lightMode}{\code{logical} or NULL, indicating whether the lightMode
parameter has been defined in the function call.
The default \code{getOption("jam.lightMode")} is NULL unless assigned.
When NULL it will use logic described in Details.
When assigned, it is used as-is.}

\item{positron_default}{\code{logical} default TRUE, applied only when
running inside Positron, and when \code{rstudioapi::getThemeInfo()} is
not available, which is default for Positron as of jamba-1.0.5.
When running inside RStudio, it calls \code{rstudioapi::getThemeInfo()}
to determine whether the theme is light or dark.}

\item{html_default}{\code{logical} default TRUE, applied when knitr
is running with html output.}

\item{knitr_default}{\code{logical} default TRUE, applied when knitr
is running.}

\item{other_default}{\code{logical} default FALSE, applied when running
outside Positron or RStudio.}

\item{...}{Additional arguments are ignored.}
}
\value{
\code{logical} indicating whether lightMode is defined
}
\description{
check lightMode for light background color
}
\details{
Check the lightMode status through options('jam.lightMode'),
or by determining the running environment.

Logic is applied as follows:
\itemize{
\item When \code{lightMode} is defined upfront as \code{logical}, it is used.
\item When \code{getOption('jam.lightMode')} is defined, it is used.
\item When \code{lightMode} is NULL, it will apply a suitable default:
\itemize{
\item If running inside knitr, it will apply \code{html_default} for html output,
and \code{knitr_default} otherwise. Typically both produce white background,
however this gives the option for custom background color.
\item If running inside Positron or RStudio, it will attempt to run
\code{rstudioapi::getThemeInfo()} and use the resulting 'theme$dark',
otherwise it applies \code{positron_default}.
\item All other cases use \code{other_default}, default is FALSE to
apply to linux remote shell session with dark background.
}
}

To set a persistent default lightMode, add this line to .Rprofile:
\itemize{
\item \code{options("jam.lightMode"=TRUE)}
}
}
\examples{
checkLightMode(FALSE);
checkLightMode();

}
\seealso{
Other jam practical functions:
\code{\link[=breakDensity]{breakDensity()}},
\code{\link[=call_fn_ellipsis]{call_fn_ellipsis()}},
\code{\link[=check_pkg_installed]{check_pkg_installed()}},
\code{\link[=colNum2excelName]{colNum2excelName()}},
\code{\link[=color_dither]{color_dither()}},
\code{\link[=exp2signed]{exp2signed()}},
\code{\link[=getAxisLabel]{getAxisLabel()}},
\code{\link[=isFALSEV]{isFALSEV()}},
\code{\link[=isTRUEV]{isTRUEV()}},
\code{\link[=jargs]{jargs()}},
\code{\link[=kable_coloring]{kable_coloring()}},
\code{\link[=lldf]{lldf()}},
\code{\link[=log2signed]{log2signed()}},
\code{\link[=middle]{middle()}},
\code{\link[=minorLogTicks]{minorLogTicks()}},
\code{\link[=newestFile]{newestFile()}},
\code{\link[=printDebug]{printDebug()}},
\code{\link[=reload_qmd_cache]{reload_qmd_cache()}},
\code{\link[=reload_rmarkdown_cache]{reload_rmarkdown_cache()}},
\code{\link[=renameColumn]{renameColumn()}},
\code{\link[=rmInfinite]{rmInfinite()}},
\code{\link[=rmNA]{rmNA()}},
\code{\link[=rmNAs]{rmNAs()}},
\code{\link[=rmNULL]{rmNULL()}},
\code{\link[=setPrompt]{setPrompt()}}
}
\concept{jam practical functions}
