% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{col2alpha}
\alias{col2alpha}
\title{get R color alpha value}
\usage{
col2alpha(x, maxValue = 1, ...)
}
\arguments{
\item{x}{\code{character} R compatible color, either a color name, hex value, or
a mixture of the two. Any value compatible with \code{grDevices::col2rgb()}.}

\item{maxValue}{\code{numeric} maximum value to return, useful when the downstream
alpha range should be 255. By default maxValue=1 is returned.}

\item{...}{Additional arguments are ignored.}
}
\value{
\code{numeric} vector of alpha values
}
\description{
Return the alpha transparency per R color
}
\examples{
col2alpha(c("red", "#99004499", "beige", "transparent", "#FFFFFF00"))

}
\seealso{
Other jam color functions:
\code{\link[=alpha2col]{alpha2col()}},
\code{\link[=applyCLrange]{applyCLrange()}},
\code{\link[=col2hcl]{col2hcl()}},
\code{\link[=col2hsl]{col2hsl()}},
\code{\link[=col2hsv]{col2hsv()}},
\code{\link[=color2gradient]{color2gradient()}},
\code{\link[=fixYellow]{fixYellow()}},
\code{\link[=fixYellowHue]{fixYellowHue()}},
\code{\link[=getColorRamp]{getColorRamp()}},
\code{\link[=hcl2col]{hcl2col()}},
\code{\link[=hsl2col]{hsl2col()}},
\code{\link[=hsv2col]{hsv2col()}},
\code{\link[=isColor]{isColor()}},
\code{\link[=kable_coloring]{kable_coloring()}},
\code{\link[=makeColorDarker]{makeColorDarker()}},
\code{\link[=rainbow2]{rainbow2()}},
\code{\link[=rgb2col]{rgb2col()}},
\code{\link[=setCLranges]{setCLranges()}},
\code{\link[=setTextContrastColor]{setTextContrastColor()}},
\code{\link[=showColors]{showColors()}},
\code{\link[=unalpha]{unalpha()}},
\code{\link[=warpRamp]{warpRamp()}}
}
\concept{jam color functions}
