% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{col2hcl}
\alias{col2hcl}
\title{convert R color to HCL color matrix}
\usage{
col2hcl(
  x,
  maxColorValue = 255,
  model = getOption("jam.model", c("hcl", "polarLUV", "polarLAB")),
  ...
)
}
\arguments{
\item{x}{\code{character} R compatible color, either a color name, hex value, or
a mixture of the two. Any value compatible with
\code{grDevices::col2rgb()}.}

\item{maxColorValue}{\code{numeric} maximum value to return, useful
when the downstream alpha range should be 255.
By default maxValue=1 is returned.}

\item{model}{\code{character} color model to use
\itemize{
\item \code{"hcl"} to use \code{farver} HCL
\item \code{"polarLUV"} for the standard R conventional HCL,
\item \code{"polarLAB"} which uses the LAB-based HCL values.
}}

\item{...}{additional arguments are ignored.}
}
\value{
\code{numeric} matrix with H, C, L values.
}
\description{
convert R color to HCL color matrix
}
\details{
This function takes an R color and converts to an HCL matrix, using
the colorspace package, and \code{colorspace::RGB()} and
\code{colorspace::polarLUV()} functions. It is also used to
maintain alpha transparency, to enable interconversion via other
color manipulation functions as well.

When \code{model="hcl"} this function uses \code{farver::decode_colour()}
and bypasses \code{colorspace}. In future the \code{colorspace} dependency
will likely be removed in favor of using \code{farver}. In any event,
\code{model="hcl"} is equivalent to using \code{model="polarLUV"} and
\code{fixup=TRUE}, except that it should be much faster.
}
\examples{
col2hcl("#FF000044")

}
\seealso{
Other jam color functions:
\code{\link[=alpha2col]{alpha2col()}},
\code{\link[=applyCLrange]{applyCLrange()}},
\code{\link[=col2alpha]{col2alpha()}},
\code{\link[=col2hsl]{col2hsl()}},
\code{\link[=col2hsv]{col2hsv()}},
\code{\link[=color2gradient]{color2gradient()}},
\code{\link[=fixYellow]{fixYellow()}},
\code{\link[=fixYellowHue]{fixYellowHue()}},
\code{\link[=getColorRamp]{getColorRamp()}},
\code{\link[=hcl2col]{hcl2col()}},
\code{\link[=hsl2col]{hsl2col()}},
\code{\link[=hsv2col]{hsv2col()}},
\code{\link[=isColor]{isColor()}},
\code{\link[=kable_coloring]{kable_coloring()}},
\code{\link[=makeColorDarker]{makeColorDarker()}},
\code{\link[=rainbow2]{rainbow2()}},
\code{\link[=rgb2col]{rgb2col()}},
\code{\link[=setCLranges]{setCLranges()}},
\code{\link[=setTextContrastColor]{setTextContrastColor()}},
\code{\link[=showColors]{showColors()}},
\code{\link[=unalpha]{unalpha()}},
\code{\link[=warpRamp]{warpRamp()}}
}
\concept{jam color functions}
