% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{color2gradient}
\alias{color2gradient}
\title{Make a color gradient}
\usage{
color2gradient(col, n = NULL, gradientWtFactor = 2/3,
  reverseGradient = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{col}{vector of one or more colors, or list of color vectors, or
vector of repeated colors. If the vector contains unique colors, each
color is expanded into a gradient of length \code{n}, where \code{n} is
recycled for each input color. If the vector contains repeated colors,
they are expanded in place.}

\item{n}{integer vector of length one or more, which defines the number
of colors to return for each gradient.}

\item{gradientWtFactor}{numeric fraction representing the amount to expand
a color toward its maximum brightness and darkness.}

\item{reverseGradient}{logical whether to return light-to-dark gradient
(TRUE) or dark-to-light gradient (FALSE).}

\item{verbose}{logical whether to print verbose output.}
}
\description{
Make a color gradient
}
\details{
This function converts a single color into a color gradient by expanding
the initial color into lighter and darker colors around the central color.
The amount of gradient expansion is controlled by gradientWtFactor, which
is a weight factor scaled to the maximum available range of bright to
dark colors.

As an extension, the function can take a vector of colors, and expand each
into its own color gradient, each with its own number of colors.
If a vector with supplied that contains repeated colors, these colors
are expanded in-place into a gradient, bypassing the value for \code{n}.

If a list is supplied, a list is returned of the same length, where
each vector inside the list is a color gradient of length specified
by \code{n}. If the input list contains multiple values, only the first
color is used to define the color gradient.
}
