% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-string.r
\name{cPasteSU}
\alias{cPasteSU}
\title{paste a list into a delimited vector using sorted, unique values}
\usage{
cPasteSU(
  x,
  sep = ",",
  doSort = TRUE,
  makeUnique = TRUE,
  na.rm = FALSE,
  keepFactors = FALSE,
  checkClass = TRUE,
  useBioc = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{list} of vectors}

\item{sep}{\code{character} delimiter used to paste multiple values together}

\item{doSort}{\code{logical} indicating whether to sort each vector
using \code{\link[=mixedOrder]{mixedOrder()}}.}

\item{makeUnique}{\code{logical} indicating whether to make each vector in
the input list unique before pasting its values together.}

\item{na.rm}{\code{logical} indicating whether to remove NA values from
each vector in the input list. When \code{na.rm} is \code{TRUE} and a
list element contains only \code{NA} values, the resulting string
will be \code{""}.}

\item{keepFactors}{\code{logical} only used when \code{useLegacy=TRUE} and
\code{doSort=TRUE}; indicating whether to preserve factors,
keeping factor level order. When
\code{keepFactors=TRUE}, if any list element is a \code{factor}, all elements
are converted to factors. Note that this step combines overall
factor levels, and non-factors will be ordered using
\code{base::order()} instead of \code{jamba::mixedOrder()} (for now.)}

\item{checkClass}{\code{logical}, default TRUE, whether to check the class
of each vector in the input list.
\itemize{
\item When TRUE, it confirms the class of each element in the \code{list}
before processing, to prevent conversion which may otherwise
lose information.
\item For all cases when a known vector is split into a \code{list},
\code{checkClass=FALSE} is preferred since there is only one class
in the resulting \code{list} elements. This approach is faster
especially for for large input lists, 10000 or more.
\item When \code{checkClass=FALSE} it assumes all entries can be
coerced to \code{character}, which is fastest, but does not preserve
factor levels due to R coersion methods used by \code{unlist()}.
}}

\item{useBioc}{\code{logical} indicating whether this function should try
to use \code{S4Vectors::unstrsplit()} when the Bioconductor package
\code{S4Vectors} is installed, otherwise it will use a less
efficient \code{mapply()} operation.}

\item{...}{additional arguments are passed to \code{mixedOrder()} when
\code{doSort=TRUE}.}
}
\description{
Paste a list of vectors into a character vector, with unique values
sorted then delimited by default with a comma.
}
\details{
This function is convenient a wrapper for \code{cPaste(..., doSort=TRUE, makeUnique=TRUE)}.
}
\seealso{
Other jam list functions: 
\code{\link{cPaste}()},
\code{\link{cPasteS}()},
\code{\link{cPasteU}()},
\code{\link{cPasteUnique}()},
\code{\link{heads}()},
\code{\link{jam_rapply}()},
\code{\link{list2df}()},
\code{\link{mergeAllXY}()},
\code{\link{mixedSorts}()},
\code{\link{rbindList}()},
\code{\link{relist_named}()},
\code{\link{rlengths}()},
\code{\link{sclass}()},
\code{\link{sdim}()},
\code{\link{uniques}()},
\code{\link{unnestList}()}
}
\concept{jam list functions}
