% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{deg2rad}
\alias{deg2rad}
\title{Convert degrees to radians}
\usage{
deg2rad(x, ...)
}
\arguments{
\item{x}{\code{numeric} vector, expected to be degree values between zero
and 360.}

\item{...}{other parameters are ignored.}
}
\value{
\code{numeric} vector after coverting degrees to radians.
}
\description{
Convert degrees to radians
}
\details{
This function simply converts degrees which range from 0 to 360,
into radians which range from zero to pi*2.
}
\examples{
deg2rad(rad2deg(c(pi*2, pi/2)))/pi;

}
\seealso{
Other jam numeric functions: 
\code{\link{noiseFloor}()},
\code{\link{normScale}()},
\code{\link{rad2deg}()},
\code{\link{rowGroupMeans}()},
\code{\link{rowRmMadOutliers}()},
\code{\link{warpAroundZero}()}
}
\concept{jam numeric functions}
