% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-diff-fns.R
\name{diff_functions}
\alias{diff_functions}
\title{Compare two functions line by line using diff}
\usage{
diff_functions(
  f1,
  f2,
  width = NULL,
  wide = TRUE,
  trim_whitespace = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{f1}{\code{function} expected to be passed by function name.}

\item{f2}{\code{function} expected to be passed by function name.}

\item{width}{\code{integer} number of columns to allow in output.}

\item{wide}{\code{logical} indicating whether to use wide output
from \code{diff}, which prints the two functions side by side.}

\item{trim_whitespace}{\code{logical} indicating whether to trim leading
and trailing whitespace on each function line to avoid differences
defined only by the indentation of text.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
None. This function is called for its output, which
can be captured with \code{utils::capture.output()}.
}
\description{
Compare two functions line by line using diff
}
\details{
This function is a utility wrapper around linux commandline
tool \code{diff}, used to compare the text representation of two
R functions.
}
\examples{
diff_functions(base::grepl, base::agrepl)

}
\seealso{
Other jam practical functions: 
\code{\link{breakDensity}()},
\code{\link{checkLightMode}()},
\code{\link{check_pkg_installed}()},
\code{\link{colNum2excelName}()},
\code{\link{color_dither}()},
\code{\link{exp2signed}()},
\code{\link{fileInfo}()},
\code{\link{fixYellow}()},
\code{\link{getAxisLabel}()},
\code{\link{handleArgsText}()},
\code{\link{heads}()},
\code{\link{isFALSEV}()},
\code{\link{isTRUEV}()},
\code{\link{jamba}},
\code{\link{jargs}()},
\code{\link{kable_coloring}()},
\code{\link{lldf}()},
\code{\link{log2signed}()},
\code{\link{make_html_styles}()},
\code{\link{make_styles}()},
\code{\link{match_unique}()},
\code{\link{mergeAllXY}()},
\code{\link{minorLogTicks}()},
\code{\link{newestFile}()},
\code{\link{printDebugI}()},
\code{\link{printDebug}()},
\code{\link{renameColumn}()},
\code{\link{rmInfinite}()},
\code{\link{rmNAs}()},
\code{\link{rmNA}()},
\code{\link{rmNULL}()},
\code{\link{sclass}()},
\code{\link{sdima}()},
\code{\link{sdim}()},
\code{\link{setPrompt}()},
\code{\link{ssdima}()},
\code{\link{ssdim}()}
}
\concept{jam practical functions}
