% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-plots.r
\name{drawLabels}
\alias{drawLabels}
\title{Draw text labels on a base R plot}
\usage{
drawLabels(
  txt = NULL,
  newCoords = NULL,
  x = NULL,
  y = NULL,
  lx = NULL,
  ly = NULL,
  segmentLwd = 1,
  segmentCol = "#00000088",
  drawSegments = TRUE,
  boxBorderColor = "#000000AA",
  boxColor = "#DDAA77",
  boxLwd = 1,
  drawBox = TRUE,
  drawLabels = TRUE,
  font = 1,
  labelCex = 0.8,
  boxCexAdjust = 1.9,
  labelCol = alpha2col(alpha = 0.8, setTextContrastColor(boxColor)),
  doPlot = TRUE,
  xpd = NA,
  preset = "default",
  adjPreset = "default",
  adjX = 0.5,
  adjY = 0.5,
  trimReturns = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{txt}{character vector of labels, length equal to \code{x} and \code{y}.}

\item{newCoords}{optional \code{data.frame} typically as a result of
a previous call to \code{drawLabels()}. In general, it should contain
colnames equivalent to the function parameters of \code{drawLabels()}.}

\item{x, y}{vector of x- and y- coordinates.}

\item{lx, ly}{optional vector of segment endpoint coordinates, used
to draw a line from x,y coordinates to the segment lx,ly coordinate.}

\item{segmentLwd, segmentCol}{vector of segment line widths, and colors,
respectively. Each vector will be recycled to \code{length(txt)} as
needed.}

\item{drawSegments}{logical whether to draw segments, where applicable.}

\item{boxBorderColor}{vector of colors used for the box border around
each label.}

\item{boxColor}{vector of colors used for the box background behind
each label.}

\item{boxLwd}{vector of box line widths, sent to \code{graphics::rect()},
this vector will be recycled to \code{length(txt)}.}

\item{drawBox}{logical whether to draw boxes behind each text label.}

\item{drawLabels}{logical whether to draw each text label.}

\item{font}{vector of font values as described in \code{graphics::par()},
where 1=normal, 2=bold, 3=italics, 4=bold-italics.}

\item{labelCex}{vector of cex values used for text labels. This vector
will be recycled to \code{length(txt)} as needed.}

\item{boxCexAdjust}{numeric vector length=2, used to expand the x-width
and y-height of the box around around text labels.}

\item{labelCol}{vector of label colors, by default it calls
\code{jamba::setTextContrastColor()} to generate a color to contrast
the background box color.}

\item{doPlot}{logical whether to perform any plot operations. Set
\code{FALSE} to calculate coordinates and return a \code{data.frame} of
label coordinates, which can then be manipulated before calling
\code{drawLabels()} again.}

\item{xpd}{value compatible with \code{par("xpd")}, where NA allows labels
anywhere in the device region, TRUE retricts labels within the figure
region, and FALSE restricts labels within the plot region.}

\item{preset}{vector of text values used to position text labels relative
to the x,y coordinate, where "topleft" will position the label so the
entire label box is top-left of the point, therefore the point will be
at the bottom-right corner of the label box. When \code{preset} is anything
by \code{"none"} the \code{adjX} and \code{adjY} values are ignored.}

\item{adjX, adjY}{the text adjustment of labels relative to the x,y
coordinate. The values are recycled to \code{length(txt)}.}

\item{trimReturns}{logical whether to trim leading and trailing return
(newline) characters from labels.}

\item{verbose}{logical whether to print verbose output.}

\item{...}{additional arguments are passed to \code{graphics::segments()} when
segments are drawn, to \code{graphics::rect()} when label boxes are drawn,
and to \code{graphics::text()} when text labels are drawn.}
}
\value{
invisible data.frame containing label coordinates used
to draw labels. This data.frame can be manipulated and provided
as input to \code{drawLabels()} for subsequent customized label
positioning.
}
\description{
Draw text labels on a base R plot
}
\details{
This function takes a vector of coordinates and text labels,
and draws the labels with colored rectangles around each label
on the plot. Each label can have unique font, cex, and color,
and are drawn using vectorized operations.

TODO: In future allow rotated text labels. Not that useful within
a plot panel, but sometimes useful when draw outside a plot, for
example axis labels.
}
\examples{
nullPlot();
drawLabels(x=par("usr")[1],
   y=par("usr")[4],
   txt="Top-left\nof plot",
   preset="topleft",
   boxColor="blue4");

drawLabels(x=par("usr")[2],
   y=par("usr")[3],
   txt="Bottom-right\nof plot",
   preset="bottomright",
   boxColor="green4");

drawLabels(x=mean(par("usr")[1:2]),
   y=mean(par("usr")[3:4]),
   txt="Center\nof plot",
   preset="center",
   boxColor="purple");

points(x=c(par("usr")[1], par("usr")[2],
      mean(par("usr")[1:2])),
   y=c(par("usr")[4], par("usr")[3],
      mean(par("usr")[3:4])),
   pch=20,
   col="red",
   xpd=NA);

}
\seealso{
Other jam plot functions: 
\code{\link{adjustAxisLabelMargins}()},
\code{\link{coordPresets}()},
\code{\link{decideMfrow}()},
\code{\link{getPlotAspect}()},
\code{\link{groupedAxis}()},
\code{\link{imageByColors}()},
\code{\link{imageDefault}()},
\code{\link{minorLogTicksAxis}()},
\code{\link{nullPlot}()},
\code{\link{plotPolygonDensity}()},
\code{\link{plotRidges}()},
\code{\link{plotSmoothScatter}()},
\code{\link{shadowText}()},
\code{\link{showColors}()},
\code{\link{smoothScatterJam}()},
\code{\link{sqrtAxis}()},
\code{\link{usrBox}()}
}
\concept{jam plot functions}
