% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{fileInfo}
\alias{fileInfo}
\title{file information in data.frame format}
\usage{
fileInfo(fileList, ...)
}
\arguments{
\item{fileList}{character vector with one or more file paths.}
}
\value{
data.frame with file information, including "filename" and "size"
as additional colnames as compared to \code{file.info} output.
}
\description{
file information in data.frame format
}
\details{
This function is a minor extension to \code{file.info} in that it
adds the filename as a proper colname, and "size" which contains a text
file size.
}
