% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{fixYellow}
\alias{fixYellow}
\title{Fix yellow color}
\usage{
fixYellow(col, Hrange = c(80, 90), Hshift = -15, ...)
}
\arguments{
\item{col}{R color, either in hex color format or using values from
\code{colors()}.}

\item{Hrange}{numeric vector whose range defines the region of hues
to be adjusted. By default hues between 80 and 90 are adjusted. If
NULL, \code{HCL} is return unchanged.}

\item{Hshift}{numeric value length one, used to adjust the hue of colors
within the range \code{Hrange}. If NULL, \code{HCL} is return unchanged.}

\item{...}{additional arguments are passed to \code{col2hcl()}, and
\code{hcl2col()}.}
}
\value{
returns a vector of R colors the same length as input \code{col}.
In the event \code{col}, \code{Hrange}, or \code{Hshift} have length 0, or if any
step in the conversion produces length 0, then the
original \code{col} is returned.
}
\description{
Fix yellow color to be less green than default "yellow"
}
\details{
This function "fixes" the color yellow, which by default appears green
especially when darkened. The effect of this function is to make yellows
appear more red, which appears more visibly yellow even when the color
is darkened.

This function is intended to be tolerant to missing values. For example if
any of the values \code{col}, \code{Hrange}, or \code{Hshift} are length 0, the original
\code{col} is returned unchanged.
}
\examples{
yellows <- vigrep("yellow", colors());
fixedYellows <- fixYellow(yellows);
showColors(list(yellows=yellows,
   fixedYellows=fixedYellows));

}
