% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-plots.r
\name{getAxisLabel}
\alias{getAxisLabel}
\title{Get axis label for minorLogTicks}
\usage{
getAxisLabel(i, asValues, logAxisType = c("normal", "flip", "pvalue"),
  logBase, base_limit = 2, offset = 0, symmetricZero = (offset > 0),
  ...)
}
\description{
Get axis label for minorLogTicks
}
\details{
This function is intended to be called internally by
\code{jamba::minorLogTicks()}.
}
