% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{getPlotAspect}
\alias{getPlotAspect}
\title{Get aspect ratio for coordinates, plot, or device}
\usage{
getPlotAspect(type = c("coords", "plot", "device"), parUsr = par("usr"),
  parPin = par("pin"), parDin = par("din"), ...)
}
\description{
Get aspect ratio for coordinates, plot, or device
}
