% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{groupedAxis}
\alias{groupedAxis}
\title{Draw groups axis labels}
\usage{
groupedAxis(side = 1, x, group_style = c("partial_grouped", "grouped",
  "centered"), las = 2, returnFractions = TRUE, nudge = 0.2,
  do_abline = FALSE, abline_lty = "solid", abline_col = "grey40",
  ...)
}
\description{
Draw grouped axis labels given a character vector.
}
\details{
This function extends \code{breaksByVector()} specifically for
axis labels. It is intended where character labels are spaced
at integer steps, and some labels are expected to be repeated.
}
\examples{
par("mar"=c(4,4,6,6));
b <- rep(LETTERS[1:5], c(2,3,5,4,3));
b2 <- c(b[1:2], makeNames(b[3:5]), b[6:16]);
nullPlot(doBoxes=FALSE,
   doUsrBox=TRUE,
   xlim=c(0,18),
   ylim=c(0,18));

groupedAxis(1, b);
groupedAxis(2, b, group_style="grouped");
groupedAxis(3, b2, do_abline=TRUE);
groupedAxis(4, b2, group_style="grouped");
mtext(side=1, "group_style='partial_grouped'", line=2, las=0);
mtext(side=2, "group_style='grouped'", line=2, las=0);
mtext(side=3, "group_style='partial_grouped'", line=2, las=0);
mtext(side=4, "group_style='grouped'", line=2, las=0);


}
\seealso{
Other jam plot functions: \code{\link{adjustAxisLabelMargins}},
  \code{\link{coordPresets}}, \code{\link{decideMfrow}},
  \code{\link{drawLabels}}, \code{\link{getPlotAspect}},
  \code{\link{imageByColors}}, \code{\link{imageDefault}},
  \code{\link{minorLogTicksAxis}}, \code{\link{nullPlot}},
  \code{\link{plotPolygonDensity}},
  \code{\link{plotSmoothScatter}},
  \code{\link{shadowText}}, \code{\link{showColors}},
  \code{\link{smoothScatterJam}}, \code{\link{sqrtAxis}},
  \code{\link{usrBox}}
}
\concept{jam plot functions}
