% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{hcl2col}
\alias{hcl2col}
\title{convert HCL to R color}
\usage{
hcl2col(x = NULL, H = NULL, C = NULL, L = NULL, ceiling = 255,
  maxColorValue = 255, alpha = NULL, fixup = NULL, ...)
}
\arguments{
\item{x}{matrix of colors, with rownames `"H"`, `"C"`, `"L"`, or if not
supplied it looks for vectors `H`, `C`, and `L` accordingly. It can
alternatively be supplied as an object of class `polarLUV`.}

\item{H, C, L}{numeric vectors supplied as an alternative to `x`, with
ranges 0 to 360, 0 to 100, and 0 to 100, respectively.}

\item{ceiling}{numeric value indicating the maximum values allowed for
`R`, `G`, and `B` after conversion by `colorspace::as(x, "RGB")`.
This ceiling is applied after the `maxColorValue` is used to scale
numeric values, and is intended to correct for the occurrence of
values above 255, which would be outside the typical color gamut
allowed for RGB colors used in R. In general, this value should not
be modified.}

\item{maxColorValue}{numeric value indicating the maximum RGB values,
typically scaling values to a range of 0 to 255, from the default
returned range of 0 to 1. In general, this value should not be
modified.}

\item{alpha}{optional vector of alpha values. If not supplied, and if
`x` is supplied as a matrix with rowname `"alpha"`, then values will
be used from `x["alpha",]`.}

\item{fixup}{boolean indicating whether to use
`colorspace::hex(...,fixup=TRUE)` for conversion to R hex colors,
**which is not recommended** since this conversion applies some
unknown non-linear transformation for colors outside the color gamut.
It is here is an option for comparison, and if specifically needed.}

\item{...}{other arguments are ignored.}
}
\value{
vector of R colors, or where the input was NA, then NA
   values are returned in the same order.
}
\description{
Convert an HCL color matrix to vector of R hex colors
}
\examples{
# Prepare a basic HCL matrix
hclM <- col2hcl(c(red="red",
   blue="blue",
   yellow="yellow",
   orange="#FFAA0066"));
hclM;

# Now convert back to R hex colors
colorV <- hcl2col(hclM);
colorV;

showColors(colorV);

}
