% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{hsv2col}
\alias{hsv2col}
\title{Convert HSV matrix to R color}
\usage{
hsv2col(hsvValue, ...)
}
\arguments{
\item{hsvValue}{\code{numeric} HSV matrix, with rownames c("h","s","v")
in any order, and optionally "alpha" rowname for alpha transparency.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{character} vector of R colors.
}
\description{
Converts a HSV color matrix to R hex color
}
\details{
This function augments the \code{grDevices::hsv()} function in that it handles
output from \code{grDevices::rgb2hsv()} or \code{col2hsv()}, sufficient to
run a series of conversion functions, e.g. \code{hsv2col(col2hsv("red"))}.
This function also maintains alpha transparency, which is not maintained
by the \code{grDevices::hsv()} function.
}
\examples{
# start with a color vector
# red and blue with partial transparency
colorV <- c("#FF000055", "#00339999");

# confirm the hsv matrix maintains transparency
col2hsv(colorV);

# convert back to the original color
hsv2col(col2hsv(colorV));

}
\seealso{
Other jam color functions:
\code{\link[=alpha2col]{alpha2col()}},
\code{\link[=applyCLrange]{applyCLrange()}},
\code{\link[=col2alpha]{col2alpha()}},
\code{\link[=col2hcl]{col2hcl()}},
\code{\link[=col2hsl]{col2hsl()}},
\code{\link[=col2hsv]{col2hsv()}},
\code{\link[=color2gradient]{color2gradient()}},
\code{\link[=fixYellow]{fixYellow()}},
\code{\link[=fixYellowHue]{fixYellowHue()}},
\code{\link[=getColorRamp]{getColorRamp()}},
\code{\link[=hcl2col]{hcl2col()}},
\code{\link[=hsl2col]{hsl2col()}},
\code{\link[=isColor]{isColor()}},
\code{\link[=kable_coloring]{kable_coloring()}},
\code{\link[=makeColorDarker]{makeColorDarker()}},
\code{\link[=rainbow2]{rainbow2()}},
\code{\link[=rgb2col]{rgb2col()}},
\code{\link[=setCLranges]{setCLranges()}},
\code{\link[=setTextContrastColor]{setTextContrastColor()}},
\code{\link[=showColors]{showColors()}},
\code{\link[=unalpha]{unalpha()}},
\code{\link[=warpRamp]{warpRamp()}}
}
\concept{jam color functions}
