% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{isColor}
\alias{isColor}
\title{detect valid R color}
\usage{
isColor(x, makeNamesFunc = c, ...)
}
\arguments{
\item{x}{character vector of potential R colors}

\item{makeNamesFunc}{function used to make names for the resulting vector}

\item{...}{additional parameters are ignored}
}
\value{
\code{logical} vector with length(x).
}
\description{
detect valid R color
}
\details{
This function determines whether each element in a vector is a valid R
color, based upon the R color names, valid hex color format, and the
word "transparent" which is valid as an R color.
}
\examples{
isColor(c("red", "blue", "beige", "#99000099", "#aa00ff", "#AAE", "bleh"))

}
\seealso{
Other jam color functions:
\code{\link[=alpha2col]{alpha2col()}},
\code{\link[=applyCLrange]{applyCLrange()}},
\code{\link[=col2alpha]{col2alpha()}},
\code{\link[=col2hcl]{col2hcl()}},
\code{\link[=col2hsl]{col2hsl()}},
\code{\link[=col2hsv]{col2hsv()}},
\code{\link[=color2gradient]{color2gradient()}},
\code{\link[=fixYellow]{fixYellow()}},
\code{\link[=fixYellowHue]{fixYellowHue()}},
\code{\link[=getColorRamp]{getColorRamp()}},
\code{\link[=hcl2col]{hcl2col()}},
\code{\link[=hsl2col]{hsl2col()}},
\code{\link[=hsv2col]{hsv2col()}},
\code{\link[=kable_coloring]{kable_coloring()}},
\code{\link[=makeColorDarker]{makeColorDarker()}},
\code{\link[=rainbow2]{rainbow2()}},
\code{\link[=rgb2col]{rgb2col()}},
\code{\link[=setCLranges]{setCLranges()}},
\code{\link[=setTextContrastColor]{setTextContrastColor()}},
\code{\link[=showColors]{showColors()}},
\code{\link[=unalpha]{unalpha()}},
\code{\link[=warpRamp]{warpRamp()}}
}
\concept{jam color functions}
