% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{isTRUEV}
\alias{isTRUEV}
\title{Vectorized isTRUE}
\usage{
isTRUEV(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{additional arguments are ignored}
}
\description{
Vectorized isTRUE
}
\details{
This function applies three criteria to an input vector, to
determine if each entry in the vector is TRUE:
\enumerate{
\item It must be class \code{logical}.
\item It must not be \code{NA}.
\item It must evaluate as \code{TRUE}.
}
}
