% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-package.R
\docType{package}
\name{jamba-package}
\alias{jamba}
\alias{jamba-package}
\title{jamba: Jam Base Methods}
\description{
The jamba package contains several jam base functions
which are re-usable for routine R analysis work, and are
important dependencies for other Jam R packages.
}
\details{
See the function reference for a complete list of functions.

The goal is to implement methods as lightweight as possible, so
so inclusion in an analysis workflow will not incur
a noticeable burden.
}
\section{plot functions}{

\itemize{
\item \code{plotSmoothScatter()} enhanced variant of \code{smoothScatter()}
to show more visual detail
\item \code{imageDefault()} enhanced rasterized \code{image()} with fixed aspect ratio
\item \code{imageByColors()} for \code{data.frame} of colors and optional
labels centered across repeated values.
\item \code{showColors()} color display for vector, list, color function,
or mixed formats.
\item \code{nullPlot()} blank plot that labels the current margin sizes
\item \code{minorLogTicksAxis()} log-scale axis ticks in base R with
custom log base, optional offset, e.g. \code{log2(1 + x)}
\item \code{shadowText()} base R text labels with shadow or outline or both,
also \code{shadowText_options()}.
\item \code{getPlotAspect()}, \code{decideMfrow()} convenience base R graphics.
}
}

\section{string functions}{

\itemize{
\item \code{mixedSort()}, \code{mixedOrder()}, \code{mixedSortDF()} - efficient
alphanumeric "version" sort, with options helpful for gene symbols.
\item \code{vgrep()}, \code{vigrep()}, \code{igrep()}, \code{vigrep()} fast grep wrappers
for value-return, case-insensitive search.
\item \code{provigrep()}, \code{proigrep()} - progressive, ordered grep to use
pattern matching to re-order a vector.
\item \code{makeNames()} create unique, versioned names with custom format
\item \code{nameVector()} apply names to vector dynamically
\item \code{nameVectorN()} vector of named names useful with \code{lapply()}.
\item \code{pasteByRow()}, \code{pasteByRowOrdered()} paste data.frame and matrix
values by row, skipping blanks, optional factor order.
\item \code{rbindList()} convert list to \code{matrix} or \code{data.frame}.
\item \code{tcount()} extends \code{table()} to sort by size and optional minimum
count filter.
}
}

\section{color functions}{

\itemize{
\item \code{rgb2col()}, \code{col2hcl()}, \code{col2hcl()}, \code{col2hsv()}, \code{hsv2col()} color
interconversion
\item \code{setTextContrastColor()} text contrast color per given background color
\item \code{getColorRamp()} catch-all to get named gradients, or expand
one or more colors to gradient.
\item \code{makeColorDarker()}, \code{color2gradient()} create light/dark altered
variants of colors.
\item \code{showColors()} versatile display of color as \code{character} vector or
\code{list} of \code{character} or \code{function}.
}
}

\section{miscellaneous helper functions}{

\itemize{
\item \code{printDebug()} colored text output to console, 'Rmarkdown', HTML
\item \code{kable_coloring()} colored \code{kableExtra::kable()} output for 'Rmarkdown'
\item \code{setPrompt()} colored R prompt
\item \code{deg2rad()}, \code{rad2deg()} interconvert degrees to radians.
\item \code{getDate()}, \code{asDate()}, \code{dateToDaysOld()} human-readable,
opinionated date formatting
\item \code{padString()}, \code{padInteger()} pad character or integer strings
\item \code{rmNA()}, \code{rmNULL()}, \code{rmInfinite()} remove or replace missing or NA
values with defined alternatives
}
}

\section{export and import functions}{

\itemize{
\item \code{readOpenxlsx()} import worksheets from 'xlsx' 'Excel' files.
\item \code{writeOpenxlsx()} export worksheets to 'xlsx' 'Excel' files with color,
formatting, and styling.
\item \code{reload_rmarkdown_cache()} load Rmarkdown cache data into an \code{environment}
for re-use.
\item \code{reload_qmd_cache()} load Quarto .qmd cache data into an \code{environment}
for re-use.
}
}

\section{Jam options}{

The \code{jamba} package recognizes some global options, but limits these
options to include only non-analysis options. For example, no global
option should change the numerical manipulation of data.
\itemize{
\item \code{jam.lightMode} - \code{logical} whether the R console or graphical
background is light or dark, \code{printDebug()} limits
the luminance range to maximize visual contrast.
\item \code{jam.Crange},\code{jam.Lrange} - numerical values used by \code{printDebug()}
to maximize visual contrast, used with \code{jam.lightMode}.
\item \code{jam.shadowColor},\code{jam.shadow.r},\code{jam.shadow.n},\code{jam.alphaShadow},
\code{jam.outline},\code{jam.alphaOutline} to customize details for \code{shadowText()},
see \code{shadowText_options()} for convenience.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://jmw86069.github.io/jamba/}
  \item Report bugs at \url{https://github.com/jmw86069/jamba/issues}
}

}
\author{
\strong{Maintainer}: James M. Ward \email{jmw86069@gmail.com} (\href{https://orcid.org/0000-0002-9510-2848}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item James M. Ward \email{jmw86069@gmail.com} (\href{https://orcid.org/0000-0002-9510-2848}{ORCID}) [copyright holder]
}

}
\keyword{internal}
