% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\docType{package}
\name{jamba}
\alias{jamba}
\alias{jamba-package}
\title{jamba: Jam Base Methods}
\description{
The jamba package contains several jam base functions
which are re-usable for routine R analysis work, and are
important dependencies for other Jam R packages.
}
\details{
The goal will be to
maintain these methods as lightweight as possible, so
their inclusion in an analysis workflow will not incur
a noticeable burden.
}
\section{miscellaneous helper functions}{

   fileInfo,
   padInteger,
   padString,
   dateToDaysOld,
   asDate,
   getDate,
   setPrompt,
   pasteByRow,
   breaksByVector,
   printDebug,
   rbindList,
   colNum2excelName,
   decideMfrow,
   getPlotAspect,
   tcount,
   make_styles,
   noiseFloor
}

\section{plot functions}{

   plotSmoothScatter,
   smoothScatterJam,
   imageDefault,
   nullPlot,
   usrBox,
   imageDefault,
   imageByColors
}

\section{string functions}{

   igrepHas,
   vigrep,
   vgrep,
   igrep,
   unigrep,
   unvigrep,
   provigrep,
   makeNames,
   nameVector,
   nameVectorN,
   rmNULL,
   rmNA,
   rmInfinite
}

\section{color functions}{

   setTextContrastColor,
   col2hcl,
   col2alpha,
   alpha2col,
   hsv2col,
   rgb2col,
   makeColorDarker,
   getColorRamp,
   isColor
}

