% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{makeColorDarker}
\alias{makeColorDarker}
\title{make R colors darker (or lighter)}
\usage{
makeColorDarker(hexColor, darkFactor = 2, sFactor = 1, fixAlpha = NULL,
  verbose = FALSE, keepNA = FALSE, useMethod = 1, ...)
}
\description{
Makes R colors darker or lighter based upon darkFactor
}
\details{
This function was originally intended to create border colors, or to
create slightly darker colors used for labels. It is also useful for
for making colors lighter, in adjusting color saturation up or down,
or applying alpha transparency during the same step.

Note when colors are brightened beyond value=1, the saturation is
gradually reduced in order to produce a visibly lighter color. The
saturation minimu is set to 0.2, to maintain at least some amount of
color.
}
