% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-middle.R
\name{middle}
\alias{middle}
\title{Return the middle portion of data similar to head and tail}
\usage{
middle(x, n = 10, evenly = TRUE, ...)
}
\arguments{
\item{x}{input data that can be subset}

\item{n}{\code{numeric} number of entries to return}

\item{evenly}{\code{logical} indicating whether to return evenly spaced
entries along the full length of \code{x}. When \code{evenly=FALSE} only
the middle \code{n} entries are returned.}

\item{...}{additional arguments are ignored.}
}
\value{
an object of class equivalent to \code{x}.
}
\description{
Return the middle portion of data similar to head and tail
}
\details{
This function is very simple, and is intended to mimic \code{head()}
and \code{tail()} to inspect data without looking at every value
}
\examples{
x <- 1:101;
middle(x);
middle(x, evenly=TRUE)

xdf <- data.frame(n=1:101,
   excel_colname=jamba::colNum2excelName(1:101));
middle(xdf)
middle(xdf, evenly=TRUE)

}
\seealso{
Other jam practical functions:
\code{\link[=breakDensity]{breakDensity()}},
\code{\link[=call_fn_ellipsis]{call_fn_ellipsis()}},
\code{\link[=checkLightMode]{checkLightMode()}},
\code{\link[=check_pkg_installed]{check_pkg_installed()}},
\code{\link[=colNum2excelName]{colNum2excelName()}},
\code{\link[=color_dither]{color_dither()}},
\code{\link[=exp2signed]{exp2signed()}},
\code{\link[=getAxisLabel]{getAxisLabel()}},
\code{\link[=isFALSEV]{isFALSEV()}},
\code{\link[=isTRUEV]{isTRUEV()}},
\code{\link[=jargs]{jargs()}},
\code{\link[=kable_coloring]{kable_coloring()}},
\code{\link[=lldf]{lldf()}},
\code{\link[=log2signed]{log2signed()}},
\code{\link[=minorLogTicks]{minorLogTicks()}},
\code{\link[=newestFile]{newestFile()}},
\code{\link[=printDebug]{printDebug()}},
\code{\link[=reload_qmd_cache]{reload_qmd_cache()}},
\code{\link[=reload_rmarkdown_cache]{reload_rmarkdown_cache()}},
\code{\link[=renameColumn]{renameColumn()}},
\code{\link[=rmInfinite]{rmInfinite()}},
\code{\link[=rmNA]{rmNA()}},
\code{\link[=rmNAs]{rmNAs()}},
\code{\link[=rmNULL]{rmNULL()}},
\code{\link[=setPrompt]{setPrompt()}}
}
\concept{jam practical functions}
