% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-plots.r
\name{minorLogTicksAxis}
\alias{minorLogTicksAxis}
\title{Display major and minor tick marks for log-scale axis}
\usage{
minorLogTicksAxis(ax = NULL, lims = NULL, logBase = 10,
  displayBase = logBase, labelValueOffset = 0,
  symmetricZero = (labelValueOffset > 0), majorCex = 1,
  minorCex = 0.65, majorN = 5, minorN = logBase, doMajor = TRUE,
  doLabels = TRUE, doMinorLabels = TRUE, asValues = TRUE,
  padj = NULL, doFormat = TRUE, big.mark = ",", scipen = 10,
  minorWhich = c(2, 5), cex = 1, las = 2, col = "black",
  col.ticks = col, minorLogTicksData = NULL, verbose = FALSE, ...)
}
\arguments{
\item{ax}{integer indicating the axis side, 1=bottom, 2=left,
3=top, 4=right.}

\item{lims}{NULL or numeric range for which the axis tick marks
will be determined. If NULL then the corresponding \code{par("usr")}
will be used.}

\item{logBase}{numeric value indicating the log base units, which
will be used similar to how \code{base} is used in \code{log(x, base)}.}

\item{displayBase}{numeric value indicating the log base units to
use when determining the numeric label position. For example,
data may be log2 scaled, and yet it is visually intuitive to
show log transformed axis units in base 10 units. See examples.}

\item{labelValueOffset}{numeric offset used in transforming the
numeric data displayed on this axis. For example, a common
technique is to transform data using \code{log2(1+x)} which adds
\code{1} to values prior to the log2 transformation. In this case,
\code{labelValueOffset=1}, which ensures the axis labels exactly
match the initial numeric value prior to the log2 transform.}

\item{symmetricZero}{logical indicating whether numeric values
are symmetric around zero. For example, log fold changes should
use \code{symmetricZero=TRUE} which ensures a log2 value of \code{-2} is
labeled \code{-4} to indicate a negative four fold change. If
\code{symmetricZero=FALSE} a log2 value of \code{-2} would be labeled
\code{0.0625}.}

\item{majorCex, minorCex}{the base text size factors, relative
to cex=1 for default text size. These factors are applied in
addition to existing \code{par("cex")} values, preserving any
global text size defined there.}

\item{majorN, minorN}{numeric values indicating the number of
axis labels, similar to how \code{n} is used by the function
\code{base::pretty()} to guide the number of labels to display.}

\item{padj}{numeric vector length 2, which is used to position
axis labels for the minor and major labels, respectively. For
example, \code{padj=c(0,1)} will position minor labels just to the
left of the tick marks, and major labels just to the right
of tick marks. This example is helpful when minor labels bunch
up on the right side of each section.}

\item{doFormat}{logical indicating whether to apply \code{base::format()} to
format numeric labels.}

\item{big.mark, scipen}{parameters passed to \code{base::format()} when
\code{doFormat=TRUE}.}

\item{minorWhich}{integer vector indicating which of the minor tick
marks should be labeled. Labels are generally numbered from \code{2}
to \code{displayBase-1}. So by default, log 10 units would add
minor tick marks and labels to the \code{c(2,5)} position. For log2
units only, the second label is defined at 1.5, which shows
minor labels at \code{c(3, 6, 12)}, which are \code{1.5 * c(2, 4, 8)}.}

\item{cex, col, col.ticks, las}{parameters used for axis label size,
axis label colors,
axis tick mark colors, and label text orientation, respectively.}

\item{minorLogTicksData}{a list object created by running
\code{jamba::minorLogTicks()}, which allows inspecting and modifying
the content for custom control.}

\item{verbose}{logical indicating whether to print verbose output.}
}
\value{
A list with vectors of majorLabels, majorTicks, minorLabels,
minorTicks, and allLabelsDF which is a \code{data.frame} containing
all axis tick positions, with corresponding labels.
}
\description{
Display major and minor tick marks for log-scale axis
}
\details{
This function displays log units on the axis of an
existing base R plot. It calls \code{minorLogTicks()} which
calculates appropriate tick and label positions. The
axis tick mark positions and labels are determined via
\code{jamba::minorLogTicks()}.

This function was motivated in order to label log-transformed
data properly in some special cases, like using \code{log2(1+x)}
where the resulting values are shifted "off by one" using
standard log-scaled axis tick marks and labels.

Also, when using log fold change values, this function
creates axis labels which indicate negative fold change
values, for example \code{-2} in log2 fold change units would
be labeled with fold change \code{-4}, and not \code{0.0625} which
represents a fractional value.

Use the argument \code{symmetricZero=TRUE} when using directional
log fold change values.
}
