% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-string.r
\name{mixedOrder}
\alias{mixedOrder}
\title{order alphanumeric values keeping numeric values in proper order}
\usage{
mixedOrder(x, ..., blanksFirst = TRUE, NAlast = TRUE,
  keepNegative = FALSE, keepInfinite = FALSE, keepDecimal = FALSE,
  verbose = FALSE, ignore.case = TRUE, useCaseTiebreak = TRUE,
  returnDebug = FALSE, returnType = c("order", "rank"))
}
\arguments{
\item{x}{input vector}

\item{...}{additional parameters are sent to \code{\link{mixedOrder}}.}

\item{blanksFirst}{logical whether to order blank entries before entries
containing a value.}

\item{NAlast}{logical whether to move NA entries to the end of the sort.}

\item{keepNegative}{logical whether to keep '-' associated with adjacent
numeric values, in order to sort them as negative values.}

\item{keepInfinite}{logical whether to allow "Inf" to be considered
a numeric infinite value.}

\item{keepDecimal}{logical whether to keep the decimal in numbers,
sorting as a true number and not as a version number. By default
keepDecimal=FALSE, which means "v1.200" should be ordered before
"v1.30". When keepDecimal=TRUE, the numeric sort considers only
"1.2" and "1.3" and sorts in that order.}

\item{verbose}{logical whether to print verbose output.}

\item{ignore.case}{logical whether to ignore uppercase and lowercase
characters when defining the sort order.}

\item{sortByName}{logical whether to sort the vector x by names(x) instead
of sorting by x itself.}
}
\value{
integer vector of orders derived from x
}
\description{
order alphanumeric values keeping numeric values in proper order
}
\details{
This function is a refactor of the \code{gtools::mixedorder}
function from the \code{gtools} package. It was extended to make it faster,
and to handle special cases slightly differently. It was driven by some
need to sort gene symbols, and miRNA symbols in numeric order, for example:
\describe{
\item{test set:}{miR-12,miR-1,miR-122,miR-1b,miR-1a,miR-2}
\item{\code{sort}:}{miR-1,miR-12,miR-122,miR-1a,miR-1b,miR-2}
\item{\code{gtools::mixedsort}:}{miR-122,miR-12,miR-2,miR-1,miR-1a,miR-1b}
\item{\code{mixedSort}:}{miR-1,miR-1a,miR-1b,miR-2,miR-12,miR-122}
}
The function does not by default recognize negative numbers as negative,
instead it treats '-' as a delimiter, unless keepNegative=TRUE.

This function also attempts to maintain '.' as part of a decimal number,
which can be problematic when sorting IP addresses, for example.

This function is also available as a sort function \code{\link{mixedSort}}.
}
\examples{
x <- c("miR-12","miR-1","miR-122","miR-1b", "miR-1a","miR-2");
mixedOrder(x);
x[mixedOrder(x)];
mixedSort(x);
order(x);
x[order(x)];
sort(x);

}
\seealso{
\code{gtools::mixedorder},
\code{gtools::mixedsort}

Other jam sort functions: \code{\link{mixedSortDF}},
  \code{\link{mixedSort}}, \code{\link{mmixedOrder}}

Other jam string functions: \code{\link{asSize}},
  \code{\link{breaksByVector}}, \code{\link{cPasteUnique}},
  \code{\link{cPaste}}, \code{\link{fillBlanks}},
  \code{\link{formatInt}}, \code{\link{makeNames}},
  \code{\link{mixedSortDF}}, \code{\link{mixedSort}},
  \code{\link{mmixedOrder}}, \code{\link{nameVectorN}},
  \code{\link{nameVector}}, \code{\link{padInteger}},
  \code{\link{padString}}, \code{\link{pasteByRowOrdered}},
  \code{\link{pasteByRow}}, \code{\link{tcount}},
  \code{\link{uniques}}
}
\concept{jam sort functions}
\concept{jam string functions}
