% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{noiseFloor}
\alias{noiseFloor}
\title{Apply noise floor to a numeric values}
\usage{
noiseFloor(x, minimum = 0, newValue = minimum, adjustNA = FALSE,
  ceiling = NULL, newCeiling = ceiling, ...)
}
\arguments{
\item{x}{numeric vector or matrix}

\item{minimum}{numeric floor value}

\item{newValue}{numeric, by default the same as the floor value. Sometimes
it can be useful to define a different value, one example is to define
values as NA, or another distinct number away from the floor.}

\item{adjustNA}{logical whether to change NA values to the newValue.}

\item{ceiling}{numeric value, optionally a ceiling. If defined, then values
above the ceiling value are set to newCeiling.}

\item{newCeiling}{numeric value when ceiling is defined, values above the
ceiling are set to this numeric value.}

\item{...}{additional parameters are ignored.}
}
\description{
Apply noise floor, setting numeric values below the floor to a minimum
value.
}
\details{
Noise floors are useful when detected numeric values are sometimes below
a clear noise threshold, and where some downstream ratio may be calculated
using these values. Applying a noise floor ensures the ratios and not
artificially higher, ironically in cases where the values involved are
least reliable. They can be said to help produce more conservative
ratios, when using noise floored values.
}
