% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{normScale}
\alias{normScale}
\title{Scale a numeric vector from 0 to 1}
\usage{
normScale(x, from = 0, to = 1, low = min(x, na.rm = TRUE), high = max(x,
  na.rm = TRUE), naValue = NA, singletMethod = c("mean", "min", "max"), ...)
}
\arguments{
\item{x}{numeric vector.}

\item{from}{the minimum numeric value to re-scale the input numeric vector.}

\item{to}{the maximum numeric value to re-scale the input numeric vector.}

\item{low}{numeric value defining the low end of the input numeric range,
intended when input values might not contain the entire numeric
range to be re-scaled.}

\item{high}{numeric value defining the high end of the input numeric range,
intended when input values might not contain the entire numeric
range to be re-scaled.}

\item{naValue}{optional numeric value used to replace \code{NA}, usually by
replacing \code{NA} with zero.}

\item{singletMethod}{character value describing how to handle singlet
input values, for example how to scale the number 5 by itself.
If "mean" then it uses the average of \code{from} and \code{to}, "min" uses
the \code{from} value, and "max" uses the \code{to} value.}

\item{...}{additional parameters are ignored.}
}
\description{
Scale a numeric vector from 0 to 1
}
\details{
This function is intended as a quick way to scale numeric values
between 0 and 1, however other ranges can be defined as needed.

NA values are ignored and will remain NA in the output. To handle
NA values, use the \code{rmNA()} function, which can optionally replace
NA with a fixed numeric value.
}
\examples{
# Notice the first value 1 is re-scaled to 0
normScale(1:11);

# Here the low value is defined as 0, even though the lowest input is 1
normScale(1:10, low=0);

normScale(c(10,20,40,30), from=50, to=65);

}
