% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{rad2deg}
\alias{rad2deg}
\title{Convert radians to degrees}
\usage{
rad2deg(x, ...)
}
\arguments{
\item{x}{numerical vector, expected to be radian values between zero
and pi*2.}

\item{...}{other parameters are ignored.}
}
\description{
Convert radians to degrees
}
\details{
This function simply converts radians which range from zero to pi*2,
into degrees which range from 0 to 360.
}
\examples{
rad2deg(c(pi*2, pi/2))

}
\seealso{
\code{rad2deg},\code{deg2rad}
}
