% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-load-qmd-cache.R
\name{reload_qmd_cache}
\alias{reload_qmd_cache}
\title{Reload Quarto 'qmd' cache}
\usage{
reload_qmd_cache(
  dir = ".",
  qmd_file = NULL,
  maxnum = 1000,
  max_cache_name = NULL,
  envir = new.env(),
  file_sort = c("globals", "objects", "ctime", "mtime"),
  preferred_load_types = c("lazyLoad", "load"),
  dryrun = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dir}{\code{character} path to the directory that contains 'Rmarkdown'
cache files. Each file is recognized by the file extension \code{".rdx"}.}

\item{qmd_file}{\code{character} path to the source '.qmd' file which
provides the order of chunks used. The chunks are expected to
have lines that begin 'r chunk_name'.
In theory, one could
create a file with only such lines, and it would be sufficient
to supply the order of chunks to be used.}

\item{maxnum}{\code{integer} indicating the maximum number of cache files
to re-load, in order.}

\item{max_cache_name}{\code{character} optional string indicating the
name of an 'Rmarkdown' cache chunk where this function will stop
loading cache data. All cache files after this point will
not be loaded. This option is intended to help recreate the
data available to a particular 'Rmarkdown' chunk during
processing.}

\item{envir}{\code{environment} in which data is populated, default \code{new.env()}
creates a new \code{environment} which is returned invisibly.}

\item{file_sort}{\code{character} string indicating how to sort cache files,
default uses best available, in order: "globals" (global index file),
"objects" (object index file), "ctime" (creation time),
"mtime" (modification time). The global index is preferred, and
other options are intended for rare scenarios when the global
index is not available. The methods using \code{"mtime"} or \code{"ctime"}
\strong{is less accurate}, yet may be sufficient for simple output.
\itemize{
\item \code{"globals"} uses the \code{"__globals"} file in the cache directory.
\item \code{"objects"} uses the \code{"__objects"} file in the cache directory.
\item \code{ctime} sorts by file creation time
\item \code{mtime} sorts by file modification time
}}

\item{preferred_load_types}{\code{character} with permitted cache load
types, in order of preference.
\itemize{
\item 'lazyLoad': Reloads the R objects as result of each chunk
using corresponding 'rdx' files. For Quarto documents, this option
is the only reliable option to access the stored R objects.
\item 'load': Reloads the 'RData' files associated with each chunk,
although in the case of Quarto documents, the 'RData' contains
a representation of the R code used to generate the R objects,
and not the R objects.
\item \strong{Important:} The 'load' operation typically used with Rmarkdown
to reload via 'RData' files is not supported for reloading the
R data environment. Use 'lazyLoad'.
}}

\item{dryrun}{\code{logical} default FALSE, whether to perform a dry-run,
which prints messages and does not load the data.}

\item{verbose}{\code{logical} default TRUE, whether to print verbose output.
This argument is not passed to other functions.}

\item{...}{additional arguments are passed to \code{lazyLoad()} or \code{load()}
as relevant to the method used to re-load the cache object data.}
}
\value{
\code{envir} is returned invisibly, with data objects populated
into that \code{environment}.
}
\description{
Reload Quarto 'qmd' cache by chunk order, into an
R environment
}
\details{
This function is intended to help re-load Quarto 'qmd' cache
created during the processing/rendering of a 'qmd' file.
It is intended to be roughly equivalent to
\code{reload_rmarkdown_cache()} with specific changes to handle
Quarto rendering, with the following requirements:
\enumerate{
\item It requires the Quarto \code{'.qmd'} or \code{'.Qmd'} file used
to produce the document.
\itemize{
\item When using default file paths, the 'dir' argument
is used to detect the corresponding 'qmd_file'.
For example \code{dir='document.name_cache/html/'} implies
\code{qmd_file='document.name.qmd'}.
\item The '.qmd' file is necessary to establish the
correct order of each output chunk. The Quarto output no
longer includes an index file with cache files in order.
\item The order of R chunk names is parsed, therefore must
follow basic formatting 'r chunkname, ...'
or 'r, chunkname, chunk.options, ...' for example.
\item Chunks display a modified time stamp, to help recognize
when the caching may have been refreshed out of order
compared to the Quarto document.
}
\item All chunks must be cached.
\itemize{
\item Any chunk with no R objects defined will indicate '(no-file)'.
\item The '(no-file)' indicates either (1) the chunk option
'cache=FALSE' or (2) the chunk made no changes to the R
environment.
}
\item It will only reload cache chunks that produce R objects.
\itemize{
\item Some chunks do not produce visible output R objects,
and therefore are not associated with a cache \code{.rdx} or \code{.RData}
cache file.
}
\item Context steps, including \code{options()} and \code{par()} are not applied.
\itemize{
\item The cache files only appear to represent R objects produced
or modified in each chunk.
}
}

When \code{dir} is provided, it can often be used to infer the source
'.qmd' file. For example:
\itemize{
\item \code{dir='some_long_filename_cache/html/'}
can be used to infer the original file:
\item \code{qmd_file='some_long_filename.qmd'}
}

This function is experimental because Quarto is not intended
to provide the final R environment. Instead, Quarto prioritizes
running the steps in a new, clean R environment, and does
not then return the resulting environment to the user
session.

An official mechanism to reload
the exact R environment would be to run \code{knitr::purl()} which
extracts the R code from a Quarto or Rmarkdown document and
evaluates the code completely. In \emph{most scenarios}, running
\code{reload_qmd_cache()} should be sufficient to reload the resulting
cached R objects suitable enough for continued exploration
or review of the data.

By default, all cached R objects are loaded into the
environment defined by \code{envir}, However,
\strong{it is recommended that \code{envir} is used to define a new environment}
into which the cached session is loaded.

\if{html}{\out{<div class="sourceCode">}}\preformatted{cache_env <- new.env()
reload_qmd_cache(cachedir, envir=cache_env)
}\if{html}{\out{</div>}}

From then on, the cached data objects can be seen with \code{ls(cache_env)}
and retrieved with \code{get("objectname", envir=cache_env)}.

If supplied with \code{maxnum} or \code{max_cache_name} then the cache
will be loaded only up to this point, and not beyond.
The recommended method to determine the cache is to use \code{dryrun=TRUE}
to view all sections, then to choose the \code{integer} number, or
\code{character} name to define the maximum chunk to load.
\subsection{Todo}{
\itemize{
\item Evaluate how best to handle 'RData' file input via the 'load'
cache type. Currently it returns each environment, each chunk
defines one hidden R object with the R code to be evaluated
by each chunk.
}
}
}
\seealso{
Other jam practical functions:
\code{\link[=breakDensity]{breakDensity()}},
\code{\link[=call_fn_ellipsis]{call_fn_ellipsis()}},
\code{\link[=checkLightMode]{checkLightMode()}},
\code{\link[=check_pkg_installed]{check_pkg_installed()}},
\code{\link[=colNum2excelName]{colNum2excelName()}},
\code{\link[=color_dither]{color_dither()}},
\code{\link[=exp2signed]{exp2signed()}},
\code{\link[=getAxisLabel]{getAxisLabel()}},
\code{\link[=isFALSEV]{isFALSEV()}},
\code{\link[=isTRUEV]{isTRUEV()}},
\code{\link[=jargs]{jargs()}},
\code{\link[=kable_coloring]{kable_coloring()}},
\code{\link[=lldf]{lldf()}},
\code{\link[=log2signed]{log2signed()}},
\code{\link[=middle]{middle()}},
\code{\link[=minorLogTicks]{minorLogTicks()}},
\code{\link[=newestFile]{newestFile()}},
\code{\link[=printDebug]{printDebug()}},
\code{\link[=reload_rmarkdown_cache]{reload_rmarkdown_cache()}},
\code{\link[=renameColumn]{renameColumn()}},
\code{\link[=rmInfinite]{rmInfinite()}},
\code{\link[=rmNA]{rmNA()}},
\code{\link[=rmNAs]{rmNAs()}},
\code{\link[=rmNULL]{rmNULL()}},
\code{\link[=setPrompt]{setPrompt()}}
}
\concept{jam practical functions}
