% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-string.r
\name{renameColumn}
\alias{renameColumn}
\title{Rename columns in a data.frame or matrix}
\usage{
renameColumn(x, from, to, verbose = FALSE, ...)
}
\arguments{
\item{x}{data.frame or matrix}

\item{from}{character vector of colnames expected to be in \code{x}.
Any values that do not match \code{colnames(x)} are ignored.}

\item{to}{character vector with \code{length(to) == length(from)}
corresponding to the target name for any colnames that
match \code{from}.}
}
\value{
data.frame (or tibble or DataFrame) with columns \code{from}
renamed to \code{to} as matched. When \code{colnames(x)} do not match
values in \code{from} the corresponding values in \code{from} are ignored.
}
\description{
Rename columns in a data.frame or matrix
}
\details{
This function is intended to rename one or more columns in a
data.frame or matrix. It will gracefully ignore columns
which do not match, in order to make it possible to call the
function multiple times without problem. It makes it possible
to verify that columns are properly named by re-running the
function as needed.

This function will also recognize input objects \code{GRanges},
\code{ucscData}, and \code{IRanges}, which store annotation in \code{DataFrame}
accessible via \code{values(x)}.
}
\examples{
df <- data.frame(A=1:5, B=6:10, C=11:15);
df;
df2 <- renameColumn(df,
   from=c("A","C"),
   to=c("a_new", "c_new"));
df2;
df3 <- renameColumn(df2,
   from=c("A","C","B"),
   to=c("a_new", "c_new","b_new"));
df3;

}
\seealso{
Other jam practical functions: \code{\link{applyCLrange}},
  \code{\link{breakDensity}}, \code{\link{checkLightMode}},
  \code{\link{colNum2excelName}}, \code{\link{exp2signed}},
  \code{\link{fileInfo}}, \code{\link{fixYellowHue}},
  \code{\link{fixYellow}}, \code{\link{getAxisLabel}},
  \code{\link{handleArgsText}}, \code{\link{isFALSEV}},
  \code{\link{isTRUEV}}, \code{\link{jamba}},
  \code{\link{jargs}}, \code{\link{kable_coloring}},
  \code{\link{log2signed}}, \code{\link{make_styles}},
  \code{\link{mergeAllXY}}, \code{\link{minorLogTicks}},
  \code{\link{newestFile}}, \code{\link{printDebugI}},
  \code{\link{printDebug}}, \code{\link{rmInfinite}},
  \code{\link{rmNA}}, \code{\link{rmNULL}},
  \code{\link{sclass}}, \code{\link{sdima}},
  \code{\link{sdim}}, \code{\link{setCLranges}},
  \code{\link{setPrompt}}, \code{\link{ssdima}},
  \code{\link{ssdim}}
}
\concept{jam practical functions}
