% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-string.r
\name{rmInfinite}
\alias{rmInfinite}
\title{remove Infinite values}
\usage{
rmInfinite(x, infiniteValue = NULL, ...)
}
\arguments{
\item{x}{vector input}

\item{infiniteValue}{NULL to remove Infinite values, or a replacement value}

\item{...}{additional parameters are ignored}
}
\value{
numeric vector with infinite values either removed, or
replaced with the supplied value.
}
\description{
remove Infinite values
}
\details{
This function removes any positive or negative infinite numerical
values, optionally replacing them with a given value or NA.
}
\examples{
rmInfinite(c(1, 5, 4, 10, Inf, 1, -Inf))

rmInfinite(c(1, 5, 4, 10, Inf, 1, -Inf), infiniteValue=1000)

}
\seealso{
Other jam practical functions:
\code{\link[=breakDensity]{breakDensity()}},
\code{\link[=call_fn_ellipsis]{call_fn_ellipsis()}},
\code{\link[=checkLightMode]{checkLightMode()}},
\code{\link[=check_pkg_installed]{check_pkg_installed()}},
\code{\link[=colNum2excelName]{colNum2excelName()}},
\code{\link[=color_dither]{color_dither()}},
\code{\link[=exp2signed]{exp2signed()}},
\code{\link[=getAxisLabel]{getAxisLabel()}},
\code{\link[=isFALSEV]{isFALSEV()}},
\code{\link[=isTRUEV]{isTRUEV()}},
\code{\link[=jargs]{jargs()}},
\code{\link[=kable_coloring]{kable_coloring()}},
\code{\link[=lldf]{lldf()}},
\code{\link[=log2signed]{log2signed()}},
\code{\link[=middle]{middle()}},
\code{\link[=minorLogTicks]{minorLogTicks()}},
\code{\link[=newestFile]{newestFile()}},
\code{\link[=printDebug]{printDebug()}},
\code{\link[=reload_qmd_cache]{reload_qmd_cache()}},
\code{\link[=reload_rmarkdown_cache]{reload_rmarkdown_cache()}},
\code{\link[=renameColumn]{renameColumn()}},
\code{\link[=rmNA]{rmNA()}},
\code{\link[=rmNAs]{rmNAs()}},
\code{\link[=rmNULL]{rmNULL()}},
\code{\link[=setPrompt]{setPrompt()}}
}
\concept{jam practical functions}
