% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{rowGroupMeans}
\alias{rowGroupMeans}
\title{Calculate row group means, or other statistics}
\usage{
rowGroupMeans(
  x,
  groups,
  na.rm = TRUE,
  useMedian = TRUE,
  rmOutliers = FALSE,
  madFactor = 5,
  returnType = c("output", "input"),
  rowStatsFunc = NULL,
  groupOrder = c("same", "sort"),
  keepNULLlevels = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{numeric data matrix}

\item{groups}{vector of group labels, either as a character
vector, or a factor. See the parameter \code{groupOrder} for
ordering of group labels in the output data matrix.}

\item{useMedian}{logical indicating whether the default
stat should be "mean" or "median".}

\item{rmOutliers}{logical indicating whether to apply outlier
detection and removal.}

\item{madFactor}{numeric value indicating the multiple of the
MAD value to define outliers. For example \code{madFactor=5}
will take the MAD value for a group multiplied by 5,
5\emph{MAD, as a threshold for outliers. So any points more than
5}MAD distance from the median per group are outliers.}

\item{returnType}{character value indicating the return data
type, \code{"output"} returns one summary stat value per group, per row;
\code{"input"} is useful when \code{rmOutliers=TRUE} in that it returns
a matrix with the same dimensions as the input, except with
outlier points replaced with NA.}

\item{rowStatsFunc}{optional function which takes a numeric matrix
as input, and returns a numeric vector equal to the number of
rows of the input data matrix. Examples: \code{base::rowMeans()},
\code{base::rowMedians()}, \code{matrixStats::rowMads}.}

\item{groupOrder}{character value indicating how character group
labels are ordered in the final data matrix, when \code{returnType="output"}.
Note that when \code{groups} is a factor, the factor levels are kept
in that order. Otherwise, \code{"same"} keeps groups in the same
order they appear in the input matrix; \code{"sort"} applies
\code{jamba::mixedSort()} to the labels.}

\item{keepNULLlevels}{logical indicating whether to keep factor
levels even when there are no corresponding columns in \code{x}.
When \code{TRUE} and \code{returnType="output"} the output matrix will
contain one colname for each factor level, with NA values used
to fill empty factor levels. This mechanism can be helpful to
ensure that output matrices have consistent colnames.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional parameters are passed to \code{rowStatsFunc},
and if \code{rmOutliers=TRUE} to \code{jamba::rowRmMadOutliers()}.}
}
\value{
When \code{returnType="output"} the output is a numeric matrix
with the same number of columns as the number of unique
\code{groups} labels. When \code{groups} is a factor and
\code{keepNULLlevels=TRUE}, the number of columns will be the
number of factor levels, otherwise it will be the number of
factor levels used in \code{groups}.

When \code{returnType="input"} the output is a numeric matrix
with the same dimensions as the input data. This output is
intended for use with \code{rmOutliers=TRUE} which will replace
outlier points with \code{NA} values. Therefore, this matrix can
be used to see the location of outliers.

The function also returns attributes that describe the
number of samples per group overall:
\describe{
\item{attr(out, "n")}{The attribute \code{"n"} is used to describe
the number of replicates per group.}
\item{attr(out, "nLabel")}{The attribute \code{"nLabel"} is
a simple text label in the form \code{"n=3"}.}
}
Note that when \code{rmOutliers=TRUE} the number of replicates per
group will vary depending upon the outliers removed. In that
case, remember that the reported \code{"n"} is always the total
possible columns available prior to outlier removal.
}
\description{
Calculate row group means, or other statistics
}
\details{
This function by default calculates group mean values
per row in a numeric matrix. However, the stat function
can be changed to calculate row medians, row MADs, etc.

An added purpose of this function is optional outlier
filtering, via calculation of MAD values and applying
a MAD threshold cutoff. The intention is to identify
technical outliers that otherwise adversely affect the
calculated group mean or median values. To inspect the
data after outlier removal, use the parameter \code{returnType="input"}
which will return the input data matrix with \code{NA}
substituted for outlier points. Outlier detection and
removal is performed by \code{jamba::rowRmMadOutliers()}.
}
\examples{
x <- matrix(ncol=9, rnorm(90));
colnames(x) <- LETTERS[1:9];
rowGroupMeans(x, groups=rep(letters[1:3], each=3))

}
\seealso{
Other jam numeric functions: 
\code{\link{deg2rad}()},
\code{\link{fix_matrix_ratio}()},
\code{\link{noiseFloor}()},
\code{\link{normScale}()},
\code{\link{rad2deg}()},
\code{\link{rowRmMadOutliers}()},
\code{\link{warpAroundZero}()}
}
\concept{jam numeric functions}
