% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{rowRmMadOutliers}
\alias{rowRmMadOutliers}
\title{Remove outlier points per row by MAD factor threshold}
\usage{
rowRmMadOutliers(
  x,
  madFactor = 5,
  na.rm = TRUE,
  minDiff = 0,
  includeAttributes = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{numeric matrix}

\item{madFactor}{numeric value to multiply by each row MAD
to define the threshold for outlier.}

\item{na.rm}{logical indicating whether to ignore NA values
when calculating the MAD value. It should probably always be
\code{TRUE}, however setting to \code{FALSE} will prevent any
calculations in rows that contain \code{NA} values, which could
be useful.}

\item{minDiff}{numeric value indicating the minimum difference
from median to qualify as an outlier. This value protects
against removing outliers which are already extremely
close together.}

\item{includeAttributes}{logical indicating whether to return
attributes that describe the threshold and type of threshold
used per row, in addition to the madFactor and minDiff values
defined.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional parameters are ignored.}
}
\value{
A numeric matrix is returned, with the same dimensions
as the input \code{x} matrix. Outliers are replaced with \code{NA}.

If \code{includeAttributes=TRUE} then four attributes will be
included as well:
\describe{
\item{attr(x, "rowThresholds")}{The row difference from
median threshold used for filtering.}
\item{attr(out, "rowThresholds")}{When the difference from
median is above \code{minDiff} it returns "madFactor",
otherwise it returns "minDiff".}
\item{attr(x, "madFactor")}{The \code{madFactor} value used.}
\item{attr(x, "minDiff")}{The \code{minDiff} value used.}
}
}
\description{
Remove outlier points per row by MAD factor threshold
}
\details{
This function applies outlier detection and removal per
row of the input numeric matrix. It first calculates MAD
per row, then defines a cutoff threshold by multiplying the
MAD by the supplied \code{madFactor}, and by ensuring the
threshold is at least as large as \code{minDiff}.

The \code{minDiff} parameter affects cases such as 3 replicates,
where all replicates are well within a known threshold
indicating low variance, but where two replicates might
be nearly identical. Consider \code{c(2.0, 2.0, 2.001)}. The
MAD is zero, therefore any difference from median might
otherwise be considered an outlier. One alternative is
to define \code{minDiff <- median(rowMads(x))} which sets the
minimum threshold to the median MAD across all rows.
Such a threshold will only be reasonable if the variance
across all rows is expected to be similar.

Note that this function is substantially faster when the
\code{matrixStats} package is installed, but will use the
\code{apply(x, 1, mad)} format as a last option.

This function assumes the input data is appropriate for
the use of MAD as a summary statistic, specifically that
the numeric values per row are expected to be roughly
normally distributed, with the occassional presence
of outlier points. The outlier points are assumed to be
technical outliers, which are often caused by some
instrument measurement failure or other upstream
protocol failure. A threshold of 5xMAD is fairly lenient
and yet is usually sufficient to identify the most
egregious outliers. Among technical replicates, often
2xMAD is more appropriate.
}
\examples{
set.seed(123);
x <- matrix(ncol=5, rnorm(25))*5 + 10;
## Define some outlier points
x[1:2,3] <- x[1:2,3]*5 + 50;
x[2:3,2] <- x[2:3,2]*5 - 100;
rowRmMadOutliers(x, madFactor=5);

}
\seealso{
Other jam numeric functions: 
\code{\link{deg2rad}()},
\code{\link{fix_matrix_ratio}()},
\code{\link{noiseFloor}()},
\code{\link{normScale}()},
\code{\link{rad2deg}()},
\code{\link{rowGroupMeans}()},
\code{\link{warpAroundZero}()}
}
\concept{jam numeric functions}
