% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{sclass}
\alias{sclass}
\title{return the classes of a list of objects}
\usage{
sclass(x, ...)
}
\arguments{
\item{x}{an S3 object inheriting from class "list", or an S4 object.}

\item{...}{additional parameters are ignored.}
}
\description{
return the classes of a list of objects
}
\details{
This function takes a list and returns the classes for each
object in the list. In the event an object class has multiple values,
the returned object is a list, otherwise is a vector.
If \code{x} is an S4 object, then \code{slotNames(x)} is used, and
the class is returned for each S4 slot.
}
\examples{
sclass(list(LETTERS=LETTERS, letters=letters));

}
