% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{sdim}
\alias{sdim}
\title{print dimensions of list object elements}
\usage{
sdim(x, includeClass = TRUE, doFormat = FALSE, big.mark = ",",
  verbose = FALSE, ...)
}
\arguments{
\item{x}{an S3 object inheriting from class "list", or an S4 object.}

\item{includeClass}{boolean indicating whether to print the class of
each element in the input \code{x} object. Note that for S4 objects,
each element will be the object returned for each of \code{slotNames(x)}.}

\item{doFormat}{boolean indicating whether to format the dimensions using
\code{format(...,big.mark=",")}, which is mainly useful for extremely
large dimensions. This parameter should probably become more broadly
useful and respectful for different locales.}

\item{big.mark}{character value used when \code{doFormat=TRUE}, used in the
call to \code{format(...,big.mark)}.}

\item{verbose}{logical whether to print verbose output}

\item{...}{additional parameters are ignored.}
}
\value{
`data.frame` where each row indicates the dimensions of
   each element in the input list. When `includeClass` is `TRUE` it
   will include a column `class` which indicates the class of each
   list element. When the input list contains arrays with more than
   two dimensions, the first two dimensions are named `"rows"` and
   `"columns"` with additional dimensions named `"dim3"` and so on.
   Any list element with fewer than that many dimensions will only have
   values populated to the relevant dimensions, for example a character
   vector will only populate the length.
}
\description{
print dimensions of list object elements, such as a list of data.frames
}
\details{
This function prints the dimensions of a list of objects, usually a list
of data.frames, but extended to handle more complicated lists, including
even S4 object slotNames().

Over time, more object types will be made compatible with this function.
Currently, igraph objects will print the number of nodes and edges, but
requires the igraph package to be installed.
}
\examples{
L <- list(LETTERS=LETTERS,
   letters=letters,
   lettersDF=data.frame(LETTERS, letters));
sdim(L);

}
