% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-showcolors.R
\name{showColors}
\alias{showColors}
\title{Show colors from a vector or list}
\usage{
showColors(
  x,
  labelCells = NULL,
  transpose = FALSE,
  srtCellnote = NULL,
  adjustMargins = TRUE,
  makeUnique = FALSE,
  doPlot = TRUE,
  ...
)
}
\arguments{
\item{x}{one of these input types:
\itemize{
\item \code{character} vector of colors
\item \code{list} with any combination of \code{character} or \code{function}.
Each element of the \code{list} is displayed on its own row.
List names are shown on the y-axis.
\item \code{function}: color function in one of two formats:
\enumerate{
\item \code{circlize::colorRamp2()} which defines numeric breaks, and
one color \emph{at} each break. This function is used by ComplexHeatmap
and is unique and useful in defining the color at each break and
not in between each break. The benefit is that a specific color
is known to mean exactly the numeric value, instead of assigning the
color to some intermediate mean of adjacent breaks, then interpolating
the color between them.
\item \code{function} as defined in color packages such as \code{viridis::viridis(10)}
where the number \code{10} defines the number of colors to produce.
For these functions, colors are displayed ranging from 0 to 1,
intending to mean lowest (0) to highest (1) color, with n steps.
}
\item 'ggplot' or 'ggproto' object, see Details.
}}

\item{labelCells}{\code{logical} whether to label colors atop the color itself.
If NULL (default) it will only display labels with 40 or fewer items
on either axis.}

\item{transpose}{\code{logical} whether to transpose the colors to display
top-to-bottom, instead of left-to-right.}

\item{srtCellnote}{\code{numeric} angle to rotate text when
\code{labelCells=TRUE}. When set to NULL, labels are vertical
srtCellnote=90 when \code{transpose=FALSE} and horizontal
srtCellnote=0 when \code{transpose=TRUE}.}

\item{adjustMargins}{\code{logical} indicating whether to call
\code{adjustAxisLabelMargins()} to adjust the x- and y-axis
label margins to accomodate the label size.
\itemize{
\item Note when an axis is hidden by using \code{xaxt="n"} or \code{xaxt="n"},
the respective margin will not be adjusted.
\item The arguments in \code{...} take precedence over \code{graphics::par()},
when deciding whether to adjust margins. However if \code{xaxt="s"} and
\code{graphics::par("xaxt"="n")} the margin will be adjusted but not
displayed.
In this way the axes can be adjusted without displaying the labels,
so the labels can be rendered later if needed.
}}

\item{makeUnique}{\code{logical} indicating whether to display only the first
unique color. When \code{x} is supplied as a \code{list} this operation will
display the first unique color for each \code{list} element.
Also, when \code{x} is a \code{list}, just to be fancy, \code{makeUnique}
is recycled to \code{length(x)} so certain list elements can display
unique values, while others display all values.}

\item{doPlot}{\code{logical} indicating whether to produce a visual plot.
Note this function returns the color matrix invisibly.}

\item{...}{additional parameters are passed to \code{imageByColors()}.}
}
\value{
invisible color \code{matrix} used by \code{imageByColors()}. When
the input \code{x} is empty, or cannot be converted to colors when
\code{x} contains a \code{function}, the output returns 'NULL'.
}
\description{
Show colors from a vector or list
}
\details{
This function simply displays colors for review, using
\code{imageByColors()} to display colors and labels across the
plot space.

When supplied a \code{list}, each row in \code{imageByColors()} represents
an entry in the \code{list}. Nothing fancy.

When input is \code{function}, it is assumed to be one of these
formats:
\enumerate{
\item \code{viridis::viridis(n)}
\itemize{
\item Sequential colors with \code{n} steps
representing colors from lowest to highest value.
\item This format does not apply any numeric range,
no numeric threshold is implied in the function at all.
It simply takes its internal range of colors and produces
\code{n} output colors, usually using \code{colorRampPalette()}
to interpolate intermediate colors when needed.
\item Common examples are R packages such as 'viridis'
which provide several different color gradients.
}
\item \code{circlize::colorRamp2()}
\itemize{
\item Colors associated with specific \code{numeric} values.
\item A distinctive feature of \code{colorRamp2()} color \code{function}
is that it assigns a vector of 'colors' to a vector
of \code{numeric} values. This mechanism is important when
it is useful to know exactly what \code{numeric} value
is represented by a color.
\item A common alternative is to assign colors \emph{between}
\code{numeric} breaks, in which case the defined color
is associated with an intermediate value betweenbreaks.
}
\item \code{ggplot} or \code{ggproto} scales object produced by 'ggplot2'.
\itemize{
\item Given a \code{ggplot2::ggplot} object, it will use
'colour' or 'fill' aesthetics present in the 'mapping',
only when there is an explicit color assignment.
It does not currently determine the default color
aesthetic function to use.
\item A \code{ggproto} object produced by a color scales function,
for example \code{ggplot2::scale_color_discrete()},
for aesthetics 'color', 'colour', or 'fill'.
\item It makes reasonable attempt to recognize custom limits,
for example '_gradientn()' functions which may have
specific 'values' (scaled from 0 to 1) buthent applied to
specific 'limits' (scaled per user coordinates).
}
}
}
\examples{
x <- color2gradient(list(Reds=c("red"), Blues=c("blue")), n=c(4,7));
showColors(x);

showColors(getColorRamp("firebrick3"))

if (requireNamespace("RColorBrewer", quietly=TRUE)) {
   RColorBrewer_namelist <- rownames(RColorBrewer::brewer.pal.info);
   y <- lapply(nameVector(RColorBrewer_namelist), function(i){
      n <- RColorBrewer::brewer.pal.info[i, "maxcolors"]
      j <- RColorBrewer::brewer.pal(n, i);
      nameVector(j, seq_along(j));
   });
   showColors(y, cexCellnote=0.6, cex.axis=0.7, main="Brewer Colors");
}
if (requireNamespace("viridisLite", quietly=TRUE)) {
   # given one function name it will display discrete colors
   showColors(viridisLite::viridis)
   # a list of functions will show each function output
   showColors(list(viridis=viridisLite::viridis,
      inferno=viridisLite::inferno))

   # grab the full viridis color map
   z <- rgb2col(viridisLite::viridis.map[,c("R","G","B")]);
   # split the colors into a list
   viridis_names <- c(A="magma",
      B="inferno",
      C="plasma",
      D="viridis",
      E="cividis",
      F="rocket",
      G="mako",
      H="turbo")
   y <- split(z,
      paste0(viridisLite::viridis.map$opt, ": ",
      viridis_names[viridisLite::viridis.map$opt]));
   showColors(y, labelCells=TRUE, xaxt="n", main="viridis.map colors");
}

# demonstrate makeUnique=TRUE
j1 <- getColorRamp("rainbow", n=7);
names(j1) <- seq_along(j1);
j2 <- rep(j1, each=3);
names(j2) <- makeNames(names(j2), suffix="_rep");
j2
showColors(list(
   j1=j1,
   j2=j2,
   j3=j2),
   makeUnique=c(FALSE, FALSE, TRUE))

}
\seealso{
Other jam plot functions:
\code{\link[=adjustAxisLabelMargins]{adjustAxisLabelMargins()}},
\code{\link[=coordPresets]{coordPresets()}},
\code{\link[=decideMfrow]{decideMfrow()}},
\code{\link[=drawLabels]{drawLabels()}},
\code{\link[=getPlotAspect]{getPlotAspect()}},
\code{\link[=groupedAxis]{groupedAxis()}},
\code{\link[=imageByColors]{imageByColors()}},
\code{\link[=imageDefault]{imageDefault()}},
\code{\link[=minorLogTicksAxis]{minorLogTicksAxis()}},
\code{\link[=nullPlot]{nullPlot()}},
\code{\link[=plotPolygonDensity]{plotPolygonDensity()}},
\code{\link[=plotRidges]{plotRidges()}},
\code{\link[=plotSmoothScatter]{plotSmoothScatter()}},
\code{\link[=shadowText]{shadowText()}},
\code{\link[=shadowText_options]{shadowText_options()}},
\code{\link[=sqrtAxis]{sqrtAxis()}},
\code{\link[=usrBox]{usrBox()}}

Other jam color functions:
\code{\link[=alpha2col]{alpha2col()}},
\code{\link[=applyCLrange]{applyCLrange()}},
\code{\link[=col2alpha]{col2alpha()}},
\code{\link[=col2hcl]{col2hcl()}},
\code{\link[=col2hsl]{col2hsl()}},
\code{\link[=col2hsv]{col2hsv()}},
\code{\link[=color2gradient]{color2gradient()}},
\code{\link[=fixYellow]{fixYellow()}},
\code{\link[=fixYellowHue]{fixYellowHue()}},
\code{\link[=getColorRamp]{getColorRamp()}},
\code{\link[=hcl2col]{hcl2col()}},
\code{\link[=hsl2col]{hsl2col()}},
\code{\link[=hsv2col]{hsv2col()}},
\code{\link[=isColor]{isColor()}},
\code{\link[=kable_coloring]{kable_coloring()}},
\code{\link[=makeColorDarker]{makeColorDarker()}},
\code{\link[=rainbow2]{rainbow2()}},
\code{\link[=rgb2col]{rgb2col()}},
\code{\link[=setCLranges]{setCLranges()}},
\code{\link[=setTextContrastColor]{setTextContrastColor()}},
\code{\link[=unalpha]{unalpha()}},
\code{\link[=warpRamp]{warpRamp()}}
}
\concept{jam color functions}
\concept{jam plot functions}
