% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{unalpha}
\alias{unalpha}
\title{Remove alpha transparency from colors}
\usage{
unalpha(x, ...)
}
\value{
character vector of R colors in hex format.
}
\description{
Remove alpha transparency from colors
}
\details{
This function simply removes the alpha transparency from
R colors, returned in hex format, for example \code{"#FF0000FF"}
becomes \code{"#FF0000"}, or \code{"blue"} becomes \code{"#0000FF"}.

It also silently converts R color names to hex format,
where applicable.
}
