% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{warpAroundZero}
\alias{warpAroundZero}
\title{Warp a vector of numeric values relative to zero}
\usage{
warpAroundZero(x, lens = 5, baseline = 0, xCeiling = NULL, ...)
}
\arguments{
\item{x}{numeric vector}

\item{lens}{numeric value which defines the lens factor,
where \code{lens > 0} will compress values near zero, and
\code{lens < 0} will expand values near zero and compress
values near the maximum value. If \code{lens == 0} the
numeric values are not changed.}

\item{baseline}{numeric value describing the baseline, for example
when the central value is non-zero. Note this baseline is applied
to the absolute value of \code{x} which means the baseline is also
symmetric around zero.}

\item{xCeiling}{numeric maximum value allowed for \code{abs(x)}, above which
values are set to \code{xCeiling} maintaining the positive or negative
sign. The intention is to prevent outlier points
from adversely affecting the numeric range used for the warping.}
}
\description{
Warp a vector of numeric values relative to zero
}
\details{
This function warps numeric values using a log curve
transformation, such that values are either more compressed
near zero, or more compressed near the maximum values.
For example, a vector of integers from -10 to 10 would be warped
so the intervals near zero were smaller than 1, and intervals
farthest from zero are greater than 1.

The main driver for this function was the desire to compress
divergent color scales used in heatmaps, in order to enhance
smaller magnitude numeric values. Existing color ramps map the
color gradient in a linear manner relative to the numeric range,
which can cause extreme values to dominate the color scale.
Further, a linear application of colors is not always appropriate.
}
\examples{
x <- c(-10:10);
xPlus10 <- warpAroundZero(x, lens=10);
xMinus10 <- warpAroundZero(x, lens=-10);

plot(x=x, y=xPlus10, type="b", pch=20, col="dodgerblue");
points(x=x, y=xMinus10, type="b", pch=20, col="orangered");

}
