% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{alpha2col}
\alias{alpha2col}
\title{set R color alpha value}
\usage{
alpha2col(x, alpha = 1, maxValue = 1, ...)
}
\arguments{
\item{x}{R compatible color, either a color name, or hex value, or
a mixture of the two. Any value compatible with \code{\link{col2rgb}}.}

\item{alpha}{numeric alpha transparency to use per x color. alpha is
recycled to length(x) as needed.}

\item{maxValue}{numeric maximum value to return, useful when the downstream
alpha range should be 255. By default maxValue=1 is returned.}
}
\description{
Define the alpha transparency per R color
}
\seealso{
Other jam color functions: \code{\link{col2alpha}},
  \code{\link{col2hcl}}, \code{\link{col2hsv}},
  \code{\link{color2gradient}}, \code{\link{getColorRamp}},
  \code{\link{hcl2col}}, \code{\link{hsv2col}},
  \code{\link{isColor}}, \code{\link{kable_coloring}},
  \code{\link{makeColorDarker}}, \code{\link{rgb2col}},
  \code{\link{setTextContrastColor}}
}
