% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{applyCLrange}
\alias{applyCLrange}
\title{Apply CL color range}
\usage{
applyCLrange(x, Crange = NULL, Lrange = NULL, Cgrey = 5,
  fixYellow = TRUE, CLmethod = c("scale", "floor", "expand"), ...)
}
\arguments{
\item{x}{vector of R colors}

\item{Crange}{NULL or numeric vector with minimum and maximum allowed
values for the chroma (C) component.}

\item{Lrange}{NULL or numeric vector with minimum and maximum allowed
values for the luminance (L) component.}

\item{Cgrey}{numeric chroma (C) value, which defines grey colors at or
below this chroma. Any colors at or below the grey cutoff will have
their C values unchanged. This mechanism prevents converting black
to red, for example. To disable the effect, set \code{Cgrey=-1}.}

\item{fixYellow}{boolean indicating whether to "fix" the darkening of
yellow, which otherwise turns to green. Instead, since JAM can,
JAM will make the yellow slightly more golden before darkening,
which is achieved by calling \code{fixYellowHue()}.}

\item{CLmethod}{character string indicating how to alter values
outside the respective \code{Crange} and \code{Lrange} ranges. "scale" will
rescale values only if any are outside of range, and will rescale
the full range of \code{c(Crange, Cvalues)} to \code{c(Crange)}. In this way,
only values outside the range are rescaled. "floor" will apply a
fixed cutoff, any values outside the range are set to equal the
range boundary itself. "expand" will rescale all values so the
range is equal to \code{Crange}.}

\item{...}{additional argyments are passed to \code{fixYellowHue()} when
\code{fixYellow} is \code{TRUE}.}
}
\value{
vector of colors after applying the chroma (C) and luminance (L)
ranges.
}
\description{
Apply chroma (C) and luminance (L) ranges to a vector of R colors
}
\details{
This function is primarily intended to restrict the range of brightness
values so they contrast with a background color, particularly when the
background color may be bright or dark in different scenarios.

The C and L values are defined by \code{colorspace::polarLUV()}, where C is
typically restricted to \code{0..100} and L is typically \code{0..100}. For some
colors, values above 100 are allowed.

Values are restricted to the given numeric range using one of three
methods, set via the \code{CLmethod} argument.

As an example, consider what should be done when \code{Crange <- c(10,70)}
and the C values are \code{Cvalues <- c(50, 60, 70, 80)}.
\enumerate{
\item "floor" uses \code{jamba::noiseFloor()} to apply fixed cutoffs at the
minimum and maximum range. This method has the effect of making all
values outside the range into an equal final value.
\item "scale" will apply \code{jamba::normScale()} to rescale only values outside
the given range. For example, \code{c(Crange, Cvalues)} as the initial range,
it constrains values to \code{c(Crange)}.  This method has the effect of
maintaining the relative difference between values.
\item "expand" will simply apply \code{jamba::normScale()} to fit the values
to the minimum and maximum range values. This method has the effect of
forcing colors to fit the full numeric range, even when the original
differences between values were small.
}

In case (1) above, Cvalues will become \code{c(50, 60, 70, 70)}.
In case (2) above, Cvalues will become \code{c(44, 53, 61, 70)}
In case (3) above, Cvalues will become \code{c(10, 30, 50, 70)}

Note that colors with C (chroma) values less than \code{Cgrey} will not have
the C value changed, in order to maintain colors at a greyscale, without
colorizing them. Particularly for pure \code{grey}, which has \code{C=0}, but
is still required to have a hue H, it is important not to increase
\code{C}.
}
\examples{
applyCLrange(c("red","blue","yellow"), Lrange=c(0,60));

}
\seealso{
Other jam practical functions: \code{\link{breakDensity}},
  \code{\link{checkLightMode}},
  \code{\link{colNum2excelName}}, \code{\link{exp2signed}},
  \code{\link{fileInfo}}, \code{\link{fixYellowHue}},
  \code{\link{fixYellow}}, \code{\link{getAxisLabel}},
  \code{\link{handleArgsText}}, \code{\link{isFALSEV}},
  \code{\link{isTRUEV}}, \code{\link{jamba}},
  \code{\link{jargs}}, \code{\link{kable_coloring}},
  \code{\link{log2signed}}, \code{\link{make_styles}},
  \code{\link{mergeAllXY}}, \code{\link{minorLogTicks}},
  \code{\link{newestFile}}, \code{\link{printDebugI}},
  \code{\link{printDebug}}, \code{\link{renameColumn}},
  \code{\link{rmInfinite}}, \code{\link{rmNA}},
  \code{\link{rmNULL}}, \code{\link{sclass}},
  \code{\link{sdima}}, \code{\link{sdim}},
  \code{\link{setCLranges}}, \code{\link{setPrompt}},
  \code{\link{ssdima}}, \code{\link{ssdim}}

Other jam color functions: \code{\link{alpha2col}},
  \code{\link{col2alpha}}, \code{\link{col2hcl}},
  \code{\link{col2hsv}}, \code{\link{color2gradient}},
  \code{\link{fixYellowHue}}, \code{\link{fixYellow}},
  \code{\link{getColorRamp}}, \code{\link{hcl2col}},
  \code{\link{hsv2col}}, \code{\link{isColor}},
  \code{\link{kable_coloring}},
  \code{\link{makeColorDarker}}, \code{\link{make_styles}},
  \code{\link{rgb2col}}, \code{\link{setCLranges}},
  \code{\link{setTextContrastColor}},
  \code{\link{showColors}}, \code{\link{unalpha}},
  \code{\link{warpRamp}}
}
\concept{jam color functions}
\concept{jam practical functions}
