% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-export.r
\name{applyXlsxCategoricalFormat}
\alias{applyXlsxCategoricalFormat}
\title{Add categorical colors to Excel xlsx worksheets}
\usage{
applyXlsxCategoricalFormat(xlsxFile, sheet = 1, rowRange = NULL,
  colRange = NULL, colorSub = NULL,
  colorSubText = setTextContrastColor(colorSub), trimCatNames = TRUE,
  overwrite = TRUE, stack = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{xlsxFile}{filename pointing to an existing Excel xlsx file.}

\item{sheet}{integer index of the worksheet or worksheets.}

\item{rowRange, colRange}{integer vectors of rows and columns
to apply categorical colors in the Excel xlsx worksheet.}

\item{colorSub}{named character vector of valid R colors, whose
names correspond to values in the worksheet cells.}

\item{colorSubText}{optional character vector of colors, whose
names correspond to values in the worksheet cells. In
absence of a specific text color, \code{jamba::setTextContrastColor()}
is used to define a contrasting text color to be visible on
the colored background.}

\item{trimCatNames}{logical whether to trim whitespace and punctuation
from \code{colorSub} and from Excel cell fields before matching colors
to Excel values.}

\item{overwrite}{logical indicating whether new cell color styles
should be forced overwrite of previous cell styles.}

\item{stack}{logical indicating whether new color rules should be
applied above existing styles, many of whose styles may not affect
the specific cell color, for example the font size and font name.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Add categorical colors to Excel xlsx worksheets
}
