% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{asDate}
\alias{asDate}
\title{convert date DDmmmYYYY to Date}
\usage{
asDate(getDateValues, dateFormat = "\%d\%b\%Y", ...)
}
\arguments{
\item{getDateValues}{character date, in format recognized by dateFormat}

\item{dateFormat}{character string representing the recognized date format,
by default DDmmmYYYY, for example 23aug2007.}

\item{...}{additional parameters are ignored.}
}
\value{
Date object
}
\description{
convert date DDmmmYYYY to Date
}
\details{
This function converts a text date string to Date object, mainly to
allow date-related math operations, for example \code{\link[base]{difftime}}.
}
\examples{
asDate(getDate());

}
