% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{checkLightMode}
\alias{checkLightMode}
\title{check lightMode for light background color}
\usage{
checkLightMode(lightMode = NULL, ...)
}
\arguments{
\item{lightMode}{logical or NULL, indicating whether the lightMode
parameter has been defined in the function call.}
}
\value{
logical or length=1, indicating whether lightMode is defined
}
\description{
check lightMode for light background color
}
\details{
Check the lightMode status through function parameter, options, or
environment variable. If the function defines lightMode, it is used as-is.
If lightMode is NULL, then options("lightMode") is used if defined.
Otherwise, it tries to detect whether the R session is running inside
Rstudio using the environmental variable "RSTUDIO", and if so it assumes
lightMode==TRUE.

To set a default lightMode, add options("lightMode"=TRUE) to .Rprofile, or
to the relevant R script.
}
\examples{
checkLightMode(TRUE);
checkLightMode();

}
