% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{checkLightMode}
\alias{checkLightMode}
\title{check lightMode for light background color}
\usage{
checkLightMode(lightMode = NULL, ...)
}
\arguments{
\item{lightMode}{logical or NULL, indicating whether the lightMode
parameter has been defined in the function call.}
}
\value{
logical or length=1, indicating whether lightMode is defined
}
\description{
check lightMode for light background color
}
\details{
Check the lightMode status through function parameter, options, or
environment variable. If the function defines lightMode, it is used as-is.
If lightMode is NULL, then options("jam.lightMode") is used if defined.
Otherwise, it tries to detect whether the R session is running inside
Rstudio using the environmental variable "RSTUDIO", and if so it assumes
lightMode==TRUE.

To set a default lightMode, add options("jam.lightMode"=TRUE) to .Rprofile, or
to the relevant R script.
}
\examples{
checkLightMode(TRUE);
checkLightMode();

}
\seealso{
Other jam practical functions: \code{\link{applyCLrange}},
  \code{\link{breakDensity}},
  \code{\link{colNum2excelName}}, \code{\link{fileInfo}},
  \code{\link{fixYellowHue}}, \code{\link{fixYellow}},
  \code{\link{getAxisLabel}}, \code{\link{handleArgsText}},
  \code{\link{isFALSEV}}, \code{\link{isTRUEV}},
  \code{\link{jamba}}, \code{\link{jargs}},
  \code{\link{kable_coloring}}, \code{\link{make_styles}},
  \code{\link{minorLogTicks}}, \code{\link{newestFile}},
  \code{\link{printDebug}}, \code{\link{renameColumn}},
  \code{\link{rmInfinite}}, \code{\link{rmNA}},
  \code{\link{rmNULL}}, \code{\link{sclass}},
  \code{\link{sdim}}, \code{\link{setCLranges}},
  \code{\link{setPrompt}}, \code{\link{ssdim}}
}
\concept{jam practical functions}
