% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-clrange.R
\name{checkLightMode}
\alias{checkLightMode}
\title{check lightMode for light background color}
\usage{
checkLightMode(lightMode = NULL, ...)
}
\arguments{
\item{lightMode}{logical or NULL, indicating whether the lightMode
parameter has been defined in the function call.}
}
\value{
logical or length=1, indicating whether lightMode is defined
}
\description{
check lightMode for light background color
}
\details{
Check the lightMode status through function parameter, options, or
environment variable. If the function defines lightMode, it is used as-is.
If lightMode is NULL, then options("jam.lightMode") is used if defined.
Otherwise, it tries to detect whether the R session is running inside
Rstudio using the environmental variable "RSTUDIO", and if so it assumes
lightMode==TRUE.

To set a default lightMode, add options("jam.lightMode"=TRUE) to .Rprofile, or
to the relevant R script.
}
\examples{
checkLightMode(TRUE);
checkLightMode();

}
\seealso{
Other jam practical functions: 
\code{\link{breakDensity}()},
\code{\link{check_pkg_installed}()},
\code{\link{colNum2excelName}()},
\code{\link{color_dither}()},
\code{\link{diff_functions}()},
\code{\link{exp2signed}()},
\code{\link{fileInfo}()},
\code{\link{fixYellow}()},
\code{\link{getAxisLabel}()},
\code{\link{handleArgsText}()},
\code{\link{heads}()},
\code{\link{isFALSEV}()},
\code{\link{isTRUEV}()},
\code{\link{jamba}},
\code{\link{jargs}()},
\code{\link{kable_coloring}()},
\code{\link{lldf}()},
\code{\link{log2signed}()},
\code{\link{make_html_styles}()},
\code{\link{make_styles}()},
\code{\link{match_unique}()},
\code{\link{mergeAllXY}()},
\code{\link{middle}()},
\code{\link{minorLogTicks}()},
\code{\link{newestFile}()},
\code{\link{printDebugI}()},
\code{\link{printDebug}()},
\code{\link{renameColumn}()},
\code{\link{rmInfinite}()},
\code{\link{rmNAs}()},
\code{\link{rmNA}()},
\code{\link{rmNULL}()},
\code{\link{sclass}()},
\code{\link{sdima}()},
\code{\link{sdim}()},
\code{\link{setPrompt}()},
\code{\link{ssdima}()},
\code{\link{ssdim}()}
}
\concept{jam practical functions}
