% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{col2alpha}
\alias{col2alpha}
\title{get R color alpha value}
\usage{
col2alpha(x, maxValue = 1, ...)
}
\arguments{
\item{x}{R compatible color, either a color name, or hex value, or
a mixture of the two. Any value compatible with \code{\link{col2rgb}}.}

\item{maxValue}{numeric maximum value to return, useful when the downstream
alpha range should be 255. By default maxValue=1 is returned.}
}
\description{
Return the alpha transparency per R color
}
\seealso{
jamba color manipulation functions: \code{\link{hsv2col}},
   \code{\link{col2hsv}},\code{\link{col2hcl}},
   \code{\link{rgb2col}},
   \code{\link{alpha2col}},\code{\link{col2alpha}},
   \code{\link{makeColorDarker}},
   \code{\link{color2gradient}},
   \code{\link{setTextContrastColor}},
   \code{\link{showColors}}
}
