% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{col2hcl}
\alias{col2hcl}
\title{convert R color to HCL color matrix}
\usage{
col2hcl(x, maxColorValue = 255, model = c("hcl", "polarLUV", "polarLAB"), ...)
}
\arguments{
\item{x}{R compatible color, either a color name, or hex value, or
a mixture of the two. Any value compatible with
\code{\link[grDevices]{col2rgb}}.}

\item{model}{the color model to use, from \code{"polarLUV"} for the
standard R conventional HCL, and \code{"polarLAB"} which uses the
LAB-based HCL values.}

\item{...}{additional arguments are ignored.}

\item{maxValue}{numeric maximum value to return, useful when the downstream
alpha range should be 255. By default maxValue=1 is returned.}
}
\description{
convert R color to HCL color matrix
}
\details{
This function takes an R color and converts to an HCL matrix, using
the colorspace package, and \code{\link[colorspace]{RGB}} and
\code{\link[colorspace]{polarLUV}} functions. It is also used to
maintain alpha transparency, to enable interconversion via other
color manipulation functions as well.
}
\examples{
col2hcl("#FF000044")

}
\seealso{
Other jam color functions: 
\code{\link{alpha2col}()},
\code{\link{applyCLrange}()},
\code{\link{col2alpha}()},
\code{\link{col2hsv}()},
\code{\link{color2gradient}()},
\code{\link{fixYellowHue}()},
\code{\link{fixYellow}()},
\code{\link{getColorRamp}()},
\code{\link{hcl2col}()},
\code{\link{hsv2col}()},
\code{\link{isColor}()},
\code{\link{kable_coloring}()},
\code{\link{makeColorDarker}()},
\code{\link{make_styles}()},
\code{\link{rgb2col}()},
\code{\link{setCLranges}()},
\code{\link{setTextContrastColor}()},
\code{\link{showColors}()},
\code{\link{unalpha}()},
\code{\link{warpRamp}()}
}
\concept{jam color functions}
