% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{col2hcl}
\alias{col2hcl}
\title{convert R color to HCL color matrix}
\usage{
col2hcl(x, maxColorValue = 255, ...)
}
\arguments{
\item{x}{R compatible color, either a color name, or hex value, or
a mixture of the two. Any value compatible with \code{\link{col2rgb}}.}

\item{maxValue}{numeric maximum value to return, useful when the downstream
alpha range should be 255. By default maxValue=1 is returned.}
}
\description{
convert R color to HCL color matrix
}
\details{
This function takes an R color and converts to an HCL matrix, using
the colorspace package, and \code{\link[colorspace]{RGB}} and
\code{\link[colorspace]{polarLUV}} functions. It is also used to
maintain alpha transparency, to enable interconversion via other
color manipulation functions as well.
}
\examples{
col2hcl("#FF000044")

}
