% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{col2hsv}
\alias{col2hsv}
\title{Convert R color to HSV matrix}
\usage{
col2hsv(x, ...)
}
\arguments{
\item{x}{R color}

\item{...}{additional parameters are ignored}
}
\value{
matrix of HSV colors
}
\description{
Convert R color to HSV matrix
}
\details{
This function takes a valid R color and converts to a HSV matrix. The
output can be effectively returned to R color with
\code{\link{hsv2col}}, usually after manipulating the
HSV color matrix.
}
\examples{
hsv2col(col2hsv("#FF000055"))

}
