% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-string.r
\name{cPasteSU}
\alias{cPasteSU}
\title{paste a list into a delimited vector using sorted, unique values}
\usage{
cPasteSU(
  x,
  sep = ",",
  doSort = TRUE,
  makeUnique = TRUE,
  na.rm = FALSE,
  keepFactors = FALSE,
  checkClass = TRUE,
  useBioc = TRUE,
  ...
)
}
\arguments{
\item{x}{input \code{list} of vectors}

\item{sep}{\code{character} delimiter used to paste multiple values together}

\item{doSort}{\code{logical} indicating whether to sort each vector
using \code{\link[=mixedOrder]{mixedOrder()}}.}

\item{makeUnique}{\code{logical} indicating whether to make each vector in
the input list unique before pasting its values together.}

\item{na.rm}{boolean indicating whether to remove NA values from
each vector in the input list. When \code{na.rm} is \code{TRUE} and a
list element contains only \code{NA} values, the resulting string
will be \code{""}.}

\item{keepFactors}{\code{logical} only used when \code{useLegacy=TRUE} and
\code{doSort=TRUE}; indicating whether to preserve factors,
keeping factor level order. When
\code{keepFactors=TRUE}, if any list element is a \code{factor}, all elements
are converted to factors. Note that this step combines overall
factor levels, and non-factors will be ordered using
\code{base::order()} instead of \code{jamba::mixedOrder()} (for now.)}

\item{useBioc}{\code{logical} indicating whether this function should try
to use \code{S4Vectors::unstrsplit()} when the Bioconductor package
\code{S4Vectors} is installed, otherwise it will use a less
efficient \code{mapply()} operation.}

\item{...}{additional arguments are passed to \code{mixedOrder()} when
\code{doSort=TRUE}.}
}
\description{
Paste a list of vectors into a character vector, with unique values
sorted then delimited by default with a comma.
}
\details{
This function is convenient a wrapper for \code{cPaste(.., doSort=TRUE, makeUnique=TRUE)}.
}
\seealso{
Other jam string functions: 
\code{\link{asSize}()},
\code{\link{breaksByVector}()},
\code{\link{cPasteS}()},
\code{\link{cPasteUnique}()},
\code{\link{cPasteU}()},
\code{\link{cPaste}()},
\code{\link{fillBlanks}()},
\code{\link{formatInt}()},
\code{\link{gsubOrdered}()},
\code{\link{makeNames}()},
\code{\link{mixedOrder}()},
\code{\link{mixedSortDF}()},
\code{\link{mixedSorts}()},
\code{\link{mixedSort}()},
\code{\link{mmixedOrder}()},
\code{\link{nameVectorN}()},
\code{\link{nameVector}()},
\code{\link{padInteger}()},
\code{\link{padString}()},
\code{\link{pasteByRowOrdered}()},
\code{\link{pasteByRow}()},
\code{\link{sizeAsNum}()},
\code{\link{tcount}()},
\code{\link{ucfirst}()},
\code{\link{uniques}()}

Other jam list functions: 
\code{\link{cPasteS}()},
\code{\link{cPasteUnique}()},
\code{\link{cPasteU}()},
\code{\link{cPaste}()},
\code{\link{heads}()},
\code{\link{jam_rapply}()},
\code{\link{list2df}()},
\code{\link{mergeAllXY}()},
\code{\link{mixedSorts}()},
\code{\link{rbindList}()},
\code{\link{relist_named}()},
\code{\link{rlengths}()},
\code{\link{sclass}()},
\code{\link{sdim}()},
\code{\link{uniques}()},
\code{\link{unnestList}()}
}
\concept{jam list functions}
\concept{jam string functions}
