% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-string.r
\name{cPasteUnique}
\alias{cPasteUnique}
\title{paste a list into a delimited vector using unique values}
\usage{
cPasteUnique(x, makeUnique = TRUE, na.rm = FALSE, ...)
}
\arguments{
\item{x}{input list of vectors}

\item{makeUnique}{boolean indicating whether to make each vector in
the input list unique before pasting its values together.}

\item{na.rm}{boolean indicating whether to remove NA values from
each vector in the input list. When \code{na.rm} is \code{TRUE} and a
list element contains only \code{NA} values, the resulting string
will be \code{""}.}

\item{..}{additional arguments are passed to \code{\link[=cPaste]{cPaste()}}.}
}
\description{
Paste a list of vectors into a character vector of unique values,
usually delimited by a comma.
}
\details{
This function is convenient a wrapper for \code{cPaste(.., makeUnique=TRUE)}.
}
\seealso{
Other jam string functions: \code{\link{asSize}},
  \code{\link{breaksByVector}}, \code{\link{cPaste}},
  \code{\link{fillBlanks}}, \code{\link{formatInt}},
  \code{\link{makeNames}}, \code{\link{mixedOrder}},
  \code{\link{mixedSortDF}}, \code{\link{mixedSort}},
  \code{\link{mmixedOrder}}, \code{\link{nameVectorN}},
  \code{\link{nameVector}}, \code{\link{padInteger}},
  \code{\link{padString}}, \code{\link{pasteByRowOrdered}},
  \code{\link{pasteByRow}}, \code{\link{tcount}},
  \code{\link{uniques}}

Other jam list functions: \code{\link{cPaste}},
  \code{\link{list2df}}, \code{\link{rbindList}},
  \code{\link{rlengths}}, \code{\link{sclass}},
  \code{\link{sdim}}, \code{\link{ssdim}},
  \code{\link{uniques}}
}
\concept{jam list functions}
\concept{jam string functions}
