% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{dateToDaysOld}
\alias{dateToDaysOld}
\title{convert date to age in days}
\usage{
dateToDaysOld(testDate, nowDate = Sys.Date(), units = "days", ...)
}
\arguments{
\item{testDate}{character date recognized by \code{\link{asDate}},
representing the test date.}

\item{nowDate}{character date recognized by \code{\link{asDate}},
representing the reference date, by default the current day.}

\item{units}{character indicating the units, as used by
\code{\link[base]{difftime}}.}

\item{...}{additional parameters are ignored.}
}
\description{
convert date to age in days
}
\examples{
dateToDaysOld("23aug2007")

}
