% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{dateToDaysOld}
\alias{dateToDaysOld}
\title{convert date to age in days}
\usage{
dateToDaysOld(testDate, nowDate = Sys.Date(), units = "days", ...)
}
\arguments{
\item{testDate}{\code{character} date recognized by \code{asDate()},
representing the test date.}

\item{nowDate}{\code{character} date recognized by \code{asDate()},
representing the reference date, by default the current day.}

\item{units}{\code{character} indicating the units, as used by
\code{difftime()}.}

\item{...}{additional parameters are ignored.}
}
\value{
integer value with the number of calendar days before the
current date, or the \code{nowDate} if supplied.
}
\description{
convert date to age in days
}
\examples{
dateToDaysOld("23aug2007")

}
\seealso{
Other jam date functions: 
\code{\link{asDate}()},
\code{\link{getDate}()}
}
\concept{jam date functions}
