% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{decideMfrow}
\alias{decideMfrow}
\title{Decide plot panel rows, columns for par(mfrow)}
\usage{
decideMfrow(n, method = c("aspect", "wide", "tall"), ...)
}
\description{
Decide plot panel rows, columns for par(mfrow)
}
