% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{fileInfo}
\alias{fileInfo}
\title{file information in data.frame format}
\usage{
fileInfo(fileList, ...)
}
\arguments{
\item{fileList}{character vector with one or more file paths.}
}
\value{
\code{data.frame} with file information, including "filename" and "size"
as additional colnames as compared to \code{file.info()} output.
}
\description{
file information in data.frame format
}
\details{
This function is a minor extension to \code{file.info()} in that it
adds the filename as a proper colname, and "size" which contains a text
file size.
}
\seealso{
Other jam practical functions: 
\code{\link{breakDensity}()},
\code{\link{checkLightMode}()},
\code{\link{check_pkg_installed}()},
\code{\link{colNum2excelName}()},
\code{\link{color_dither}()},
\code{\link{diff_functions}()},
\code{\link{exp2signed}()},
\code{\link{fixYellow}()},
\code{\link{getAxisLabel}()},
\code{\link{handleArgsText}()},
\code{\link{heads}()},
\code{\link{isFALSEV}()},
\code{\link{isTRUEV}()},
\code{\link{jamba}},
\code{\link{jargs}()},
\code{\link{kable_coloring}()},
\code{\link{lldf}()},
\code{\link{log2signed}()},
\code{\link{make_html_styles}()},
\code{\link{make_styles}()},
\code{\link{match_unique}()},
\code{\link{mergeAllXY}()},
\code{\link{middle}()},
\code{\link{minorLogTicks}()},
\code{\link{newestFile}()},
\code{\link{printDebug}()},
\code{\link{renameColumn}()},
\code{\link{rmInfinite}()},
\code{\link{rmNAs}()},
\code{\link{rmNA}()},
\code{\link{rmNULL}()},
\code{\link{sclass}()},
\code{\link{sdim}()},
\code{\link{setPrompt}()}
}
\concept{jam practical functions}
