% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-plots.r
\name{getAxisLabel}
\alias{getAxisLabel}
\title{Get axis label for minorLogTicks}
\usage{
getAxisLabel(i, asValues, logAxisType, logBase, base_limit = 2,
  labelValueOffset = 0, symmetricZero = (labelValueOffset > 0), ...)
}
\description{
Get axis label for minorLogTicks
}
\details{
This function is intended to be called internally by
\code{jamba::minorLogTicks()}.
}
