% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-plots.r
\name{getAxisLabel}
\alias{getAxisLabel}
\title{Get axis label for minorLogTicks}
\usage{
getAxisLabel(
  i,
  asValues,
  logAxisType = c("normal", "flip", "pvalue"),
  logBase,
  base_limit = 2,
  offset = 0,
  symmetricZero = (offset > 0),
  ...
)
}
\arguments{
\item{i}{\code{numeric} axis value}

\item{asValues}{\code{logical} indicating whether the value should be
evaluated.}

\item{logAxisType}{\code{character} string with the type of axis values:
\itemize{
\item \code{"normal"}: axis values as-is.
\item \code{"flip"}: inverted axis values, for example where negative values
should be displayed as negative log-transformed values.
\item \code{"pvalue"}: for values transformed as \code{-log10(pvalue)}
}}

\item{logBase}{\code{numeric} logarithmic base}

\item{base_limit}{\code{numeric} value indicating the minimum value that
should be written as an exponential.}

\item{offset}{\code{numeric} value of offset used for log transformation.}

\item{symmetricZero}{\code{logical} indicating whether negative values
should be displayed as negative log-transformed values.}

\item{...}{additional arguments are ignored.}
}
\value{
axis label as \code{character} or \code{expression} where necessary.
}
\description{
Get axis label for minorLogTicks
}
\details{
This function is intended to be called internally by
\code{jamba::minorLogTicks()}.
}
\seealso{
Other jam practical functions: 
\code{\link{breakDensity}()},
\code{\link{checkLightMode}()},
\code{\link{check_pkg_installed}()},
\code{\link{colNum2excelName}()},
\code{\link{color_dither}()},
\code{\link{diff_functions}()},
\code{\link{exp2signed}()},
\code{\link{fileInfo}()},
\code{\link{fixYellow}()},
\code{\link{handleArgsText}()},
\code{\link{heads}()},
\code{\link{isFALSEV}()},
\code{\link{isTRUEV}()},
\code{\link{jamba}},
\code{\link{jargs}()},
\code{\link{kable_coloring}()},
\code{\link{lldf}()},
\code{\link{log2signed}()},
\code{\link{make_html_styles}()},
\code{\link{make_styles}()},
\code{\link{match_unique}()},
\code{\link{mergeAllXY}()},
\code{\link{middle}()},
\code{\link{minorLogTicks}()},
\code{\link{newestFile}()},
\code{\link{printDebug}()},
\code{\link{renameColumn}()},
\code{\link{rmInfinite}()},
\code{\link{rmNAs}()},
\code{\link{rmNA}()},
\code{\link{rmNULL}()},
\code{\link{sclass}()},
\code{\link{sdim}()},
\code{\link{setPrompt}()}
}
\concept{jam practical functions}
