% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{getColorRamp}
\alias{getColorRamp}
\title{get color ramp by name, color, or function}
\usage{
getColorRamp(col, n = 15, trimRamp = FALSE, verbose = FALSE,
  defaultBaseColor = "#BBABAB77", reverseRamp = FALSE, ...)
}
\arguments{
\item{col}{accepts:
\describe{
   \item{"character vector"}{one or more colors used to define a color
      gradient. Where one color is supplied, a gradient is created from
      defaultBaseColor to the supplied color.}
   \item{"character vector"}{one name matching a known color ramp either
      from \code{\link[RColorBrewer]{brewer.pal.info}}, or from the
      \code{\link[viridis]{viridis}} package.}
}}

\item{trimRamp}{logical whether to trim off the extreme values of the
color ramp, for example to remove the very lightest and very darkest
colors. If trimRamp is an integer, then that many colors are removed.}

\item{verbose}{logical whether to print verbose output}

\item{defaultBaseColor}{character vector indicating a color from which to
begin a color gradient, only used when col is a single color.}

\item{reverseRamp}{logical indicating whether to reverse the resulting
color ramp. This value is ignored when a single value is supplied for
col, and where "_r" or "_rev" is detected as a substring at the end
of the character value.}
}
\description{
get color ramp by name, color, or function
}
\details{
This function accepts a color ramp name, a single color,
a vector of colors, or a function names, and returns a simple
vector of colors of the appropriate length, suitable as input
to a number of plotting functions.

For example, it recognizes \code{\link{RColorBrewer}} color ramp
names, but can reverse those color ramps with a suffix "_r" at the
end, e.g. "RdBu_r" will supply a color ramp from blue to red, suitable
for heatmaps where red is typically associated with heat and high
numeric values.
}
\examples{
getColorRamp("red");
printDebug(getColorRamp("red"));
printDebug(getColorRamp("RdBu"));
printDebug(getColorRamp("RdBu_r"));

}
