% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{getColorRamp}
\alias{getColorRamp}
\title{get color ramp by name, color, or function}
\usage{
getColorRamp(col, n = 15, trimRamp = FALSE, verbose = FALSE,
  defaultBaseColor = "grey90", reverseRamp = FALSE, alpha = TRUE, ...)
}
\arguments{
\item{col}{accepts
\itemize{
\item{"character vector"}{one or more colors used to define a color
gradient. Where one color is supplied, a gradient is created from
defaultBaseColor to the supplied color.}
\item{"character vector length 1"}{one name matching a known color ramp either
from \code{RColorBrewer::brewer.pal.info}, or from the
\code{viridis::viridis} package.}
}}

\item{trimRamp}{logical whether to trim off the extreme values of the
color ramp, for example to remove the very lightest and very darkest
colors. If trimRamp is an integer, then that many colors are removed.}

\item{verbose}{logical whether to print verbose output}

\item{defaultBaseColor}{character vector indicating a color from which to
begin a color gradient, only used when col is a single color.}

\item{reverseRamp}{logical indicating whether to reverse the resulting
color ramp. This value is ignored when a single value is supplied for
col, and where "_r" or "_rev" is detected as a substring at the end
of the character value.}

\item{alpha}{logical indicating whether to honor alpha transparency
whenever \code{colorRampPalette} is called. If colors contain
no alpha transparency, this setting has no effect, otherwise the
alpha value appears to be applied using a linear gradient between
colors.}
}
\description{
get color ramp by name, color, or function
}
\details{
This function accepts a color ramp name, a single color,
a vector of colors, or a function names, and returns a simple
vector of colors of the appropriate length, suitable as input
to a number of plotting functions.

For example, it recognizes \code{\link[RColorBrewer]{RColorBrewer}} color ramp
names, but can reverse those color ramps with a suffix "_r" at the
end, e.g. "RdBu_r" will supply a color ramp from blue to red, suitable
for heatmaps where red is typically associated with heat and high
numeric values.
}
\examples{
# get a gradient using red4
red4 <- getColorRamp("red4");
showColors(getColorRamp(red4));

# make a custom gradient
BuOr <- getColorRamp(c("dodgerblue","grey10","orange"));
showColors(BuOr);
colorList <- list(red4=red4, BuOr=BuOr);

# If RColorBrewer is available, use a brewer name
if (suppressPackageStartupMessages(require(RColorBrewer))) {
   RdBu <- getColorRamp("RdBu");
   RdBu_r <- getColorRamp("RdBu_r");
   colorList <- c(colorList, list(RdBu=RdBu, RdBu_r=RdBu_r));
   showColors(RdBu);
}

if (suppressPackageStartupMessages(require(viridis))) {
   viridisV <- getColorRamp("viridis");
   colorList <- c(colorList, list(viridis=viridisV));
}

# for fun, put a few color ramps onto one plot
showColors(colorList, cexCellnote=0.7);

}
\seealso{
Other jam color functions: \code{\link{alpha2col}},
  \code{\link{col2alpha}}, \code{\link{col2hcl}},
  \code{\link{col2hsv}}, \code{\link{color2gradient}},
  \code{\link{hcl2col}}, \code{\link{hsv2col}},
  \code{\link{isColor}}, \code{\link{kable_coloring}},
  \code{\link{makeColorDarker}}, \code{\link{rgb2col}},
  \code{\link{setTextContrastColor}}
}
