% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{getDate}
\alias{getDate}
\title{get simple date string}
\usage{
getDate(t = Sys.time(), trim = TRUE, ...)
}
\arguments{
\item{t}{current time, by default the output of \code{\link{Sys.time}}.}

\item{trim}{logical whether to trim the output of \code{\link{format}} in
the event that multiple values are sent for \code{t}.}

\item{...}{additional parameters sent to \code{\link{format}}}
}
\value{
character vector with simplified date string
}
\description{
get simple date string
}
\details{
Gets the current date in a simplified text string. Use
\code{\link{asDate}} to convert back to Date object.
}
\examples{
getDate();

}
