% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{handleArgsText}
\alias{handleArgsText}
\title{Handle function arguments as text}
\usage{
handleArgsText(argTextA, name = "", col1 = "mediumpurple2",
  col2 = "mediumaquamarine", colT = "dodgerblue3", colF = "red1",
  colNULL = "grey60", lightMode = checkLightMode(),
  adjustRgb = getOption("jam.adjustRgb"), Crange = NULL, Lrange = NULL,
  indent = "", useCollapseList = ",\\n      ", useCollapseBase = ", ",
  level = 1, debug = 0, useColor = TRUE, verbose = FALSE, ...)
}
\description{
Handles a list or list of lists, converting to human-readable text format
}
\details{
This function is a rare non-exported function intended to be called by
\code{jargs()}, but separated in order to help isolate the logical
steps required.
}
