% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{hsv2col}
\alias{hsv2col}
\title{Convert HSV matrix to R color}
\usage{
hsv2col(hsvValue, ...)
}
\arguments{
\item{hsvValue}{HSV matrix, with rownames c("h","s","v") in any order,
and optionally "alpha" rowname for alpha transparency.}
}
\description{
Converts a HSV color matrix to R hex color
}
\details{
This function augments the \code{\link{hsv}} function in that it handles
output from \code{\link{rgb2hsv}} or \code{\link{col2hsv}}, sufficient to
run a series of conversion functions, e.g. \code{hsv2col(col2hsv("red"))}.
This function also maintains alpha transparency, which is not maintained
by the \code{\link{hsv}} function.
}
\examples{
hsv2col(col2hsv("#FF000055"))

}
