% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{isFALSEV}
\alias{isFALSEV}
\title{Vectorized isFALSE}
\usage{
isFALSEV(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{additional arguments are ignored}
}
\description{
Vectorized isFALSE
}
\details{
This function applies three criteria to an input vector, to
determine if each entry in the vector is FALSE:
\enumerate{
\item It must be class \code{logical}.
\item It must not be \code{NA}.
\item It must evaluate as \code{FALSE}.
}
}
\seealso{
Other jam practical functions: \code{\link{applyCLrange}},
  \code{\link{breakDensity}}, \code{\link{checkLightMode}},
  \code{\link{colNum2excelName}}, \code{\link{fileInfo}},
  \code{\link{fixYellowHue}}, \code{\link{fixYellow}},
  \code{\link{getAxisLabel}}, \code{\link{handleArgsText}},
  \code{\link{isTRUEV}}, \code{\link{jamba}},
  \code{\link{jargs}}, \code{\link{kable_coloring}},
  \code{\link{make_styles}}, \code{\link{minorLogTicks}},
  \code{\link{newestFile}}, \code{\link{printDebug}},
  \code{\link{renameColumn}}, \code{\link{rmInfinite}},
  \code{\link{rmNA}}, \code{\link{rmNULL}},
  \code{\link{sclass}}, \code{\link{sdim}},
  \code{\link{setCLranges}}, \code{\link{setPrompt}},
  \code{\link{ssdim}}
}
\concept{jam practical functions}
