% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{isFALSEV}
\alias{isFALSEV}
\title{Vectorized isFALSE}
\usage{
isFALSEV(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{additional arguments are ignored}
}
\description{
Vectorized isFALSE
}
\details{
This function applies three criteria to an input vector, to
determine if each entry in the vector is FALSE:
\enumerate{
\item It must be class \code{logical}.
\item It must not be \code{NA}.
\item It must evaluate as \code{FALSE}.
}
}
\seealso{
Other jam practical functions: 
\code{\link{breakDensity}()},
\code{\link{checkLightMode}()},
\code{\link{check_pkg_installed}()},
\code{\link{colNum2excelName}()},
\code{\link{color_dither}()},
\code{\link{diff_functions}()},
\code{\link{exp2signed}()},
\code{\link{fileInfo}()},
\code{\link{fixYellow}()},
\code{\link{getAxisLabel}()},
\code{\link{handleArgsText}()},
\code{\link{heads}()},
\code{\link{isTRUEV}()},
\code{\link{jamba}},
\code{\link{jargs}()},
\code{\link{kable_coloring}()},
\code{\link{lldf}()},
\code{\link{log2signed}()},
\code{\link{make_html_styles}()},
\code{\link{make_styles}()},
\code{\link{match_unique}()},
\code{\link{mergeAllXY}()},
\code{\link{middle}()},
\code{\link{minorLogTicks}()},
\code{\link{newestFile}()},
\code{\link{printDebug}()},
\code{\link{renameColumn}()},
\code{\link{rmInfinite}()},
\code{\link{rmNAs}()},
\code{\link{rmNA}()},
\code{\link{rmNULL}()},
\code{\link{sclass}()},
\code{\link{sdim}()},
\code{\link{setPrompt}()}
}
\concept{jam practical functions}
