% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{isTRUEV}
\alias{isTRUEV}
\title{Vectorized isTRUE}
\usage{
isTRUEV(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{additional arguments are ignored}
}
\description{
Vectorized isTRUE
}
\details{
This function applies three criteria to an input vector, to
determine if each entry in the vector is TRUE:
\enumerate{
\item It must be class \code{logical}.
\item It must not be \code{NA}.
\item It must evaluate as \code{TRUE}.
}
}
\seealso{
Other jam practical functions: 
\code{\link{applyCLrange}()},
\code{\link{breakDensity}()},
\code{\link{checkLightMode}()},
\code{\link{colNum2excelName}()},
\code{\link{exp2signed}()},
\code{\link{fileInfo}()},
\code{\link{fixYellowHue}()},
\code{\link{fixYellow}()},
\code{\link{getAxisLabel}()},
\code{\link{handleArgsText}()},
\code{\link{isFALSEV}()},
\code{\link{jamba}},
\code{\link{jargs}()},
\code{\link{kable_coloring}()},
\code{\link{lldf}()},
\code{\link{log2signed}()},
\code{\link{make_html_styles}()},
\code{\link{make_styles}()},
\code{\link{mergeAllXY}()},
\code{\link{minorLogTicks}()},
\code{\link{newestFile}()},
\code{\link{printDebugI}()},
\code{\link{printDebug}()},
\code{\link{renameColumn}()},
\code{\link{rmInfinite}()},
\code{\link{rmNA}()},
\code{\link{rmNULL}()},
\code{\link{sclass}()},
\code{\link{sdima}()},
\code{\link{sdim}()},
\code{\link{setCLranges}()},
\code{\link{setPrompt}()},
\code{\link{ssdima}()},
\code{\link{ssdim}()}
}
\concept{jam practical functions}
