% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{isTRUEV}
\alias{isTRUEV}
\title{Vectorized isTRUE}
\usage{
isTRUEV(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{additional arguments are ignored}
}
\value{
\code{logical} vector with length matching \code{x}.
}
\description{
Vectorized isTRUE
}
\details{
This function applies three criteria to an input vector, to
determine if each entry in the vector is TRUE:
\enumerate{
\item It must be class \code{logical}.
\item It must not be \code{NA}.
\item It must evaluate as \code{TRUE}.
}
}
\examples{
isTRUEV(c(TRUE, FALSE, NA, TRUE))

}
\seealso{
Other jam practical functions: 
\code{\link{breakDensity}()},
\code{\link{call_fn_ellipsis}()},
\code{\link{checkLightMode}()},
\code{\link{check_pkg_installed}()},
\code{\link{colNum2excelName}()},
\code{\link{color_dither}()},
\code{\link{exp2signed}()},
\code{\link{getAxisLabel}()},
\code{\link{isFALSEV}()},
\code{\link{jargs}()},
\code{\link{kable_coloring}()},
\code{\link{lldf}()},
\code{\link{log2signed}()},
\code{\link{middle}()},
\code{\link{minorLogTicks}()},
\code{\link{newestFile}()},
\code{\link{printDebug}()},
\code{\link{reload_rmarkdown_cache}()},
\code{\link{renameColumn}()},
\code{\link{rmInfinite}()},
\code{\link{rmNA}()},
\code{\link{rmNAs}()},
\code{\link{rmNULL}()},
\code{\link{setPrompt}()}
}
\concept{jam practical functions}
