% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-calc-density.R
\name{jamCalcDensity}
\alias{jamCalcDensity}
\title{Calculate scatter plot point density}
\usage{
jamCalcDensity(x, nbin, bandwidth = NULL, range.x)
}
\arguments{
\item{x}{\code{numeric} matrix with two columns representing x,y coordinates.}

\item{nbin}{\code{integer} number of bins to subdivide the scatterplot,
expanded to length 2 to accommodate x and y axis bins.}

\item{bandwidth}{\code{numeric} or \code{NULL} representing the bandwidth used
for point density determination.}

\item{range.x}{\code{numeric} vector length 2 representing the range of
values to consider for point density.}
}
\description{
Calculate scatter plot point density
}
\details{
This function is called internally by \code{plotSmoothScatter()},
and is an equivalent replacement for
\code{grDevices:::.smoothScatterCalcDensity()}.
}
\concept{jam internal functions}
