% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{jargs}
\alias{jargs}
\title{Show R function arguments jam-style}
\usage{
jargs(
  x,
  grepString = NULL,
  sortVars = FALSE,
  asList = TRUE,
  useColor = TRUE,
  lightMode = NULL,
  Crange = getOption("jam.Crange"),
  Lrange = getOption("jam.Lrange"),
  adjustRgb = getOption("jam.adjustRgb"),
  useCollapseBase = ", ",
  verbose = FALSE,
  debug = 0,
  ...
)
}
\arguments{
\item{x}{function or character name of a function.}

\item{grepString}{NULL, logical, or character grep string used to filter
names of function parameters. If logical, it is assumed to be
sortVars, and indicates whether to sort the parameter names.}

\item{sortVars}{logical whether to sort the function parameter names.}

\item{asList}{logical whether to display one entry per line (TRUE), or
display results as a data.frame.}

\item{useColor}{logical whether to display results in color, if the crayon
package is available, and terminal console is capable.}

\item{lightMode}{boolean or NULL, indicating whether the text background
color is light, thus imposing a maximum brightness for colors displayed.
It use lightMode if defined by the function caller, otherwise it will
use options("jam.lightMode") if defined, lastly it will attempt to detect
whether running inside Rstudio by checking the environment variable
"RSTUDIO", and if so it will assume lightMode==TRUE.}

\item{Crange}{numeric range of chroma values, ranging
between 0 and 100. When NULL, default values will be
assigned to Crange by \code{setCLranges()}.}

\item{Lrange}{numeric range of luminance values, ranging
between 0 and 100. When NULL, default values will be
assigned to Lrange by \code{setCLranges()}.}

\item{adjustRgb}{numeric value adjustment used during the conversion of
RGB colors to ANSI colors, which is inherently lossy. If not defined,
it uses the default returned by \code{setCLranges()} which itself uses
\code{getOption("jam.adjustRgb")} with default=0. In order to boost
color contrast, an alternate value of -0.1 is suggested.}

\item{useCollapseBase}{character string used to combine multiple parameter
values.}

\item{verbose}{logical whether to print verbose output.}

\item{debug}{integer value, greater than 0 will cause debug-type verbose
output, useful because parameters are hard!}
}
\description{
Show R function arguments jam-style
}
\details{
This function displays R function arguments, organized with one argument
per line, and colorized using the \code{crayon} package if
installed.

It has an optional subset ability, using \code{grepString} which filters
argument names by pattern matching. This feature is intended to help find
the parameter name using a substring, which is particularly helpful for
R functions which contain numerous parameters, and which have not obscured
them by use of generic function wrapper. Generic functions very often
contain no useful parameters, making it difficult to discover required
parameters without reading the function documentation from the proper
dispatched function and calling package.

It also by default sorts parameters by name, to make it easier to find
from among many parameters.
}
\examples{
args(jargs)
jargs(jargs)

# retrieve parameters involving notes from imageByColors
jargs(imageByColors, "note")

}
\seealso{
Other jam practical functions: 
\code{\link{applyCLrange}()},
\code{\link{breakDensity}()},
\code{\link{checkLightMode}()},
\code{\link{check_pkg_installed}()},
\code{\link{colNum2excelName}()},
\code{\link{color_dither}()},
\code{\link{diff_functions}()},
\code{\link{exp2signed}()},
\code{\link{fileInfo}()},
\code{\link{fixYellowHue}()},
\code{\link{fixYellow}()},
\code{\link{getAxisLabel}()},
\code{\link{handleArgsText}()},
\code{\link{heads}()},
\code{\link{isFALSEV}()},
\code{\link{isTRUEV}()},
\code{\link{jamba}},
\code{\link{kable_coloring}()},
\code{\link{lldf}()},
\code{\link{log2signed}()},
\code{\link{make_html_styles}()},
\code{\link{make_styles}()},
\code{\link{match_unique}()},
\code{\link{mergeAllXY}()},
\code{\link{minorLogTicks}()},
\code{\link{newestFile}()},
\code{\link{printDebugI}()},
\code{\link{printDebug}()},
\code{\link{renameColumn}()},
\code{\link{rmInfinite}()},
\code{\link{rmNA}()},
\code{\link{rmNULL}()},
\code{\link{sclass}()},
\code{\link{sdima}()},
\code{\link{sdim}()},
\code{\link{setCLranges}()},
\code{\link{setPrompt}()},
\code{\link{ssdima}()},
\code{\link{ssdim}()}
}
\concept{jam practical functions}
