% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{kable_coloring}
\alias{kable_coloring}
\title{Extend kableExtra colorization of tables}
\usage{
kable_coloring(df, colorSub = NULL, background_as_tile = TRUE,
  color_cells = TRUE, row_color_by = NULL, sep = "_",
  return_type = c("kable", "data.frame"), verbose = FALSE, ...)
}
\arguments{
\item{df}{`data.frame`}

\item{colorSub}{named vector of R colors, whose names match entries
in the `data.frame` which are given these assigned colors.}

\item{background_as_tile}{boolean defining whether a cell background
color will appear as a rounded tile (if TRUE), or a rectangle
(if FALSE.)}

\item{color_cells}{boolean indicating whether to color individual cells}

\item{row_color_by}{optional vector of `colnames` found in `df`
indicating how to colorize entire rows of a table. The typical
colorization by cell will not fill the whole cell per row, so
this colorization can be helpful to fill the remaining area.}

\item{sep}{character delimiter used to combine values in multiple
columns when `row_color_by` is supplied and contains multiple
`colnames`. The delimited character strings are compared to
`colorSub` to assign colors.}

\item{return_type}{character string indicating the type of data
to return. If `row_color_by` is supplied, then the only
`"kable"` is allowed, otherwise `"data.frame"` is returned.
Note that the `"kable"` option runs `%>% row_spec()` which
returns an object only usable in downstream `kable` calls.}

\item{verbose}{boolean indicating whether to print verbose output.}
}
\value{
`data.frame` or `kable` data.frame dependent upon the
   `return_type` argument. Note that even the `data.frame` returned
   will have colors encoded into each cell, so it will likely
   not be possible to manipulate as normal.
}
\description{
Extend kableExtra colorization of tables in Rmarkdown files
}
\details{
This function extends the `kableExtra` package, and is only
available for use if the `kableExtra` package is installed. It is
intended to allow specific color assignment of elements in a
data.frame, but otherwise uses the `kableExtra` functions to
apply those colors.
}
\examples{
# Note the output will produce HTML code, which will
# open the Rstudio viewer when run inside Rstudio.
kable_coloring(
   df=data.frame(A=letters[1:5], B=LETTERS[1:5]),
   colorSub=nameVector(rainbow(10), c(letters[1:5], LETTERS[1:5])),
   row_color_by="B")

}
