% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{make_styles}
\alias{make_styles}
\title{vectorized make_styles for crayon output}
\usage{
make_styles(style, text, bg = FALSE, grey = FALSE, colors = num_colors(),
  checkSat = TRUE, satCutoff = 0.01, lightMode = checkLightMode(),
  adjustRgb = getOption("jam.adjustRgb"), adjustPower = 1.5,
  Lrange = NULL, Crange = NULL, subTransparent = "grey45",
  alphaPower = 2, verbose = FALSE, ...)
}
\arguments{
\item{style}{vector of one or more styles}

\item{text}{vector of one or more character values}

\item{bg}{NULL or a vector of one or more background styles}

\item{colors}{integer number of colors allowed for console output}

\item{adjustRgb}{numeric value adjustment used during the conversion of
RGB colors to ANSI colors, which is inherently lossy. If not defined,
it uses the default returned by `setCLranges()` which itself uses
\code{getOption("jam.adjustRgb")} with default=0. In order to boost
color contrast, an alternate value of -0.1 is suggested.}

\item{adjustPower}{numeric adjustment power factor}

\item{subTransparent}{color used to substitute for "transparent" which
a valid R color, but not a valid color for the crayon package.}

\item{alphaPower}{numeric value, used to adjust the RGB values for alpha
values less than 255, by raising the ratio to 1/alphaPower, which takes
the ratio of square roots.  alphaPower=100 for minimal adjustment.}

\item{verbose}{logical whether to print verbose output}

\item{...}{additional parameters are ignored}

\item{checkSet}{logical whether to check if the color is extremely
low saturation, in which case grey==TRUE is set.}

\item{setCutoff}{numeric cutoff for color saturation when checkSat==TRUE.}

\item{lRange}{lightness range, default c(0,1)}

\item{sRange}{saturation range, default c(0,1)}
}
\description{
vectorized make_styles for crayon output
}
\details{
This function is essentially a vectorized version of
\code{\link[crayon]{make_style}} in order to style a vector of
character strings with a vector of foreground and background styles.
}
