% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-string.r
\name{mixedSortDF}
\alias{mixedSortDF}
\title{sort data.frame keeping numeric values in proper order}
\usage{
mixedSortDF(df, byCols = 1:ncol(df), na.last = TRUE, decreasing = NULL,
  useRownames = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{df}{data.frame input}

\item{byCols}{integer vector referring to the order of columns to be
used by \code{\link{mmixedOrder}} to order the data.frame. Note that
negative values will reverse the sort order for the corresponding
column number.}

\item{na.last}{logical whether NA values should be ranked last}

\item{decreasing}{NULL or a logical vector indicating which columns
in \code{byCols} should be sorted in decreasing order. By default, the
sign(byCols) is used to define the sort order of each column, but it
can be explicitly overridden with this \code{decreasing} parameter.}

\item{useRownames}{logical whether to use \code{rownames(df)} as a last
tiebreaker in the overall rank ordering. This parameter has the primary
effect of assuring a reproducible result, provided the rownames are
consistently defined, or if rownames are actually row numbers.}

\item{verbose}{logical whether to print verbose output}
}
\value{
data.frame whose rows are ordered using \code{\link{mmixedOrder}}
}
\description{
sort data.frame keeping numeric values in proper order
}
\details{
This function is a wrapper around \code{\link{mmixedOrder}} so it operates
on data.frame columns in the proper order, using logic similar that used
by \code{\link{order}} when operating on data.frames. The sort order logic
is fully described in \code{\link{mixedSort}} and \code{\link{mixedOrder}}.
}
\examples{
x <- c("miR-12","miR-1","miR-122","miR-1b", "miR-1a","miR-2");
g <- rep(c("groupB", "groupA"), c(2,4));
df <- data.frame(group=g, miRNA=x, stringsAsFactors=FALSE);
mixedSortDF(df);

}
\seealso{
\code{\link{mixedOrder}}, \code{\link{mixedSort}}
}
