% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-string.r
\name{mmixedOrder}
\alias{mmixedOrder}
\title{order alphanumeric values from a list}
\usage{
mmixedOrder(..., na.last = TRUE, decreasing = FALSE, verbose = FALSE,
  ignore.case = TRUE, matrixAsDF = TRUE)
}
\arguments{
\item{...}{parameters treated as a list of vectors to be ordered in
proper order, based upon the mechanism by \code{\link[base]{order}}.}

\item{na.last, decreasing, verbose, ignore.case}{parameters sent to
\code{\link{mixedOrder}}.}

\item{matrixAsDF}{logical if \code{...} supplies only one matrix object,
whether to convert it to data.frame, the coerce to a list, before
processing. By default, in the event only one matrix object is supplied,
this conversion is performed, in order to define a sort order based upon
each column in order.}
}
\value{
integer vector of row orders
}
\description{
order alphanumeric values from a list
}
\details{
This function is a minor extension to \code{\link{mixedOrder}} in that
it accepts list input, similar to how \code{\link[base]{order}} operates.
This function is mainly useful when sorting something like a data.frame,
where ties in column 1 should be maintained then broken by non-equal
values in column 2.

In fact, \code{\link{mixedSortDF}} calls this \code{mmixedOrder} function,
in order to sort a data.frame properly by column.

See \code{\link{mixedOrder}} and \code{\link{mixedSort}} for a better
description of how the sort order logic operates.
}
\seealso{
\code{\link{mixedOrder}}, \code{\link{mixedSort}}

Other jam string functions: \code{\link{cPasteUnique}},
  \code{\link{cPaste}}, \code{\link{igrepHas}},
  \code{\link{igrep}}, \code{\link{mixedOrder}},
  \code{\link{mixedSortDF}}, \code{\link{mixedSort}},
  \code{\link{provigrep}}, \code{\link{unigrep}},
  \code{\link{uniques}}, \code{\link{unvigrep}},
  \code{\link{vgrep}}, \code{\link{vigrep}}
}
\concept{jam string functions}
