% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-string.r
\name{nameVectorN}
\alias{nameVectorN}
\title{define a named vector using vector names}
\usage{
nameVectorN(x, makeNamesFunc = makeNames, ...)
}
\arguments{
\item{x}{vector or any object which has names available via \code{names(x)}}

\item{makeNamesFunc}{function used to create unique names, in the event that
the names(x) are not unique.}
}
\value{
vector of names, whose names are uniquely assigned using
   \code{\link{makeNames}} using the values of the vector.
}
\description{
define a named vector using vector names
}
\details{
This function creates a vector from the names of the input vector,
then assigns the same as names. The utility is mainly for
\code{\link{lapply}} functions which maintain the name of a vector
in its output. The reason to run \code{\link{lapply}} using names
is so the lapply function is operating only on the name and not the
data it references, which can be convenient when the name of the element
is useful to known inside the function body. The reason to name the names,
is so the list object returned by \code{\link{lapply}} is also named
with these same consistent names.

Consider a list of data.frames, each of which represents stats results
from a contrast and fold change. The data.frame may not indicate the name
of the contrast, while the list itself may be named by the contrast.
One would \code{lapply(nameVectorN(listDF), function(iName)iName)} which
allows the internal function access to the name of each list element. This
could for example be added to the data.frame.
}
\examples{
L <- nameVector(1:5, LETTERS[1:5]);
nameVectorN(L);

}
