% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{padString}
\alias{padString}
\title{pad a character string to a fixed length}
\usage{
padString(
  x,
  stringLength = max(nchar(x)),
  padCharacter = " ",
  justify = "left",
  ...
)
}
\arguments{
\item{x}{character vector}

\item{stringLength}{integer length for the resulting character strings
in \code{x}. By default, all strings are padded to the length of the
longest entry, however stringLength can be defined to impose strict
number of characters for all entries.}

\item{padCharacter}{single character used for padding.}

\item{justify}{character value with "left", "right", "center" to indicate
alignment of the resulting text string.}

\item{...}{additional parameters are ignored.}
}
\value{
character vector of length(x)
}
\description{
pad a character string to a fixed length
}
\examples{
padString(c("one","two","three"));
padString(c("one","two","three","four"), padCharacter="_", justify="center");

}
\seealso{
Other jam string functions: 
\code{\link{asSize}()},
\code{\link{breaksByVector}()},
\code{\link{cPasteSU}()},
\code{\link{cPasteS}()},
\code{\link{cPasteUnique}()},
\code{\link{cPasteU}()},
\code{\link{cPaste}()},
\code{\link{fillBlanks}()},
\code{\link{formatInt}()},
\code{\link{gsubOrdered}()},
\code{\link{makeNames}()},
\code{\link{mixedOrder}()},
\code{\link{mixedSortDF}()},
\code{\link{mixedSorts}()},
\code{\link{mixedSort}()},
\code{\link{mmixedOrder}()},
\code{\link{nameVectorN}()},
\code{\link{nameVector}()},
\code{\link{padInteger}()},
\code{\link{pasteByRowOrdered}()},
\code{\link{pasteByRow}()},
\code{\link{sizeAsNum}()},
\code{\link{tcount2}()},
\code{\link{tcount}()},
\code{\link{ucfirst}()},
\code{\link{uniques}()}
}
\concept{jam string functions}
