% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{pasteByRowOrdered}
\alias{pasteByRowOrdered}
\title{Paste data.frame rows into an ordered factor}
\usage{
pasteByRowOrdered(x, sep = "_", na.rm = TRUE, condenseBlanks = TRUE,
  includeNames = FALSE, ...)
}
\arguments{
\item{x}{data.frame}

\item{sep}{character separator to use between columns}

\item{na.rm}{logical whether to remove NA values, or include them as "NA"}

\item{condenseBlanks}{logical whether to condense blank or empty values
without including an extra delimiter between columns.}

\item{includeNames}{logical whether to include the colname delimited
prior to the value, using sepName as the delimiter.}

\item{...}{additional arguments are passed to \code{jamba::pasteByRow()}.}

\item{verbose}{logical whether to print verbose output.}
}
\description{
Paste data.frame rows into an ordered factor
}
\details{
This function is an extension to \code{jamba::pasteByRow()} which
pastes rows from a \code{data.frame} into a character vector. This
function orders the resulting character vector after ordering
the \code{data.frame} with \code{jamba::mixedSortDF()}, which maintains
the proper factor level ordering of each column.

Note that the order of columns in the data.frame is the sort
order of factors. In future, the column sort order may become
an optional parameter.

Note also that \code{jamba::mixedSortDF()} uses \code{jamba::mixedSort()}
which itself performs alphanumeric sort in order to keep
values in proper numeric order where possible.
}
