% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-plots.r
\name{plotPolygonDensity}
\alias{plotPolygonDensity}
\title{Plot distribution and histogram overlay}
\usage{
plotPolygonDensity(x, doHistogram = TRUE, doPolygon = TRUE,
  barCol = "#00337799", col = NULL,
  histBorder = makeColorDarker(barCol, darkFactor = 1.5),
  colAlphas = c(0.8, 0.6, 0.9), hueShift = -0.1,
  darkFactors = c(-1.3, 1, 3), polyCol = "#00449977",
  polyBorder = makeColorDarker(polyCol), lwd = 2, las = 2,
  u5.bias = 0, pretty.n = 10, bw = NULL, breaks = 100,
  width = NULL, densityBreaksFactor = 3, axisFunc = axis,
  bty = "l", cex.axis = 1.5, doPar = TRUE, heightFactor = 0.95,
  weightFactor = NULL, main = "Histogram distribution", xaxs = "i",
  yaxs = "i", log = NULL, xScale = c("default", "log10", "sqrt"),
  logFloorMethod = c("detect", "+1", "floor1", "none"),
  usePanels = TRUE, useOnePanel = FALSE, ablineV = NULL,
  ablineH = NULL, ablineVcol = "#44444499", ablineHcol = "#44444499",
  ablineVlty = "solid", ablineHlty = "solid", removeNA = TRUE,
  add = FALSE, ylimQuantile = 0.99, ylim = NULL, verbose = FALSE,
  ...)
}
\arguments{
\item{x}{numeric vector, or numeric matrix.}

\item{doHistogram}{logical indicating whether to plot histogram bars.}

\item{doPolygon}{logical indicating whether to plot the density polygon.}

\item{barCol, polyCol, polyBorder, histBorder}{colors used when \code{col} is
not supplied. They define colors for the histogram bars, polygon
fill, polygon border, and histogram bar border, respectively.}

\item{col}{color or vector of colors to apply to plot panels.}

\item{u5.bias, pretty.n}{parameters sent to \code{base::pretty()} for axis
label positioning.}

\item{bw}{text bandwidth name, used in the density calculation, sent
to \code{jamba::breakDensity()}. By default \code{stats::density()} calls a
very smooth density kernel, which obscures finer details. By default,
\code{jamba::breakDensity()} uses a more detailed kernel.}

\item{breaks}{numeric breaks sent to \code{hist} to define the number of
histogram bars.}

\item{densityBreaksFactor}{numeric factor controlling the level of
detail in the density, sent to \code{jamba::breakDensity()}.}

\item{log}{character vector, optionally containing "x" and/or "y" to
apply the appropriate transformation. If "x" then it sets
\code{xScale="log10"}.}

\item{xScale}{character string defining the x-axis transformation:
"default" applies no transform; "log10" applies a log10 transform;
"sqrt" applies a sqrt transform.}

\item{usePanels}{logical indicating whether to separate
the density plots into panels when \code{x} contains multiple columns.
When \code{useOnePanel=FALSE} the panels will be defined so that all
columns will fit on one page.}

\item{useOnePanel}{logical indicating whether to define multiple panels
on one page. Therefore \code{useOnePanel=TRUE} will create multiple
pages with one panel on each page, which may work well for
output in multi-page PDF files.}

\item{ablineV, ablineH}{abline vertical and horizontal positions,
respectively. These values are mostly helpful in multi-panel plots,
since they draw consistent lines on each panel.}

\item{removeNA}{logical indicating whether to remove NA values
prior to running histogram and density calculations. Presence
of NA values generally causes both functions to fail.}

\item{ylimQuantile}{numeric value between 0 and 1, indicating the
quantile value of the density \code{y} values to use for the ylim. This
threshold is only applied when \code{ylim} is NULL.}

\item{ylim, xlim}{numeric y-axis and x-axis ranges, respectively. When NULL,
the x-axis range is determined for each plot panel.}

\item{verbose}{logical indicating whether to print verbose output.}
}
\description{
Plot distribution and histogram overlay
}
\details{
This function is a wrapper around \code{graphics::hist()} and
\code{stats::density()}, with enough customization to cover
most of the situations that need customization.

For example \code{log="x"} will automatically log-scale the x-axis,
keeping the histogram bars uniformly sized. Alternatively,
\code{xScale="sqrt"} will square root transform the data, and
transform the x-axis while keeping the numeric values constant.

It also takes
care of scaling the density height to be reasonably similar to
the histogram bar height, using the 99th quantile of the y-axis
value, which helps prevent outlier peaks from dominating the
y-axis range, thus obscuring interesting smaller features.

If supplied with a data matrix, this function will create a layout
with ncol(x) panels, and plot the distribution of each column
in its own panel, using categorical colors from
\code{colorjam::rainbowJam()}.

By default NA values are ignored, and the distributions are the
non-NA values.

Colors can be controlled using the parameter \code{col}, but can
be specifically defined for bars with \code{barCol} and the polygon
with \code{polyCol}.
}
\examples{
# basic density plot
plotPolygonDensity(rnorm(2000), main="basic polygon density plot");

# fewer breaks
plotPolygonDensity(rnorm(2000), breaks=20,
   main="breaks=20");

# log-scaled x-axis
plotPolygonDensity(10^(3+rnorm(2000)), log="x",
   breaks=50,
   main="log-scaled x-axis");

}
\seealso{
Other jam plot functions: \code{\link{adjustAxisLabelMargins}},
  \code{\link{decideMfrow}}, \code{\link{getPlotAspect}},
  \code{\link{imageByColors}}, \code{\link{imageDefault}},
  \code{\link{minorLogTicksAxis}}, \code{\link{nullPlot}},
  \code{\link{plotSmoothScatter}},
  \code{\link{shadowText}}, \code{\link{showColors}},
  \code{\link{smoothScatterJam}}, \code{\link{sqrtAxis}},
  \code{\link{usrBox}}
}
\concept{jam plot functions}
