% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{printDebug}
\alias{printDebug}
\title{print colorized output to R console}
\usage{
printDebug(
  ...,
  fgText = NULL,
  bgText = NULL,
  fgTime = "cyan",
  timeStamp = TRUE,
  comment = TRUE,
  formatNumbers = TRUE,
  trim = TRUE,
  digits = NULL,
  nsmall = 0L,
  justify = "left",
  big.mark = "",
  small.mark = "",
  zero.print = NULL,
  width = NULL,
  doColor = NULL,
  splitComments = FALSE,
  collapse = "",
  sep = ",",
  doReset = NULL,
  detectColors = TRUE,
  darkFactor = c(1, 1.5),
  sFactor = c(1, 1.5),
  lightMode = NULL,
  Crange = getOption("jam.Crange"),
  Lrange = getOption("jam.Lrange"),
  removeNA = FALSE,
  replaceNULL = NULL,
  adjustRgb = getOption("jam.adjustRgb"),
  byLine = FALSE,
  verbose = FALSE,
  indent = "",
  keepNA = TRUE,
  file = "",
  append = TRUE,
  invert = FALSE,
  htmlOut = FALSE,
  x
)
}
\arguments{
\item{...}{text to be printed to the R console.}

\item{fgText}{vector of R compatible colors, or a list of vectors of
R compatible colors, to define the foreground colors. In the latter
case, each vector is applied to each list item from '...'}

\item{bgText}{vector of R compatible colors, or a list of vectors,
to define the background color.}

\item{fgTime}{character R color to colorize the time}

\item{timeStamp}{logical whether to include a time stamp in output}

\item{comment}{logical whether to prefix output with '##' as a comment}

\item{formatNumbers}{logical whether to format numbers using
\code{\link[base]{format}} which controls the number of digits displayed.}

\item{trim, digits, nsmall, justify, big.mark, small.mark, zero.print, width}{parameters sent to the \code{\link[base]{format}} function.}

\item{doColor}{NULL or logical indicating whether to colorize output. If
NULL it detects whether the crayon package is available and console
color is enabled.}

\item{splitComments}{logical whether to color each element independently
without light-dark alternating pattern.}

\item{collapse}{character collapse string used to separate list items,
by default "" so text separation is expected in the input data.}

\item{sep}{character separator used to separate vector elements, when
a list items contains a vector.}

\item{doReset}{logical or \code{NULL}, indicating whether to apply
\code{crayon::reset()} to the delimiter \code{sep}. When \code{doReset=TRUE} the
style on the delimiter is forced to reset, using \code{crayon::reset()},
or to remove pre-existing style with \code{crayon::strip_style()}. When
\code{doReset=NULL} and \code{sep} contains ANSI escape characters, they are
left as-is; when \code{doReset=NULL} and \code{sep} does not contain ANSI escape
characters, \code{sep} becomes \code{crayon::reset(sep)} which forces the style
to be reset between printed values.}

\item{detectColors}{logical whether to detect and potentially try to
correct console color capabilities.}

\item{darkFactor}{numeric darkness to apply to alternative vector values
when using alternating light-dark color shading.}

\item{sFactor}{numeric color saturation to apply to alternative vector
values when using alternating light-dark color shading.}

\item{lightMode}{boolean or NULL, indicating whether the text background
color is light, thus imposing a maximum brightness for colors displayed.
It use lightMode if defined by the function caller, otherwise it will
use options("jam.lightMode") if defined, lastly it will attempt to detect
whether running inside Rstudio by checking the environment variable
"RSTUDIO", and if so it will assign lightMode TRUE.}

\item{Crange}{numeric range of chroma values, ranging
between 0 and 100. When NULL, default values will be
assigned to Crange by \code{setCLranges()}.}

\item{Lrange}{numeric range of luminance values, ranging
between 0 and 100. When NULL, default values will be
assigned to Lrange by \code{setCLranges()}.}

\item{removeNA}{logical whether to remove NA values and not print to
the console.}

\item{replaceNULL}{character or NULL, optionally replace NULL elements
with non-NULL character value.}

\item{adjustRgb}{numeric value adjustment used during the conversion of
RGB colors to ANSI colors, which is inherently lossy. If not defined,
it uses the default returned by \code{setCLranges()} which itself uses
\code{getOption("jam.adjustRgb")} with default=0. In order to boost
color contrast, an alternate value of -0.1 is suggested.}

\item{byLine}{logical whether to delimit lists by line instead of
using collapse to combine them onto one line.}

\item{verbose}{logical whether to print verbose output}

\item{indent}{character optional characters used as a prefix to indent
output.}

\item{file}{passed to \code{cat}, to allow sending output to
a specified file.}

\item{append}{logical whether to append output, relevant only when
\code{file} specifies a filename.}

\item{invert}{logical indicating whether foreground and background
colors should be switched.}

\item{htmlOut}{logical indicating whether to print HTML span
output, using format
\verb{<span style="color:fg;background-color:bg">text</span>}.}
}
\value{
This function is called for the by-product of printing
debug output, it returns \code{invisible(NULL)}, no output.
}
\description{
print colorized output to R console
}
\details{
This function prints colorized output to the R console, with some
rules for colorizing the output to help visually distinguish items.
Its output also by default begins with comment '#' characters, a
datetimestamp, so it the output is copied back into the R console it
will not cause a new command to be run.

The colorization uses a vector or list of colors for fgText (foreground)
and bgText (background.), applied to each item in '...'. When an item in
'...' is a vector, each vector element is colored alternating light
and dark from that base color, to give visual indication of each element.
The next item in '...' receives the next color from fgText, and so on.
Colors in fgText are recycled to the length of '...'
}
\examples{
printDebug("Testing ", "default ", "printDebug().");
printDebug("List of vectors:", c("one", "two", "three"));

# By default, there is no space between separate elements in `...`
printDebug("List of vectors:", c("one", "two", "three"),
   c("four", "five", "six"));
# To add a space " " between elements, use collapse
printDebug("List of vectors:", c("one", "two", "three"),
   c("four", "five", "six"), collapse=" ");

# slightly different style, one entry per line, indented:
printDebug("List of vectors:", c("one", "two", "three"),
   c("four", "five", "six"), collapse="\n   ");

# when a vector entirely contains recognized colors,
# the colors are used in the output
printDebug(c("red", "blue", "yellow"));

# When the vector contains colors, the names are used as the label
color_vector <- jamba::nameVector(c("red", "blue", "green","orange"),
   c("group_A", "group_B", "group_C", "group_D"));
printDebug(color_vector);

# Remember the sister function that inverses the colors
printDebugI(color_vector);

}
\seealso{
Other jam practical functions: 
\code{\link{applyCLrange}()},
\code{\link{breakDensity}()},
\code{\link{checkLightMode}()},
\code{\link{colNum2excelName}()},
\code{\link{exp2signed}()},
\code{\link{fileInfo}()},
\code{\link{fixYellowHue}()},
\code{\link{fixYellow}()},
\code{\link{getAxisLabel}()},
\code{\link{handleArgsText}()},
\code{\link{isFALSEV}()},
\code{\link{isTRUEV}()},
\code{\link{jamba}},
\code{\link{jargs}()},
\code{\link{kable_coloring}()},
\code{\link{log2signed}()},
\code{\link{make_html_styles}()},
\code{\link{make_styles}()},
\code{\link{mergeAllXY}()},
\code{\link{minorLogTicks}()},
\code{\link{newestFile}()},
\code{\link{printDebugI}()},
\code{\link{renameColumn}()},
\code{\link{rmInfinite}()},
\code{\link{rmNA}()},
\code{\link{rmNULL}()},
\code{\link{sclass}()},
\code{\link{sdima}()},
\code{\link{sdim}()},
\code{\link{setCLranges}()},
\code{\link{setPrompt}()},
\code{\link{ssdima}()},
\code{\link{ssdim}()}
}
\concept{jam practical functions}
