% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{printDebug}
\alias{printDebug}
\title{print colorized output to R console}
\usage{
printDebug(..., fgText = NULL, bgText = NULL, fgTime = "cyan",
  timeStamp = TRUE, comment = TRUE, formatNumbers = TRUE, trim = TRUE,
  digits = NULL, nsmall = 0L, justify = "left", big.mark = "",
  small.mark = "", zero.print = NULL, width = NULL, doColor = NULL,
  splitComments = FALSE, collapse = "", sep = ",", detectColors = TRUE,
  darkFactor = c(1, 1.5), sFactor = c(1, 1.5), removeNA = FALSE,
  replaceNULL = NULL, adjustRgb = -0.1, byLine = FALSE, verbose = FALSE,
  indent = "", keepNA = TRUE, file = "", append = TRUE, x)
}
\arguments{
\item{...}{text to be printed to the R console.}

\item{fgText}{vector of R compatible colors, or a list of vectors of
R compatible colors, to define the foreground colors. In the latter
case, each vector is applied to each list item from '...'}

\item{bgText}{vector of R compatible colors, or a list of vectors,
to define the background color.}

\item{fgTime}{character R color to colorize the time}

\item{timeStamp}{logical whether to include a time stamp in output}

\item{comment}{logical whether to prefix output with '##' as a comment}

\item{formatNumbers}{logical whether to format numbers using
\link{\code{format}} which controls the number of digits displayed.}

\item{trim, digits, nsmall, justify, big.mark, small.mark, zero.print, width}{parameters sent to the \link{\code{format}} function.}

\item{doColor}{NULL or logical indicating whether to colorize output. If
NULL it detects whether the crayon package is available and console
color is enabled.}

\item{splitComments}{logical whether to color each element independently
without light-dark alternating pattern.}

\item{collapse}{character collapse string used to separate list items,
by default "" so text separation is expected in the input data.}

\item{sep}{character separator used to separate vector elements, when
a list items contains a vector.}

\item{detectColors}{logical whether to detect and potentially try to
correct console color capabilities.}

\item{darkFactor}{numeric darkness to apply to alternative vector values
when using alternating light-dark color shading.}

\item{sFactor}{numeric color saturation to apply to alternative vector
values when using alternating light-dark color shading.}

\item{removeNA}{logical whether to remove NA values and not print to
the console.}

\item{replaceNULL}{character or NULL, optionally replace NULL elements
with non-NULL character value.}

\item{adjustRgb}{logical sent to \code{\link{make_styles}}, whether to
adjust RGB to prevent loss of detail with low color brightness.}

\item{byLine}{logical whether to delimit lists by line instead of
using collapse to combine them onto one line.}

\item{verbose}{logical whether to print verbose output}

\item{indent}{character optional characters used as a prefix to indent
output.}

\item{file}{passed to \code{\link{cat}}, to allow sending output to
a specified file.}

\item{append}{logical whether to append output, relevant only when
\code{file} specifies a filename.}
}
\description{
print colorized output to R console
}
\details{
This function prints colorized output to the R console, with some
rules for colorizing the output to help visually distinguish items.
Its output also by default begins with comment '#' characters, a
datetimestamp, so it the output is copied back into the R console it
will not cause a new command to be run.

The colorization uses a vector or list of colors for fgText (foreground)
and bgText (background.), applied to each item in '...'. When an item in
'...' is a vector, each vector element is colored alternating light
and dark from that base color, to give visual indication of each element.
The next item in '...' receives the next color from fgText, and so on.
Colors in fgText are recycled to the length of '...'
}
\examples{
printDebug("Testing ", "default ", "printDebug().");
printDebug("List of vectors:", c("one", "two", "three"));
printDebug("List of vectors:", c("one", "two", "three"),
   c("four", "five", "six"), collapse=" ");

# slightly different style, one entry per line, indented:
printDebug("List of vectors:", c("one", "two", "three"),
   c("four", "five", "six"), collapse="\\n   ");

# in an R console, or when writing to a log file, the
# following output text is colored
printDebug(c("red", "blue", "yellow"));

}
