% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-load-cache.R
\name{reload_rmarkdown_cache}
\alias{reload_rmarkdown_cache}
\title{Reload Rmarkdown cache}
\usage{
reload_rmarkdown_cache(
  dir = ".",
  maxnum = 1000,
  max_cache_name = NULL,
  envir = globalenv(),
  file_sort = c("ctime", "mtime"),
  dryrun = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dir}{\code{character} path to the directory that contains Rmarkdown
cache files. Each file is recognized by the file extension \code{".rdx"}.}

\item{maxnum}{\code{integer} indicating the maximum number of cache files
to re-load, in order.}

\item{max_cache_name}{\code{character} optional string indicating the
name of an Rmarkdown cache chunk where this function will stop
loading cache data. All cache files after this point will
not be loaded. This option is intended to help recreate the
data available to a particular Rmarkdown chunk during
processing.}

\item{envir}{\code{environment} where cache data will be loaded.}

\item{file_sort}{\code{character} string indicating how to sort cache files
to place them in proper order for re-loading. Note that
file modification time \code{mtime} may incorrectly sort files
that were modified after the initial processing of the
Rmarkdown file.
\itemize{
\item \code{ctime} sorts by file creation time, default
\item \code{mtime} sorts by file modification time
}}

\item{dryrun}{\code{logical} indicating whether to perform a dry-run,
which prints messages but does not process the data.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{lazyLoad()}.}
}
\value{
this function does not return data, but instead is called
for the by-product of loading data into the given \code{envir}
environment.
}
\description{
Reload Rmarkdown cache in the order files were created
}
\details{
This function is intended to help re-load Rmarkdown cache files
created during the processing/rendering of an Rmarkdown file.

By default, all cached R objects are loaded into the
global environment \code{globalenv()}.

It can be given an argumen \code{envir} to store R objects inside
a specific \code{environment}.

If supplied with \code{max_cache_name} then this function will only
load the cache chunks in order, until it recognizes that
chunk name. This option is intended to help restore the R
data available for a particular Rmarkdown chunk.
}
\seealso{
Other jam utility functions: 
\code{\link{call_fn_ellipsis}()}
}
\concept{jam utility functions}
