% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{rgb2col}
\alias{rgb2col}
\title{Convert RGB to R color}
\usage{
rgb2col(red, green = NULL, blue = NULL, alpha = NULL, names = NULL,
  maxColorValue = NULL, keepNA = TRUE, ...)
}
\arguments{
\item{red}{numeric vector of red values; or RGB numeric matrix with
rownames c("red","green","blue") in any order, with optional rowname
"alpha"; or character strings with comma-separated rgb values, in 
format "100,20,10". The latter input is designed to handle web rgb
values.}

\item{green}{numeric vector, or when red is a matrix or comma-delimited
character string, this parameter is ignored.}

\item{blue}{numeric vector, or when red is a matrix or comma-delimited
character string, this parameter is ignored.}

\item{alpha}{numeric vector, or when red is a matrix or comma-delimited
character string, this parameter is ignored.}

\item{maxColorValue}{numeric maximum value for colors. If NULL then it
defaults to 1 unless there are values above 1, in which case it defaults
to 255.}

\item{keepNA}{logical whether to keep NA values, returning NA for any
input where red, green, and/or blue are NA. If keepNA==FALSE then it
substitutes 0 for any NA values.}

\item{alpha}{numeric vector, or when red is a matrix or comma-delimited
character string, this parameter is ignored.}
}
\description{
Converts RGB color matrix to R hex color
}
\details{
This function intends to augment the \code{\link{rgb}} function, which
does not handle output from \code{\link{col2rgb}}. The goal is to handle
multiple color conversions, e.g. \code{rgb2col(col2rgb("red"))}. This
function also maintains alpha transparency when supplied.

The output is named either by names(red), rownames(red), or if supplied,
the value of the parameter \code{names}.
}
\examples{
rgb2col(col2rgb("#FF000099", alpha=TRUE))

}
