% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-colors.r
\name{rgb2col}
\alias{rgb2col}
\title{Convert RGB color matrix to R color}
\usage{
rgb2col(red, green = NULL, blue = NULL, alpha = NULL, names = NULL,
  maxColorValue = NULL, keepNA = TRUE, ...)
}
\arguments{
\item{red}{numeric vector of red values; or RGB numeric matrix with
rownames c("red","green","blue") in any order, with optional rowname
"alpha"; or character strings with comma-separated rgb values, in
format "100,20,10". The latter input is designed to handle web rgb
values.}

\item{green}{numeric vector, or when red is a matrix or comma-delimited
character string, this parameter is ignored.}

\item{blue}{numeric vector, or when red is a matrix or comma-delimited
character string, this parameter is ignored.}

\item{alpha}{numeric vector, or when red is a matrix or comma-delimited
character string, this parameter is ignored.}

\item{maxColorValue}{numeric maximum value for colors. If NULL then it
defaults to 1 unless there are values above 1, in which case it defaults
to 255.}

\item{keepNA}{logical whether to keep NA values, returning NA for any
input where red, green, and/or blue are NA. If keepNA==FALSE then it
substitutes 0 for any NA values.}

\item{alpha}{numeric vector, or when red is a matrix or comma-delimited
character string, this parameter is ignored.}
}
\description{
Convert RGB color matrix to R color
}
\details{
This function intends to augment the \code{\link[grDevices]{rgb}} function, which
does not handle output from \code{\link[grDevices]{col2rgb}}. The goal is to handle
multiple color conversions, e.g. \code{rgb2col(col2rgb("red"))}. This
function also maintains alpha transparency when supplied.

The output is named either by names(red), rownames(red), or if supplied,
the value of the parameter \code{names}.
}
\examples{
# start with a color vector
# red and blue with partial transparency
colorV <- c("#FF000055", "#00339999");

# Show the output of rgb2col
# make sure to include alpha=TRUE to maintain alpha transparency
col2rgb(colorV, alpha=TRUE);

# confirm we can convert from RGB back to the same colors
rgb2col(col2rgb(colorV, alpha=TRUE));

}
\seealso{
Other jam color functions: \code{\link{alpha2col}},
  \code{\link{col2alpha}}, \code{\link{col2hcl}},
  \code{\link{col2hsv}}, \code{\link{color2gradient}},
  \code{\link{getColorRamp}}, \code{\link{hcl2col}},
  \code{\link{hsv2col}}, \code{\link{isColor}},
  \code{\link{kable_coloring}},
  \code{\link{makeColorDarker}},
  \code{\link{setTextContrastColor}}
}
\concept{jam color functions}
