% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-string.r
\name{rmInfinite}
\alias{rmInfinite}
\title{remove Infinite values}
\usage{
rmInfinite(x, infiniteValue = NULL, ...)
}
\arguments{
\item{x}{vector input}

\item{infiniteValue}{NULL to remove Infinite values, or a replacement value}

\item{...}{additional parameters are ignored}
}
\description{
remove Infinite values
}
