% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-string.r
\name{rmInfinite}
\alias{rmInfinite}
\title{remove Infinite values}
\usage{
rmInfinite(x, infiniteValue = NULL, ...)
}
\arguments{
\item{x}{vector input}

\item{infiniteValue}{NULL to remove Infinite values, or a replacement value}

\item{...}{additional parameters are ignored}
}
\value{
numeric vector with infinite values either removed, or
replaced with the supplied value.
}
\description{
remove Infinite values
}
\details{
This function removes any positive or negative infinite numerical
values, optionally replacing them with a given value or NA.
}
\seealso{
Other jam practical functions: 
\code{\link{breakDensity}()},
\code{\link{checkLightMode}()},
\code{\link{check_pkg_installed}()},
\code{\link{colNum2excelName}()},
\code{\link{color_dither}()},
\code{\link{diff_functions}()},
\code{\link{exp2signed}()},
\code{\link{fileInfo}()},
\code{\link{fixYellow}()},
\code{\link{getAxisLabel}()},
\code{\link{handleArgsText}()},
\code{\link{heads}()},
\code{\link{isFALSEV}()},
\code{\link{isTRUEV}()},
\code{\link{jamba}},
\code{\link{jargs}()},
\code{\link{kable_coloring}()},
\code{\link{lldf}()},
\code{\link{log2signed}()},
\code{\link{make_html_styles}()},
\code{\link{make_styles}()},
\code{\link{match_unique}()},
\code{\link{mergeAllXY}()},
\code{\link{minorLogTicks}()},
\code{\link{newestFile}()},
\code{\link{printDebugI}()},
\code{\link{printDebug}()},
\code{\link{renameColumn}()},
\code{\link{rmNAs}()},
\code{\link{rmNA}()},
\code{\link{rmNULL}()},
\code{\link{sclass}()},
\code{\link{sdima}()},
\code{\link{sdim}()},
\code{\link{setPrompt}()},
\code{\link{ssdima}()},
\code{\link{ssdim}()}
}
\concept{jam practical functions}
