% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-string.r
\name{rmInfinite}
\alias{rmInfinite}
\title{remove Infinite values}
\usage{
rmInfinite(x, infiniteValue = NULL, ...)
}
\arguments{
\item{x}{vector input}

\item{infiniteValue}{NULL to remove Infinite values, or a replacement value}

\item{...}{additional parameters are ignored}
}
\value{
numeric vector with infinite values either removed, or
replaced with the supplied value.
}
\description{
remove Infinite values
}
\details{
This function removes any positive or negative infinite numerical
values, optionally replacing them with a given value or NA.
}
\examples{
rmInfinite(c(1, 5, 4, 10, Inf, 1, -Inf))

rmInfinite(c(1, 5, 4, 10, Inf, 1, -Inf), infiniteValue=1000)

}
\seealso{
Other jam practical functions: 
\code{\link{breakDensity}()},
\code{\link{call_fn_ellipsis}()},
\code{\link{checkLightMode}()},
\code{\link{check_pkg_installed}()},
\code{\link{colNum2excelName}()},
\code{\link{color_dither}()},
\code{\link{exp2signed}()},
\code{\link{getAxisLabel}()},
\code{\link{isFALSEV}()},
\code{\link{isTRUEV}()},
\code{\link{jargs}()},
\code{\link{kable_coloring}()},
\code{\link{lldf}()},
\code{\link{log2signed}()},
\code{\link{middle}()},
\code{\link{minorLogTicks}()},
\code{\link{newestFile}()},
\code{\link{printDebug}()},
\code{\link{reload_rmarkdown_cache}()},
\code{\link{renameColumn}()},
\code{\link{rmNA}()},
\code{\link{rmNAs}()},
\code{\link{rmNULL}()},
\code{\link{setPrompt}()}
}
\concept{jam practical functions}
