% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-string.r
\name{rmNA}
\alias{rmNA}
\title{remove NA values}
\usage{
rmNA(x, naValue = NULL, rmNULL = FALSE, nullValue = naValue,
  rmInfinite = TRUE, infiniteValue = NULL, rmNAnames = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{x}{vector input}

\item{naValue}{NULL or single replacement value for NA entries. If NULL,
then NA entries are removed from the result.}

\item{rmNULL}{logical whether to replace NULL entries with naValue}

\item{nullValue}{NULL or single replacement value for NULL entries. If NULL,
then NULL entries are removed from the result.}

\item{rmInfinite}{logical whether to replace Infinite values with
infiniteValue}

\item{infiniteValue}{value to use when rmInfinite==TRUE to replace
entries which are Inf or -Inf.}

\item{rmNAnames}{logical whether to remove entries which have NA as the
name, regardless whether the entry itself is NA.}

\item{verbose}{logical whether to print verbose output}
}
\value{
vector with NA entries either removed, or replaced with naValue,
   and NULL entries either removed or replaced by nullValue.
}
\description{
remove NA values
}
\details{
This function removes NA values, by default shortening a vector as a result,
but optionally replacing NA and Infinite values with fixed values.
}
