% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-string.r
\name{rmNULL}
\alias{rmNULL}
\title{remove NULL entries from list}
\usage{
rmNULL(x, nullValue = NULL, ...)
}
\arguments{
\item{x}{list or other object which may contain NULL.}
}
\value{
list with NULL entries either removed, or replaced with nullValue.
This function is typically called so it removed list elements which are
NULL, resulting in a list that contains non-NULL entries. This function
can also be useful when NULL values should be changed to something else,
perhaps a character value "NULL" to be used as a label.
}
\description{
remove NULL entries from list
}
\details{
This function is a simple helper function to remove NULL from a list,
optionally replacing it with another value
}
