% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{sclass}
\alias{sclass}
\title{return the classes of a list of objects}
\usage{
sclass(x, ...)
}
\arguments{
\item{x}{an S3 object inheriting from class "list", or an S4 object.}

\item{...}{additional parameters are ignored.}
}
\description{
return the classes of a list of objects
}
\details{
This function takes a list and returns the classes for each
object in the list. In the event an object class has multiple values,
the returned object is a list, otherwise is a vector.
If \code{x} is an S4 object, then \code{slotNames(x)} is used, and
the class is returned for each S4 slot.
}
\examples{
sclass(list(LETTERS=LETTERS, letters=letters));

}
\seealso{
Other jam practical functions: \code{\link{applyCLrange}},
  \code{\link{breakDensity}}, \code{\link{checkLightMode}},
  \code{\link{colNum2excelName}}, \code{\link{fileInfo}},
  \code{\link{fixYellowHue}}, \code{\link{fixYellow}},
  \code{\link{getAxisLabel}}, \code{\link{handleArgsText}},
  \code{\link{isFALSEV}}, \code{\link{isTRUEV}},
  \code{\link{jamba}}, \code{\link{jargs}},
  \code{\link{kable_coloring}}, \code{\link{make_styles}},
  \code{\link{minorLogTicks}}, \code{\link{newestFile}},
  \code{\link{printDebug}}, \code{\link{renameColumn}},
  \code{\link{rmInfinite}}, \code{\link{rmNA}},
  \code{\link{rmNULL}}, \code{\link{sdim}},
  \code{\link{setCLranges}}, \code{\link{setPrompt}},
  \code{\link{ssdim}}

Other jam list functions: \code{\link{cPasteUnique}},
  \code{\link{cPaste}}, \code{\link{list2df}},
  \code{\link{rbindList}}, \code{\link{rlengths}},
  \code{\link{sdim}}, \code{\link{ssdim}},
  \code{\link{uniques}}
}
\concept{jam list functions}
\concept{jam practical functions}
