% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-export.r
\name{set_xlsx_colwidths}
\alias{set_xlsx_colwidths}
\title{Set column widths in Xlsx files}
\usage{
set_xlsx_colwidths(
  xlsxFile,
  sheet = 1,
  cols = seq_along(widths),
  widths = 11,
  ...
)
}
\arguments{
\item{xlsxFile}{\code{character} path to a file saved as \code{xlsx} format.}

\item{sheet}{value passed to \code{openxlsx::setColWidths()} indicating
the worksheet to affect. It can either be an \code{integer} value, or
the \code{character} name of a sheet.}

\item{cols}{\verb{integer vector} indicating the column numbers to affect.}

\item{widths}{\verb{numeric vector} indicating the width of each column
defined by \code{cols}.}

\item{...}{additional arguments are passed to \code{openxlsx::setColWidths()}.}
}
\description{
Set column widths in Xlsx files
}
\details{
This function is a light wrapper to perform these steps
from the very useful \code{openxlsx} R package:
\itemize{
\item \code{openxlsx::loadWorkbook()}
\item \code{openxlsx::setColWidths()}
\item \code{openxlsx::saveWorkbook()}
}
}
\examples{
\dontrun{
   df <- data.frame(a=LETTERS[1:5], b=1:5);
   jamba::writeOpenxlsx(x=df,
      file="jamba_test.xlsx",
      sheetName="test_jamba");

   ## By default, cols starts at column 1 and continues to length(widths)
   jamba::set_xlsx_colwidths(file="jamba_test.xlsx",
      sheetName="test_jamba",
      widths=rep(20, ncol(df))
   )
}

}
\seealso{
Other jam export functions: 
\code{\link{applyXlsxCategoricalFormat}()},
\code{\link{applyXlsxConditionalFormat}()},
\code{\link{set_xlsx_rowheights}()},
\code{\link{writeOpenxlsx}()}
}
\concept{jam export functions}
