% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-export.r
\name{set_xlsx_rowheights}
\alias{set_xlsx_rowheights}
\title{Set row heights in Xlsx files}
\usage{
set_xlsx_rowheights(
  xlsxFile,
  sheet = 1,
  rows = seq_along(heights) + 1,
  heights = 17,
  ...
)
}
\arguments{
\item{xlsxFile}{\code{character} filename to a file with ".xlsx" extension,
or \code{Workbook} object defined in the \code{openxlsx} package. When
\code{xlsxFile} is a \code{Workbook} the output is not saved to a file.}

\item{sheet}{value passed to \code{openxlsx::setRowHeights()} indicating
the worksheet to affect. It can either be an \code{integer} value, or
the \code{character} name of a sheet.}

\item{rows}{\verb{integer vector} indicating the row numbers to affect.}

\item{heights}{\verb{numeric vector} indicating the height of each column
defined by \code{rows}.}

\item{...}{additional arguments are passed to \code{openxlsx::setRowHeights()}.}
}
\value{
\code{Workbook} object as defined by the \code{openxlsx} package
is returned invisibly with \code{invisible()}. This \code{Workbook}
can be used in argument \code{wb} to provide a speed boost when
saving multiple sheets to the same file.
}
\description{
This function is a light wrapper to perform these steps
from the very useful \code{openxlsx} R package:
}
\details{
\itemize{
\item \code{openxlsx::loadWorkbook()}
\item \code{openxlsx::setRowHeights()}
\item \code{openxlsx::saveWorkbook()}
}

Note that when only the argument \code{heights} is defined,
the argument \code{rows} will point to row 2 and lower, thus
skipping the first (header) row. Define \code{rows} specifically
in order to affect the header row as well.
}
\examples{
\dontrun{
   df <- data.frame(a=LETTERS[1:5], b=1:5);
   jamba::writeOpenxlsx(x=df,
      file="jamba_test.xlsx",
      sheetName="test_jamba");

   ## by default, rows will start at row 2, skipping the header
   jamba::set_xlsx_rowheights(file="jamba_test.xlsx",
      sheetName="test_jamba",
      heights=rep(17, nrow(df))
   )

   ## to include the header row
   jamba::set_xlsx_rowheights(file="jamba_test.xlsx",
      sheetName="test_jamba",
      rows=seq_len(nrow(df)+1),
      heights=rep(17, nrow(df)+1)
   )
}

}
\seealso{
Other jam export functions: 
\code{\link{applyXlsxCategoricalFormat}()},
\code{\link{applyXlsxConditionalFormat}()},
\code{\link{readOpenxlsx}()},
\code{\link{set_xlsx_colwidths}()},
\code{\link{writeOpenxlsx}()}
}
\concept{jam export functions}
