% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{setCLranges}
\alias{setCLranges}
\title{Get chroma (C) and luminance (L) ranges for the given lightMode}
\usage{
setCLranges(lightMode = checkLightMode(), Crange = NULL, Lrange = NULL,
  adjustRgb = getOption("jam.adjustRgb"), verbose = FALSE, ...)
}
\arguments{
\item{lightMode}{boolean indicating whether the background color
is light (TRUE is bright), or dark (FALSE is dark.) By default
it calls \code{checkLightMode()} which queries \code{getOption("lightMode")}.}

\item{Crange}{numeric range of chroma values, ranging
between 0 and 100. When NULL, default values will be
assigned to Crange. When supplied, range(Crange) is used.}

\item{Lrange}{numeric range of luminance values, ranging
between 0 and 100. When NULL, default values will be
assigned to Lrange. When supplied, range(Lrange) is used.}

\item{adjustRgb}{numeric color adjustment factor, used during the
conversion of RGB colors to the ANSI-compatible colors used
by the \code{crayon} pacakge. The ANSI color range does not include
a full RGB palette, and the conversion is somewhat lossy.
By default, \code{getOptions("jam.adjustRgb")} is used to store a
globally re-usable value.}

\item{verbose}{boolean indicating whether to print verbose output.}

\item{...}{additional parameters are ignored.}
}
\value{
list with elements
\describe{
\item{Crange}{Numeric vector of length 2, defining the
chroma (C) range.}
\item{Lrange}{Numeric vector of length 2, defining the
luminance (L) range.}
\item{adjustRgb}{Numeric vector of length 1, defining the
adjustment to apply during RGB-to-ANSI color conversion.}
}
}
\description{
Return Crange, Lrange, and adjustRgb values for the given lightMode,
which helps define sensible default ranges for contrasting colors.
}
\details{
This function is intended mainly for internal use by \code{jamba}
such as \code{printDebug()}, and \code{make_styles()}. It is split into its
own function in order to help share the logic of determining some
sensible default ranges.
}
\examples{
setCLranges(lightMode=FALSE)

}
