% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{setCLranges}
\alias{setCLranges}
\title{Get chroma (C) and luminance (L) ranges for the given lightMode}
\usage{
setCLranges(
  lightMode = checkLightMode(),
  Crange = getOption("jam.Crange", c(30, 190)),
  Lrange = getOption("jam.Lrange"),
  Cgrey = getOption("jam.Cgrey", 5),
  adjustRgb = getOption("jam.adjustRgb", 0),
  setOptions = c("ifnull", "TRUE", "FALSE"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{lightMode}{boolean indicating whether the background color
is light (TRUE is bright), or dark (FALSE is dark.) By default
it calls \code{checkLightMode()} which queries \code{getOption("lightMode")}.}

\item{Crange}{numeric range of chroma values, ranging
between 0 and 100. By default, \code{getOptions("Crange")} is used,
otherwise defaults will be assigned based upon \code{lightMode}.}

\item{Lrange}{numeric range of luminance values, ranging
between 0 and 100. By default, \code{getOptions("Crange")} is used,
otherwise defaults will be assigned based upon \code{lightMode}.}

\item{Cgrey}{numeric chroma (C) value, which defines grey colors at or
below this chroma. Any colors at or below the grey cutoff will have
their C values unchanged. This mechanism prevents converting black
to red, for example. To disable the effect, set \code{Cgrey=-1}.}

\item{adjustRgb}{numeric color adjustment factor, used during the
conversion of RGB colors to the ANSI-compatible colors used
by the \code{crayon} pacakge. The ANSI color range does not include
a full RGB palette, and the conversion is somewhat lossy.
By default, \code{getOptions("jam.adjustRgb")} is used to store a
globally re-usable value.}

\item{setOptions}{character or logical whether to update \code{Crange} and \code{Lrange}
options during the subsequent call to \code{setCLranges()}. By default,
\code{"ifnull"} will update only options which were previously \code{NULL};
\code{"FALSE"} prevents modifying the global options; \code{"TRUE"} will
update these options with the current values.}

\item{verbose}{boolean indicating whether to print verbose output.}

\item{...}{additional parameters are ignored.}
}
\value{
list with elements
\describe{
\item{Crange}{Numeric vector of length 2, defining the
chroma (C) range.}
\item{Lrange}{Numeric vector of length 2, defining the
luminance (L) range.}
\item{adjustRgb}{Numeric vector of length 1, defining the
adjustment to apply during RGB-to-ANSI color conversion.}
\item{Cgrey}{Numeric vector of length 1, defining the
chroma value below which colors are considered greyscale,
where chroma ranges from 0 to 100. Value of -1 (or FALSE)
disables this logic.}
}
}
\description{
Return Crange, Lrange, and adjustRgb values for the given lightMode,
which helps define sensible default ranges for contrasting colors.
}
\details{
This function is intended mainly for internal use by \code{jamba}
such as \code{printDebug()}, and \code{make_styles()}. It is split into its
own function in order to help share the logic of determining some
sensible default ranges.
}
\examples{
setCLranges(lightMode=FALSE)

}
\seealso{
Other jam practical functions: 
\code{\link{applyCLrange}()},
\code{\link{breakDensity}()},
\code{\link{checkLightMode}()},
\code{\link{colNum2excelName}()},
\code{\link{exp2signed}()},
\code{\link{fileInfo}()},
\code{\link{fixYellowHue}()},
\code{\link{fixYellow}()},
\code{\link{getAxisLabel}()},
\code{\link{handleArgsText}()},
\code{\link{isFALSEV}()},
\code{\link{isTRUEV}()},
\code{\link{jamba}},
\code{\link{jargs}()},
\code{\link{kable_coloring}()},
\code{\link{log2signed}()},
\code{\link{make_styles}()},
\code{\link{mergeAllXY}()},
\code{\link{minorLogTicks}()},
\code{\link{newestFile}()},
\code{\link{printDebugI}()},
\code{\link{printDebug}()},
\code{\link{renameColumn}()},
\code{\link{rmInfinite}()},
\code{\link{rmNA}()},
\code{\link{rmNULL}()},
\code{\link{sclass}()},
\code{\link{sdima}()},
\code{\link{sdim}()},
\code{\link{setPrompt}()},
\code{\link{ssdima}()},
\code{\link{ssdim}()}

Other jam color functions: 
\code{\link{alpha2col}()},
\code{\link{applyCLrange}()},
\code{\link{col2alpha}()},
\code{\link{col2hcl}()},
\code{\link{col2hsv}()},
\code{\link{color2gradient}()},
\code{\link{fixYellowHue}()},
\code{\link{fixYellow}()},
\code{\link{getColorRamp}()},
\code{\link{hcl2col}()},
\code{\link{hsv2col}()},
\code{\link{isColor}()},
\code{\link{kable_coloring}()},
\code{\link{makeColorDarker}()},
\code{\link{make_styles}()},
\code{\link{rgb2col}()},
\code{\link{setTextContrastColor}()},
\code{\link{showColors}()},
\code{\link{unalpha}()},
\code{\link{warpRamp}()}
}
\concept{jam color functions}
\concept{jam practical functions}
