% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{setPrompt}
\alias{setPrompt}
\title{set R prompt with project name and R version}
\usage{
setPrompt(projectName = NULL, useColor = TRUE, projectColor = "yellow",
  bracketColor = "white", Rcolors = c("white", "white", "white"),
  PIDcolor = NA, promptColor = "white", usePid = TRUE,
  resetPrompt = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{projectName}{character string representing the active project.}

\item{useColor}{logical whether to define a color prompt if the
\code{crayon} package is installed.}

\item{projectColor, bracketColor, Rcolors, PIDcolor, promptColor}{colors
used when useColor==TRUE and the \code{\link{crayon}} package
is installed. \code{projectColor} colors the project name; \code{bracketColor}
colors the curly brackets around the project; \code{Rcolors} can be
a vector of 3 colors, colorizing "R", the "-" divider, and the
R version; \code{PIDcolor} colors the PID when \code{usePid=TRUE}; and
\code{promptColor} colors the \code{">"} at the end of the prompt.}

\item{usePid}{logical whether to include the process ID in the prompt.}

\item{resetPrompt}{logical whether to revert all changes to the prompt
back to the default R prompt, that is, no color and no projectName.}

\item{verbose}{logical whether to print verbose output}

\item{...}{additional parameters are passed to \code{make_styles()} which is
only relevant with the argument \code{useColor=TRUE}.}
}
\value{
invisible character string representing the prompt used.
}
\description{
set R prompt with project name and R version
}
\details{
This function sets the R prompt including project name, the R
version, and the process ID. It is intended to be useful by
reinforcing the active project for an R session, particularly when
there may be multiple R sessions active. The R version can be useful
when running R on different machines, to reinforce which version of
R is active on the given machine. The process ID is mainly helpful in
the event an R process spins out of control, and it would be useful
to know definitively which exact process ID is stuck, so that it can
be killed without affecting other R sessions inadvertently.

The prompt is defined in \code{options("prompt")}.

Note that in some cases, the color encoding of the prompt interferes
with word wrapping, the symptom is that when typing text into the R console
a long line will begin to word wrap prematurely, before the text reaches
the edge of the screen. There are two frequent causes of this issue:

\describe{
\item{options("width")}{is sometimes defined too narrow for the
screen, which can happen when resizing the console, or when
accessing an R session via GNU screen, or tmux, and the environment
variable has not been propagated to the terminal window. Usually
this issue is resolved by defining \code{options("width")} manually,
or by simply resizing the terminal window, which may trigger the
appropriate environment variable updates.}
\item{The locale}{can sometimes be mismatched with the terminal window,
usually caused by some terminal emulation layer which is not
properly detecting the compatibility of the server. It may happen
for example, when using PuTTY on Windows, or when using GNU screen or
tmux on linux or Mac OSX. To troubleshoot, check
\code{Sys.env("LC_ALL")} which may be \code{"C"} or another locale such as
\code{"en_US.UTF-8"}. Note that switching locale may have the effect of
correcting the word wrap, but may adversely affect display of
non-standard unicode characters.}
}

In any event, R uses readline for unix-like systems by default, and
issues related to using color prompt are handled at that level. For example,
in some Mac OSX consoles, there are alternate color escape sequences which
are used to tell readline to ignore an escape sequence when it counts the
number of characters being displayed by the prompt.
}
\examples{
\dontrun{
setPrompt("jamba", projectColor="purple");
setPrompt("jamba", usePid=FALSE);
}

}
