% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-plots.r
\name{showColors}
\alias{showColors}
\title{Show colors from a vector or list}
\usage{
showColors(x, labelCells = NULL, transpose = FALSE, srtCellnote = NULL,
  ...)
}
\arguments{
\item{x}{vector of colors, or list of color vectors.}

\item{labelCells}{logical whether to label colors atop the color itself.
If NULL (default) it will only display labels with 40 or fewer items
on either axis.}

\item{transpose}{logical whether to transpose the colors to display
top-to-bottom, instead of left-to-right.}

\item{srtCellnote}{numerical angle to rotate text when
\code{labelCells=TRUE}. When set to NULL, labels are vertical
srtCellnote=90 when \code{transpose=FALSE} and horizontal
srtCellnote=0 when \code{transpose=TRUE}.}

\item{...}{additional parameters are sent to \code{\link{imageByColors}}.}
}
\value{
invisible color matrix used in \code{\link{imageByColors}}.
}
\description{
Show colors from a vector or list
}
\details{
This function simply displays colors for review, using
\code{\link{imageByColors}} to display colors and labels across the
plot space.
}
\examples{
x <- color2gradient(list(Reds=c("red"), Blues=c("blue")), n=c(4,7));
showColors(x);

if (suppressPackageStartupMessages(require(RColorBrewer))) {
   y <- lapply(nameVector(RColorBrewer:::namelist), function(i){
      brewer.pal(20,i);
   });
   showColors(y, cexCellnote=0.6, main="Brewer Colors");
}
if (suppressPackageStartupMessages(require(viridis))) {
   z <- rgb2col(viridis.map[,c("R","G","B")]);
   names(z) <- viridis.map[,"opt"];
   showColors(z, labelCells=TRUE, xaxt="n", main="viridis.map colors");
}

}
