% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-plots.r
\name{showColors}
\alias{showColors}
\title{Show colors from a vector or list}
\usage{
showColors(x, labelCells = NULL, transpose = FALSE,
  srtCellnote = NULL, adjustMargins = TRUE, ...)
}
\arguments{
\item{x}{vector of colors, or list of color vectors.}

\item{labelCells}{logical whether to label colors atop the color itself.
If NULL (default) it will only display labels with 40 or fewer items
on either axis.}

\item{transpose}{logical whether to transpose the colors to display
top-to-bottom, instead of left-to-right.}

\item{srtCellnote}{numerical angle to rotate text when
\code{labelCells=TRUE}. When set to NULL, labels are vertical
srtCellnote=90 when \code{transpose=FALSE} and horizontal
srtCellnote=0 when \code{transpose=TRUE}.}

\item{adjustMargins}{logical indicating whether to call
\code{\link{adjustAxisLabelMargins}} to adjust the x- and y-axis
label margins to accomodate the label size.}

\item{...}{additional parameters are sent to \code{\link{imageByColors}}.}
}
\value{
invisible color matrix used in \code{\link{imageByColors}}.
}
\description{
Show colors from a vector or list
}
\details{
This function simply displays colors for review, using
\code{\link{imageByColors}} to display colors and labels across the
plot space.
}
\examples{
x <- color2gradient(list(Reds=c("red"), Blues=c("blue")), n=c(4,7));
showColors(x);

if (suppressPackageStartupMessages(require(RColorBrewer))) {
   y <- lapply(nameVector(RColorBrewer:::namelist), function(i){
      brewer.pal(20,i);
   });
   showColors(y, cexCellnote=0.6, cex.axis=0.7, main="Brewer Colors");
}
if (suppressPackageStartupMessages(require(viridis))) {
   # Grab the full viridis color map
   z <- rgb2col(viridis.map[,c("R","G","B")]);
   # label the colors using viridis.map$opt
   names(z) <- viridis.map$opt;
   showColors(z, labelCells=TRUE, xaxt="n", main="viridis.map colors");
   # or split the colors into a list
   zl <- split(z, viridis.map[,"opt"]);
   showColors(zl, labelCells=TRUE, srtCellnote=0, xaxt="n",
      main="Viridis colors as a list");
}

}
\seealso{
Other jam plot functions: \code{\link{adjustAxisLabelMargins}},
  \code{\link{coordPresets}}, \code{\link{decideMfrow}},
  \code{\link{drawLabels}}, \code{\link{getPlotAspect}},
  \code{\link{imageByColors}}, \code{\link{imageDefault}},
  \code{\link{minorLogTicksAxis}}, \code{\link{nullPlot}},
  \code{\link{plotPolygonDensity}},
  \code{\link{plotSmoothScatter}},
  \code{\link{shadowText}}, \code{\link{smoothScatterJam}},
  \code{\link{sqrtAxis}}, \code{\link{usrBox}}

Other jam color functions: \code{\link{alpha2col}},
  \code{\link{applyCLrange}}, \code{\link{col2alpha}},
  \code{\link{col2hcl}}, \code{\link{col2hsv}},
  \code{\link{color2gradient}}, \code{\link{fixYellowHue}},
  \code{\link{fixYellow}}, \code{\link{getColorRamp}},
  \code{\link{hcl2col}}, \code{\link{hsv2col}},
  \code{\link{isColor}}, \code{\link{kable_coloring}},
  \code{\link{makeColorDarker}}, \code{\link{make_styles}},
  \code{\link{rgb2col}}, \code{\link{setCLranges}},
  \code{\link{setTextContrastColor}},
  \code{\link{unalpha}}, \code{\link{warpRamp}}
}
\concept{jam color functions}
\concept{jam plot functions}
