% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-plots.r
\name{sqrtAxis}
\alias{sqrtAxis}
\title{Determine square root axis tick mark positions}
\usage{
sqrtAxis(
  side = 1,
  x = NULL,
  pretty.n = 10,
  u5.bias = 0,
  big.mark = ",",
  plot = TRUE,
  las = 2,
  cex.axis = 0.6,
  ...
)
}
\arguments{
\item{side}{integer value indicating the axis position, as used
by \code{axis()}, 1=bottom, 2=left, 3=top, 4=right.}

\item{x}{optional numeric vector representing the numeric range
to be labeled.}

\item{pretty.n}{numeric value indicating the number of desired
tick marks, passed to \code{pretty()}.}

\item{u5.bias}{numeric value passed to \code{pretty()} to influence the
frequency of intermediate tick marks.}

\item{big.mark}{character value passed to \code{format()} which helps
visually distinguish numbers larger than 1000.}

\item{plot}{logical indicating whether to plot the axis tick
marks and labels.}

\item{las, cex.axis}{numeric values passed to \code{axis()} when drawing
the axis, by default \code{las=2} plots labels rotated
perpendicular to the axis.}

\item{...}{additional parameters are passed to \code{pretty()}.}
}
\value{
Invisibly returns a numeric vector of axis tick positions,
named by the display label.
The axis values are in square root space while the labels represent
the normal space values.
}
\description{
Determine square root axis tick mark positions
}
\details{
This function calculates positions for tick marks for data
that has been transformed with \code{sqrt()}, specifically a directional
transformation like \code{sqrt(abs(x)) * sign(x)}.

The main goal of this function is to provide reasonably placed
tick marks using integer values.
}
\seealso{
Other jam plot functions: 
\code{\link{adjustAxisLabelMargins}()},
\code{\link{coordPresets}()},
\code{\link{decideMfrow}()},
\code{\link{drawLabels}()},
\code{\link{getPlotAspect}()},
\code{\link{groupedAxis}()},
\code{\link{imageByColors}()},
\code{\link{imageDefault}()},
\code{\link{minorLogTicksAxis}()},
\code{\link{nullPlot}()},
\code{\link{plotPolygonDensity}()},
\code{\link{plotRidges}()},
\code{\link{plotSmoothScatter}()},
\code{\link{shadowText_options}()},
\code{\link{shadowText}()},
\code{\link{showColors}()},
\code{\link{smoothScatterJam}()},
\code{\link{usrBox}()}
}
\concept{jam plot functions}
