% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba.r
\name{tcount}
\alias{tcount}
\title{frequency of entries, ordered by frequency}
\usage{
tcount(x, doSort = TRUE, minCount = NULL, maxCount = NULL,
  nameSortFunc = sort, ...)
}
\description{
frequency of entries, ordered by frequency
}
