% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-string.r
\name{uniques}
\alias{uniques}
\title{apply unique to each element of a list}
\usage{
uniques(x, keepNames = TRUE, incomparables = FALSE, useBioc = TRUE, ...)
}
\arguments{
\item{x}{input list of vectors}

\item{keepNames}{boolean indicating whether to keep the list element
names in the returned results.}

\item{incomparables}{see \code{\link[=unique]{unique()}} for details, this value is only
sent to \code{\link[S4Vectors:unique]{S4Vectors::unique()}} when the Bioconductor package
\code{S4Vectors} is installed, and is ignored otherwise for efficiency.}

\item{useBioc}{boolean indicating whether this function should try
to use \code{\link[S4Vectors:unique]{S4Vectors::unique()}} when the Bioconductor package
\code{S4Vectors} is installed, otherwise it will use a somewhat less
efficient bulk operation.}
}
\value{
\code{list} with unique values in each list element.
}
\description{
Apply unique to each element of a list, usually a list of vectors
}
\details{
This function will attempt to use \code{\link[S4Vectors:unique]{S4Vectors::unique()}} which is
substantially faster than any \code{apply} family function, especially
for very long lists. However, when \code{S4Vectors} is not installed,
it applies uniqueness to the \code{unlist}ed vector of values, which is
also substantially faster than the \code{apply} family functions for
long lists, but which may still be less efficient than the
C implementation provided by \code{S4Vectors}.

This function will attempt to use \code{\link[S4Vectors:unique]{S4Vectors::unique()}} which is
substantially faster than the alternative \code{apply} family function,
especially for very long lists.

This function also by default applies a sort to each
list element, using \code{\link[=mixedOrder]{mixedOrder()}} for alphanumeric sorting. This
sort is applied to the entire \code{unlist}ed vector, so it is also
substantially faster than applying \code{sort} to each list element
individually.

When \code{makeUnique} is \code{TRUE}, it will make each list element unique,
either using the \code{\link[S4Vectors:unique]{S4Vectors::unique()}} function which is highly
optimized, or the fallback is to use a substantially slower \code{apply}
family function.
}
\seealso{
Other jam string functions: \code{\link{cPasteUnique}},
  \code{\link{cPaste}}, \code{\link{igrepHas}},
  \code{\link{igrep}}, \code{\link{mixedOrder}},
  \code{\link{mixedSortDF}}, \code{\link{mixedSort}},
  \code{\link{mmixedOrder}}, \code{\link{provigrep}},
  \code{\link{unigrep}}, \code{\link{unvigrep}},
  \code{\link{vgrep}}, \code{\link{vigrep}}
}
