% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-string.r
\name{uniques}
\alias{uniques}
\title{apply unique to each element of a list}
\usage{
uniques(x, keepNames = TRUE, incomparables = FALSE, useBioc = TRUE,
  ...)
}
\arguments{
\item{x}{input list of vectors}

\item{keepNames}{boolean indicating whether to keep the list element
names in the returned results.}

\item{incomparables}{see \code{\link[=unique]{unique()}} for details, this value is only
sent to \code{\link[S4Vectors:unique]{S4Vectors::unique()}} when the Bioconductor package
\code{S4Vectors} is installed, and is ignored otherwise for efficiency.}

\item{useBioc}{boolean indicating whether this function should try
to use \code{\link[S4Vectors:unique]{S4Vectors::unique()}} when the Bioconductor package
\code{S4Vectors} is installed, otherwise it will use a somewhat less
efficient bulk operation.}
}
\value{
\code{list} with unique values in each list element.
}
\description{
Apply unique to each element of a list, usually a list of vectors
}
\details{
This function will attempt to use \code{\link[S4Vectors:unique]{S4Vectors::unique()}} which is
substantially faster than any \code{apply} family function, especially
for very long lists. However, when \code{S4Vectors} is not installed,
it applies uniqueness to the \code{unlist}ed vector of values, which is
also substantially faster than the \code{apply} family functions for
long lists, but which may still be less efficient than the
C implementation provided by \code{S4Vectors}.
}
\seealso{
Other jam string functions: \code{\link{asSize}},
  \code{\link{breaksByVector}}, \code{\link{cPasteSU}},
  \code{\link{cPasteS}}, \code{\link{cPasteUnique}},
  \code{\link{cPasteU}}, \code{\link{cPaste}},
  \code{\link{fillBlanks}}, \code{\link{formatInt}},
  \code{\link{gsubOrdered}}, \code{\link{makeNames}},
  \code{\link{mixedOrder}}, \code{\link{mixedSortDF}},
  \code{\link{mixedSorts}}, \code{\link{mixedSort}},
  \code{\link{mmixedOrder}}, \code{\link{nameVectorN}},
  \code{\link{nameVector}}, \code{\link{padInteger}},
  \code{\link{padString}}, \code{\link{pasteByRowOrdered}},
  \code{\link{pasteByRow}}, \code{\link{tcount}},
  \code{\link{ucfirst}}

Other jam list functions: \code{\link{cPasteSU}},
  \code{\link{cPasteS}}, \code{\link{cPasteUnique}},
  \code{\link{cPasteU}}, \code{\link{cPaste}},
  \code{\link{list2df}}, \code{\link{mergeAllXY}},
  \code{\link{mixedSorts}}, \code{\link{rbindList}},
  \code{\link{relist_named}}, \code{\link{rlengths}},
  \code{\link{sclass}}, \code{\link{sdim}},
  \code{\link{ssdim}}, \code{\link{unnestList}}
}
\concept{jam list functions}
\concept{jam string functions}
