% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamba-string.r
\name{vigrep}
\alias{vigrep}
\title{case-insensitive grep, returning values}
\usage{
vigrep(..., value = TRUE, ignore.case = TRUE)
}
\arguments{
\item{..., value, ignore.case}{parameters sent to \code{base::grep()}}
}
\value{
vector of matching values
}
\description{
case-insensitive grep, returning values
}
\details{
This function is a simple wrapper around \code{base::grep()} which
runs in case-insensitive mode, and returns matching values. It is
particularly helpful when grabbing values from a vector.
}
\examples{
V <- paste0(LETTERS[1:5], LETTERS[4:8]);
vigrep("d", V);

}
\seealso{
Other jam grep functions: 
\code{\link{grepls}()},
\code{\link{igrep}()},
\code{\link{igrepHas}()},
\code{\link{igrepl}()},
\code{\link{provigrep}()},
\code{\link{unigrep}()},
\code{\link{unvigrep}()},
\code{\link{vgrep}()}
}
\concept{jam grep functions}
